==> 4_The_Final_Model/01b_Data_Import_Local.Rmd <==

Clearing environmnent
```{r}
cat("\f")
rm(list=ls())
```

Loading packages
```{r}
library(bigQueryR)
library(woe)

library(data.table)
library(caret)
library(glmnet)
library(ROCR)
library(knitr)
library(xlsx)

library(skimr)
library(gridExtra)
library(grid)
library(ggplot2)
library(lattice)
library(rpart)
library(randomForest)
library(dplyr)

specify_decimal <- function(x, k) trimws(format(round(x, k), nsmall=k))
```

Reading Bigquery data (read from folder)
```{r}
load("c:/Users/radics.gabor/Documents/icom/v07_Full_db_new/data/JZ_import_1.rda")
load("c:/Users/radics.gabor/Documents/icom/v07_Full_db_new/data/JZ_import_2.rda")
load("c:/Users/radics.gabor/Documents/icom/v07_Full_db_new/data/JZ_external_1.rda")
```
