==> 4_The_Final_Model/03.2_Feature_Engineering_Factor.Rmd <==
#1 week performance period (f_Target)

Train
```{r}
JZ_train <- JZ_train[, `:=`(
  f_apEV_event_autosuggest = ifelse(apEV_event_autosuggest <   0.5 , '1', ifelse(apEV_event_autosuggest <  2.5 , '2', '3'))
  , f_everEV_event_Mobilelhelyezkedes = ifelse(everEV_event_Mobilelhelyezkedes <   3.5 , '1', ifelse(everEV_event_Mobilelhelyezkedes <  14.5 , '2', '3'))
  , f_everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   2.5 , '1', ifelse(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <  11.5 , '2', '3'))
  , f_apEV_event_Mobilelhelyezkedes = ifelse(apEV_event_Mobilelhelyezkedes <   4.5 , '1', '2')
  , f_everEV_event_autosuggest = ifelse(everEV_event_autosuggest <   2.5 , '1', ifelse(everEV_event_autosuggest <  31.5 , '2', '3'))
  , f_wEV_event_autosuggest = ifelse(wEV_event_autosuggest <   0.5 , '1', ifelse(wEV_event_autosuggest <  2.5 , '2', '3'))
  , f_wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   2.5 , '1', '2')
  , f_apEV_timeOnSite = ifelse(apEV_timeOnSite <   81110 , '1', ifelse(apEV_timeOnSite <  877100 , '2', '3'))
  , f_apEV_pageviews = ifelse(apEV_pageviews <   3288 , '1', ifelse(apEV_pageviews <  14920 , '2', '3'))
  , f_apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   4.5 , '1', '2')
  , f_wEV_event_Mobilelhelyezkedes = ifelse(wEV_event_Mobilelhelyezkedes <   2.5 , '1', '2')
  , f_apEV_HirdetesTipus_Lakossag = ifelse(apEV_HirdetesTipus_Lakossag <   127.5 , '1', ifelse(apEV_HirdetesTipus_Lakossag <  486.5 , '2', '3'))
  , f_apEV_HirdetesEladasmodja_Elado = ifelse(apEV_HirdetesEladasmodja_Elado <   127.5 , '1', ifelse(apEV_HirdetesEladasmodja_Elado <  486.5 , '2', '3'))
  , f_apEV_device_desktop = ifelse(apEV_device_desktop <   164.5 , '1', ifelse(apEV_device_desktop <  444.5 , '2', '3'))
  , f_apEV_hits = ifelse(apEV_hits <   4267 , '1', ifelse(apEV_hits <  27120 , '2', '3'))
  , f_wEV_timeOnSite = ifelse(wEV_timeOnSite <   91650 , '1', ifelse(wEV_timeOnSite <  864100 , '2', '3'))
  , f_wEV_HirdetesTipus_Lakossag = ifelse(wEV_HirdetesTipus_Lakossag <   83.5 , '1', ifelse(wEV_HirdetesTipus_Lakossag <  348.5 , '2', '3'))
  , f_apEV_Honnan_Hungary = ifelse(apEV_Honnan_Hungary <   157.5 , '1', ifelse(apEV_Honnan_Hungary <  707.5 , '2', '3'))
  , f_wEV_HirdetesEladasmodja_Elado = ifelse(wEV_HirdetesEladasmodja_Elado <   80.5 , '1', ifelse(wEV_HirdetesEladasmodja_Elado <  338.5 , '2', '3'))
  , f_apEV_event_ReszletekOldalMegjelenes = ifelse(apEV_event_ReszletekOldalMegjelenes <   51.5 , '1', '2')
  , f_wEV_device_desktop = ifelse(wEV_device_desktop <   71.5 , '1', ifelse(wEV_device_desktop <  289.5 , '2', '3'))
  , f_wEV_pageviews = ifelse(wEV_pageviews <   2060 , '1', ifelse(wEV_pageviews <  11980 , '2', '3'))
  , f_wEV_Honnan_Hungary = ifelse(wEV_Honnan_Hungary <   122.5 , '1', ifelse(wEV_Honnan_Hungary <  366.5 , '2', '3'))
  , f_everEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(everEV_action_Kereso_ReszletesSzukitoLenyit <   1.5 , '1', ifelse(everEV_action_Kereso_ReszletesSzukitoLenyit <  10.5 , '2', '3'))
  , f_apEV_event_Kereso = ifelse(apEV_event_Kereso <   2.5 , '1', '2')
  , f_everEV_event_Kereso = ifelse(everEV_event_Kereso <   1.5 , '1', ifelse(everEV_event_Kereso <  10.5 , '2', '3'))
  , f_wEV_hits = ifelse(wEV_hits <   5662 , '1', '2')
  , f_apEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(apEV_action_Kereso_ReszletesSzukitoLenyit <   2.5 , '1', '2')
  , f_wEV_event_Kepgaleria = ifelse(wEV_event_Kepgaleria <   19.5 , '1', ifelse(wEV_event_Kepgaleria <  144.5 , '2', '3'))
  , f_wEV_event_ReszletekOldalMegjelenes = ifelse(wEV_event_ReszletekOldalMegjelenes <   52.5 , '1', '2')
  , f_apEV_event_Kepgaleria = ifelse(apEV_event_Kepgaleria <   47.5 , '1', ifelse(apEV_event_Kepgaleria <  229.5 , '2', '3'))
  , f_apEV_Visits = ifelse(apEV_Visits <   5.5 , '1', ifelse(apEV_Visits <  24.5 , '2', '3'))
  , f_everEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(everEV_action_Mobilelhelyezkedes_Linkethasznalt <   1.5 , '1', '2')
  , f_everEV_hits = ifelse(everEV_hits <   16800 , '1', ifelse(everEV_hits <  259900 , '2', '3'))
  , f_everEV_HirdetesEladasmodja_Elado = ifelse(everEV_HirdetesEladasmodja_Elado <   584.5 , '1', ifelse(everEV_HirdetesEladasmodja_Elado <  2830 , '2', '3'))
  , f_apEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(apEV_action_Mobilelhelyezkedes_Linkethasznalt <   0.5 , '1', '2')
  , f_everEV_event_Kereses = ifelse(everEV_event_Kereses <   1.5 , '1', '2')
  , f_wEV_Visits = ifelse(wEV_Visits <   9.5 , '1', '2')
  , f_everEV_HirdetesTipus_Lakossag = ifelse(everEV_HirdetesTipus_Lakossag <   584.5 , '1', '2')
  , f_apEV_event_Nincstalalatfrontend = ifelse(apEV_event_Nincstalalatfrontend <   0.5 , '1', '2')
  , f_wEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(wEV_action_Kereso_ReszletesSzukitoLenyit <   1.5 , '1', '2')
  , f_wEV_event_Kereso = ifelse(wEV_event_Kereso <   1.5 , '1', '2')
  , f_apEV_event_Erdeklodes = ifelse(apEV_event_Erdeklodes <   0.5 , '1', '2')
  , f_apEV_Bejelentkezve_Magan = ifelse(apEV_Bejelentkezve_Magan <   97.5 , '1', '2')
  , f_wEV_Bejelentkezve_Magan = ifelse(wEV_Bejelentkezve_Magan <   23.5 , '1', '2')
  , f_everEV_pageviews = ifelse(everEV_pageviews <   49070 , '1', '2')
  , f_apEV_referer_icom = ifelse(apEV_referer_icom <   0.5 , '1', '2')
  , f_everEV_timeOnSite = ifelse(everEV_timeOnSite <   2665000 , '1', '2')
  , f_wEV_Budapestrol_Budapestre = ifelse(wEV_Budapestrol_Budapestre <   81.5 , '1', '2')
  , f_everEV_event_Kepgaleria = ifelse(everEV_event_Kepgaleria <   104.5 , '1', ifelse(everEV_event_Kepgaleria <  1703 , '2', '3'))
  , f_everEV_Bejelentkezve_Magan = ifelse(everEV_Bejelentkezve_Magan <   238.5 , '1', '2')
  , f_apEV_Budapestrol_Budapestre = ifelse(apEV_Budapestrol_Budapestre <   168.5 , '1', '2')
  , f_apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   1.5 , '1', '2')
  , f_everEV_event_Nincstalalatfrontend = ifelse(everEV_event_Nincstalalatfrontend <   4.5 , '1', '2')
  , f_everEV_event_Eszkoztar = ifelse(everEV_event_Eszkoztar <   0.5 , '1', ifelse(everEV_event_Eszkoztar <  12.5 , '2', '3'))
  , f_apEV_channel_Organic_Search = ifelse(apEV_channel_Organic_Search <   104.5 , '1', '2')
  , f_wEV_action_Eszkoztar_Kattintas = ifelse(wEV_action_Eszkoztar_Kattintas <   0.5 , '1', '2')
  , f_apEV_action_Eszkoztar_Kattintas = ifelse(apEV_action_Eszkoztar_Kattintas <   1.5 , '1', '2')
  , f_wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   0.5 , '1', '2')
  , f_everEV_device_desktop = ifelse(everEV_device_desktop <   729.5 , '1', '2')
  , f_apEV_trafficSource_google_organic = ifelse(apEV_trafficSource_google_organic <   104.5 , '1', '2')
  , f_everEV_Honnan_Hungary = ifelse(everEV_Honnan_Hungary <   734.5 , '1', '2')
  , f_wEV_event_Erdeklodes = ifelse(wEV_event_Erdeklodes <   0.5 , '1', '2')
  , f_wEV_referer_icom = ifelse(wEV_referer_icom <   0.5 , '1', '2')
  , f_wEV_event_Eszkoztar = ifelse(wEV_event_Eszkoztar <   0.5 , '1', '2')
  , f_everEV_ReszletekOldalMegjelenes = ifelse(everEV_ReszletekOldalMegjelenes <   216.5 , '1', '2')
  , f_everEV_event_ReszletekOldalMegjelenes = ifelse(everEV_event_ReszletekOldalMegjelenes <   216.5 , '1', '2')
  , f_everEV_action_Kereses_ujEpitesuSzuro = ifelse(everEV_action_Kereses_ujEpitesuSzuro <   1.5 , '1', '2')
  , f_everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   2.5 , '1', '2')
  , f_everEV_action_Kereses_mobil = ifelse(everEV_action_Kereses_mobil <   0.5 , '1', '2')
  , f_everEV_channel_Organic_Search = ifelse(everEV_channel_Organic_Search <   292.5 , '1', '2')
  , f_everEV_event_M1X = ifelse(everEV_event_M1X <   1.5 , '1', '2')
  , f_apEV_event_Eszkoztar = ifelse(apEV_event_Eszkoztar <   3.5 , '1', '2')
  , f_wEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(wEV_action_Mobilelhelyezkedes_Linkethasznalt <   0.5 , '1', '2')
  , f_apEV_event_M1X = ifelse(apEV_event_M1X <   1.5 , '1', '2')
  , f_everEV_trafficSource_google_organic = ifelse(everEV_trafficSource_google_organic <   523.5 , '1', '2')
  , f_everEV_action_Eszkoztar_Kattintas = ifelse(everEV_action_Eszkoztar_Kattintas <   3.5 , '1', '2')
  , f_everEV_event_Erdeklodes = ifelse(everEV_event_Erdeklodes <   0.5 , '1', '2')
  , f_everEV_action_Mentes_Sikeres = ifelse(everEV_action_Mentes_Sikeres <   2.5 , '1', '2')
  , f_apEV_event_FrontendBelepes = ifelse(apEV_event_FrontendBelepes <   0.5 , '1', '2')
  , f_everEV_event_FrontendBelepes = ifelse(everEV_event_FrontendBelepes <   1.5 , '1', '2')
  , f_everEV_event_Mentes = ifelse(everEV_event_Mentes <   2.5 , '1', '2')
  , f_apEV_referer_google = ifelse(apEV_referer_google <   6.5 , '1', '2')
  , f_everEV_socialNetwork_Facebook = ifelse(everEV_socialNetwork_Facebook <   313.5 , '1', '2')
  , f_apEV_trafficSource_google_cpc = ifelse(apEV_trafficSource_google_cpc <   136.5 , '1', '2')
  , f_everEV_SocialSourceReferral = ifelse(everEV_SocialSourceReferral <   313.5 , '1', '2')
  , f_everEV_channel_Social = ifelse(everEV_channel_Social <   313.5 , '1', '2')
  , f_apEV_channel_Paid_Search = ifelse(apEV_channel_Paid_Search <   136.5 , '1', '2')
  , f_everEV_action_FrontendBelepes_Megjelenes = ifelse(everEV_action_FrontendBelepes_Megjelenes <   4.5 , '1', '2')
  , f_everEV_trafficSource_facebook_referral = ifelse(everEV_trafficSource_facebook_referral <   313.5 , '1', '2')
  , f_wEV_trafficSource_facebook_referral = ifelse(wEV_trafficSource_facebook_referral <   29.5 , '1', '2')
  , f_wEV_socialNetwork_Facebook = ifelse(wEV_socialNetwork_Facebook <   29.5 , '1', '2')
  , f_wEV_SocialSourceReferral = ifelse(wEV_SocialSourceReferral <   29.5 , '1', '2')
  , f_wEV_channel_Social = ifelse(wEV_channel_Social <   29.5 , '1', '2')
  , f_apEV_action_Erdeklodes_OfficeUtcapontos = ifelse(apEV_action_Erdeklodes_OfficeUtcapontos <   0.5 , '1', '2')
  , f_wEV_channel_Organic_Search = ifelse(wEV_channel_Organic_Search <   55.5 , '1', '2')
  , f_apEV_action_FrontendBelepes_Megjelenes = ifelse(apEV_action_FrontendBelepes_Megjelenes <   0.5 , '1', '2')
  , f_everEV_Budapestrol_Budapestre = ifelse(everEV_Budapestrol_Budapestre <   469.5 , '1', '2')
  , f_wEV_trafficSource_google_organic = ifelse(wEV_trafficSource_google_organic <   55.5 , '1', '2')
  , f_wEV_referer_google = ifelse(wEV_referer_google <   7.5 , '1', '2')
  , f_apEV_trafficSource_facebook_referral = ifelse(apEV_trafficSource_facebook_referral <   55.5 , '1', '2')
  , f_apEV_socialNetwork_Facebook = ifelse(apEV_socialNetwork_Facebook <   55.5 , '1', '2')
  , f_everEV_label_FrontendBelepes_Megjelenes_fejlec = ifelse(everEV_label_FrontendBelepes_Megjelenes_fejlec <   2.5 , '1', '2')
  , f_apEV_SocialSourceReferral = ifelse(apEV_SocialSourceReferral <   55.5 , '1', '2')
  , f_apEV_channel_Social = ifelse(apEV_channel_Social <   55.5 , '1', '2')
  , f_wEV_event_FrontendBelepes = ifelse(wEV_event_FrontendBelepes <   0.5 , '1', '2')
  , f_everEV_channel_Paid_Search = ifelse(everEV_channel_Paid_Search <   301.5 , '1', '2')
  , f_everEV_trafficSource_google_cpc = ifelse(everEV_trafficSource_google_cpc <   301.5 , '1', '2')
  , f_wEV_channel_Email = ifelse(wEV_channel_Email <   13.5 , '1', '2')
  , f_apEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(apEV_action_M1X_talalkozottDuplikaltHirdetessel <   4.5 , '1', '2')
  , f_everEV_event_OfficeLogo = ifelse(everEV_event_OfficeLogo <   0.5 , '1', '2')
  , f_wEV_trafficSource_google_cpc = ifelse(wEV_trafficSource_google_cpc <   69.5 , '1', '2')
  , f_everEV_channel_Email = ifelse(everEV_channel_Email <   280.5 , '1', '2')
  , f_everEV_event_AI = ifelse(everEV_event_AI <   0.5 , '1', '2')
  , f_everEV_event_lakoparkListaKattintas = ifelse(everEV_event_lakoparkListaKattintas <   0.5 , '1', '2')
  , f_wEV_event_M1X = ifelse(wEV_event_M1X <   2.5 , '1', '2')
  , f_everEV_event_lakoparkMegjelenes = ifelse(everEV_event_lakoparkMegjelenes <   1.5 , '1', '2')
  , f_everEV_referer_creativecdn = ifelse(everEV_referer_creativecdn <   2.5 , '1', '2')
  , f_wEV_channel_Paid_Search = ifelse(wEV_channel_Paid_Search <   68.5 , '1', '2')
  , f_everEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(everEV_action_M1X_talalkozottDuplikaltHirdetessel <   4.5 , '1', '2')
  , f_everEV_action_Erdeklodes_OfficeUtcapontos = ifelse(everEV_action_Erdeklodes_OfficeUtcapontos <   1.5 , '1', '2')
  , f_everEV_action_FrontendBelepes_Sikeres = ifelse(everEV_action_FrontendBelepes_Sikeres <   2.5 , '1', '2')
  , f_apEV_channel_Email = ifelse(apEV_channel_Email <   47.5 , '1', '2')
  , f_everEV_label_FrontendBelepes_Sikeres_fejlec = ifelse(everEV_label_FrontendBelepes_Sikeres_fejlec <   1.5 , '1', '2')
  , f_wEV_trafficSource_Hirdetesfigyelo_email = ifelse(wEV_trafficSource_Hirdetesfigyelo_email <   5.5 , '1', '2')
  , f_wEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(wEV_action_M1X_talalkozottDuplikaltHirdetessel <   2.5 , '1', '2')
  , f_apEV_action_Erdeklodes_OfficeNemUtcapontos = ifelse(apEV_action_Erdeklodes_OfficeNemUtcapontos <   0.5 , '1', '2')
  , f_apEV_referer_creativecdn = ifelse(apEV_referer_creativecdn <   0.5 , '1', '2')
  , f_apEV_trafficSource_Hirdetesfigyelo_email = ifelse(apEV_trafficSource_Hirdetesfigyelo_email <   18.5 , '1', '2')
  , f_everEV_action_Erdeklodes_MaganFizetos = ifelse(everEV_action_Erdeklodes_MaganFizetos <   0.5 , '1', '2')
  , f_everEV_referer_icom = ifelse(everEV_referer_icom <   8.5 , '1', '2')
  , f_everEV_action_Erdeklodes_OfficeNemUtcapontos = ifelse(everEV_action_Erdeklodes_OfficeNemUtcapontos <   3.5 , '1', '2')
  , f_everEV_trafficSource_Hirdetesfigyelo_email = ifelse(everEV_trafficSource_Hirdetesfigyelo_email <   233.5 , '1', '2')
  , f_everEV_referer_google = ifelse(everEV_referer_google <   40.5 , '1', '2')
  , f_wEV_action_Mobilreszletek_Kepgaleria = ifelse(wEV_action_Mobilreszletek_Kepgaleria <   1.5 , '1', '2')
  , f_everEV_Visits = ifelse(everEV_Visits <   49.5 , '1', '2')
  , f_everEV_Felhasznalo_Magan = ifelse(everEV_Felhasznalo_Magan <   0.5 , '1', '2')
  , f_everEV_event_Mobilreszletek = ifelse(everEV_event_Mobilreszletek <   33.5 , '1', '2')
  , f_wEV_event_Mobilreszletek = ifelse(wEV_event_Mobilreszletek <   3.5 , '1', '2')
  , f_everEV_action_Mobilreszletek_Kepgaleria = ifelse(everEV_action_Mobilreszletek_Kepgaleria <   25.5 , '1', '2')
  , f_everRE_AVG_Kepgaleria = ifelse(everRE_AVG_Kepgaleria <   0.06639 , '1', '2')
  , f_everEV_event_Piac = ifelse(everEV_event_Piac <   0.5 , '1', '2')
  , f_wRE_AVG_Kepgaleria = ifelse(wRE_AVG_Kepgaleria <   0.1235 , '1', '2')
  , f_wEV_trafficSource_rtbhouse_retargeting = ifelse(wEV_trafficSource_rtbhouse_retargeting <   0.5 , '1', '2')
  , f_everEV_channel_Direct = ifelse(everEV_channel_Direct <   92.5 , '1', '2')
  , f_wRE_SUM_Kepgaleria = ifelse(wRE_SUM_Kepgaleria <   2.5 , '1', '2')
  , f_everEV_action_Piac_undefined = ifelse(everEV_action_Piac_undefined <   0.5 , '1', '2')
  , f_wRE_MAX_Kepgaleria = ifelse(wRE_MAX_Kepgaleria <   0.5 , '1', '2')
  , f_apEV_trafficSource_rtbhouse_retargeting = ifelse(apEV_trafficSource_rtbhouse_retargeting <   0.5 , '1', '2')
  , f_apEV_action_Mobilreszletek_Kepgaleria = ifelse(apEV_action_Mobilreszletek_Kepgaleria <   1.5 , '1', '2')
  , f_apEV_event_Mobilreszletek = ifelse(apEV_event_Mobilreszletek <   12.5 , '1', '2')
  , f_apRE_AVG_Kepgaleria = ifelse(apRE_AVG_Kepgaleria <   0.2153 , '1', '2')
  , f_everEV_event_MaganhirdetoInfok = ifelse(everEV_event_MaganhirdetoInfok <   0.5 , '1', '2')
  , f_everRE_MAX_Kepgaleria = ifelse(everRE_MAX_Kepgaleria <   0.5 , '1', '2')
  , f_everEV_trafficSource_rtbhouse_retargeting = ifelse(everEV_trafficSource_rtbhouse_retargeting <   2.5 , '1', '2')
  , f_apRE_SUM_Kepgaleria = ifelse(apRE_SUM_Kepgaleria <   4.5 , '1', '2')
  , f_everRE_SUM_Kepgaleria = ifelse(everRE_SUM_Kepgaleria <   25.5 , '1', '2')
  , f_apRE_MAX_Kepgaleria = ifelse(apRE_MAX_Kepgaleria <   0.5 , '1', '2')
  , f_everEV_action_Erdeklodes_MaganIngyenes = ifelse(everEV_action_Erdeklodes_MaganIngyenes <   0.5 , '1', '2')
  , f_everRE_ikerhaz = ifelse(everRE_ikerhaz <   2.5 , '1', '2')
  , f_everEV_Videkrol_Budapestre = ifelse(everEV_Videkrol_Budapestre <   140.5 , '1', '2')
),]



```

Test
```{r}
JZ_test <- JZ_test[, `:=`(
  f_apEV_event_autosuggest = ifelse(apEV_event_autosuggest <   0.5 , '1', ifelse(apEV_event_autosuggest <  2.5 , '2', '3'))
  , f_everEV_event_Mobilelhelyezkedes = ifelse(everEV_event_Mobilelhelyezkedes <   3.5 , '1', ifelse(everEV_event_Mobilelhelyezkedes <  14.5 , '2', '3'))
  , f_everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   2.5 , '1', ifelse(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <  11.5 , '2', '3'))
  , f_apEV_event_Mobilelhelyezkedes = ifelse(apEV_event_Mobilelhelyezkedes <   4.5 , '1', '2')
  , f_everEV_event_autosuggest = ifelse(everEV_event_autosuggest <   2.5 , '1', ifelse(everEV_event_autosuggest <  31.5 , '2', '3'))
  , f_wEV_event_autosuggest = ifelse(wEV_event_autosuggest <   0.5 , '1', ifelse(wEV_event_autosuggest <  2.5 , '2', '3'))
  , f_wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   2.5 , '1', '2')
  , f_apEV_timeOnSite = ifelse(apEV_timeOnSite <   81110 , '1', ifelse(apEV_timeOnSite <  877100 , '2', '3'))
  , f_apEV_pageviews = ifelse(apEV_pageviews <   3288 , '1', ifelse(apEV_pageviews <  14920 , '2', '3'))
  , f_apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   4.5 , '1', '2')
  , f_wEV_event_Mobilelhelyezkedes = ifelse(wEV_event_Mobilelhelyezkedes <   2.5 , '1', '2')
  , f_apEV_HirdetesTipus_Lakossag = ifelse(apEV_HirdetesTipus_Lakossag <   127.5 , '1', ifelse(apEV_HirdetesTipus_Lakossag <  486.5 , '2', '3'))
  , f_apEV_HirdetesEladasmodja_Elado = ifelse(apEV_HirdetesEladasmodja_Elado <   127.5 , '1', ifelse(apEV_HirdetesEladasmodja_Elado <  486.5 , '2', '3'))
  , f_apEV_device_desktop = ifelse(apEV_device_desktop <   164.5 , '1', ifelse(apEV_device_desktop <  444.5 , '2', '3'))
  , f_apEV_hits = ifelse(apEV_hits <   4267 , '1', ifelse(apEV_hits <  27120 , '2', '3'))
  , f_wEV_timeOnSite = ifelse(wEV_timeOnSite <   91650 , '1', ifelse(wEV_timeOnSite <  864100 , '2', '3'))
  , f_wEV_HirdetesTipus_Lakossag = ifelse(wEV_HirdetesTipus_Lakossag <   83.5 , '1', ifelse(wEV_HirdetesTipus_Lakossag <  348.5 , '2', '3'))
  , f_apEV_Honnan_Hungary = ifelse(apEV_Honnan_Hungary <   157.5 , '1', ifelse(apEV_Honnan_Hungary <  707.5 , '2', '3'))
  , f_wEV_HirdetesEladasmodja_Elado = ifelse(wEV_HirdetesEladasmodja_Elado <   80.5 , '1', ifelse(wEV_HirdetesEladasmodja_Elado <  338.5 , '2', '3'))
  , f_apEV_event_ReszletekOldalMegjelenes = ifelse(apEV_event_ReszletekOldalMegjelenes <   51.5 , '1', '2')
  , f_wEV_device_desktop = ifelse(wEV_device_desktop <   71.5 , '1', ifelse(wEV_device_desktop <  289.5 , '2', '3'))
  , f_wEV_pageviews = ifelse(wEV_pageviews <   2060 , '1', ifelse(wEV_pageviews <  11980 , '2', '3'))
  , f_wEV_Honnan_Hungary = ifelse(wEV_Honnan_Hungary <   122.5 , '1', ifelse(wEV_Honnan_Hungary <  366.5 , '2', '3'))
  , f_everEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(everEV_action_Kereso_ReszletesSzukitoLenyit <   1.5 , '1', ifelse(everEV_action_Kereso_ReszletesSzukitoLenyit <  10.5 , '2', '3'))
  , f_apEV_event_Kereso = ifelse(apEV_event_Kereso <   2.5 , '1', '2')
  , f_everEV_event_Kereso = ifelse(everEV_event_Kereso <   1.5 , '1', ifelse(everEV_event_Kereso <  10.5 , '2', '3'))
  , f_wEV_hits = ifelse(wEV_hits <   5662 , '1', '2')
  , f_apEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(apEV_action_Kereso_ReszletesSzukitoLenyit <   2.5 , '1', '2')
  , f_wEV_event_Kepgaleria = ifelse(wEV_event_Kepgaleria <   19.5 , '1', ifelse(wEV_event_Kepgaleria <  144.5 , '2', '3'))
  , f_wEV_event_ReszletekOldalMegjelenes = ifelse(wEV_event_ReszletekOldalMegjelenes <   52.5 , '1', '2')
  , f_apEV_event_Kepgaleria = ifelse(apEV_event_Kepgaleria <   47.5 , '1', ifelse(apEV_event_Kepgaleria <  229.5 , '2', '3'))
  , f_apEV_Visits = ifelse(apEV_Visits <   5.5 , '1', ifelse(apEV_Visits <  24.5 , '2', '3'))
  , f_everEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(everEV_action_Mobilelhelyezkedes_Linkethasznalt <   1.5 , '1', '2')
  , f_everEV_hits = ifelse(everEV_hits <   16800 , '1', ifelse(everEV_hits <  259900 , '2', '3'))
  , f_everEV_HirdetesEladasmodja_Elado = ifelse(everEV_HirdetesEladasmodja_Elado <   584.5 , '1', ifelse(everEV_HirdetesEladasmodja_Elado <  2830 , '2', '3'))
  , f_apEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(apEV_action_Mobilelhelyezkedes_Linkethasznalt <   0.5 , '1', '2')
  , f_everEV_event_Kereses = ifelse(everEV_event_Kereses <   1.5 , '1', '2')
  , f_wEV_Visits = ifelse(wEV_Visits <   9.5 , '1', '2')
  , f_everEV_HirdetesTipus_Lakossag = ifelse(everEV_HirdetesTipus_Lakossag <   584.5 , '1', '2')
  , f_apEV_event_Nincstalalatfrontend = ifelse(apEV_event_Nincstalalatfrontend <   0.5 , '1', '2')
  , f_wEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(wEV_action_Kereso_ReszletesSzukitoLenyit <   1.5 , '1', '2')
  , f_wEV_event_Kereso = ifelse(wEV_event_Kereso <   1.5 , '1', '2')
  , f_apEV_event_Erdeklodes = ifelse(apEV_event_Erdeklodes <   0.5 , '1', '2')
  , f_apEV_Bejelentkezve_Magan = ifelse(apEV_Bejelentkezve_Magan <   97.5 , '1', '2')
  , f_wEV_Bejelentkezve_Magan = ifelse(wEV_Bejelentkezve_Magan <   23.5 , '1', '2')
  , f_everEV_pageviews = ifelse(everEV_pageviews <   49070 , '1', '2')
  , f_apEV_referer_icom = ifelse(apEV_referer_icom <   0.5 , '1', '2')
  , f_everEV_timeOnSite = ifelse(everEV_timeOnSite <   2665000 , '1', '2')
  , f_wEV_Budapestrol_Budapestre = ifelse(wEV_Budapestrol_Budapestre <   81.5 , '1', '2')
  , f_everEV_event_Kepgaleria = ifelse(everEV_event_Kepgaleria <   104.5 , '1', ifelse(everEV_event_Kepgaleria <  1703 , '2', '3'))
  , f_everEV_Bejelentkezve_Magan = ifelse(everEV_Bejelentkezve_Magan <   238.5 , '1', '2')
  , f_apEV_Budapestrol_Budapestre = ifelse(apEV_Budapestrol_Budapestre <   168.5 , '1', '2')
  , f_apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   1.5 , '1', '2')
  , f_everEV_event_Nincstalalatfrontend = ifelse(everEV_event_Nincstalalatfrontend <   4.5 , '1', '2')
  , f_everEV_event_Eszkoztar = ifelse(everEV_event_Eszkoztar <   0.5 , '1', ifelse(everEV_event_Eszkoztar <  12.5 , '2', '3'))
  , f_apEV_channel_Organic_Search = ifelse(apEV_channel_Organic_Search <   104.5 , '1', '2')
  , f_wEV_action_Eszkoztar_Kattintas = ifelse(wEV_action_Eszkoztar_Kattintas <   0.5 , '1', '2')
  , f_apEV_action_Eszkoztar_Kattintas = ifelse(apEV_action_Eszkoztar_Kattintas <   1.5 , '1', '2')
  , f_wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   0.5 , '1', '2')
  , f_everEV_device_desktop = ifelse(everEV_device_desktop <   729.5 , '1', '2')
  , f_apEV_trafficSource_google_organic = ifelse(apEV_trafficSource_google_organic <   104.5 , '1', '2')
  , f_everEV_Honnan_Hungary = ifelse(everEV_Honnan_Hungary <   734.5 , '1', '2')
  , f_wEV_event_Erdeklodes = ifelse(wEV_event_Erdeklodes <   0.5 , '1', '2')
  , f_wEV_referer_icom = ifelse(wEV_referer_icom <   0.5 , '1', '2')
  , f_wEV_event_Eszkoztar = ifelse(wEV_event_Eszkoztar <   0.5 , '1', '2')
  , f_everEV_ReszletekOldalMegjelenes = ifelse(everEV_ReszletekOldalMegjelenes <   216.5 , '1', '2')
  , f_everEV_event_ReszletekOldalMegjelenes = ifelse(everEV_event_ReszletekOldalMegjelenes <   216.5 , '1', '2')
  , f_everEV_action_Kereses_ujEpitesuSzuro = ifelse(everEV_action_Kereses_ujEpitesuSzuro <   1.5 , '1', '2')
  , f_everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   2.5 , '1', '2')
  , f_everEV_action_Kereses_mobil = ifelse(everEV_action_Kereses_mobil <   0.5 , '1', '2')
  , f_everEV_channel_Organic_Search = ifelse(everEV_channel_Organic_Search <   292.5 , '1', '2')
  , f_everEV_event_M1X = ifelse(everEV_event_M1X <   1.5 , '1', '2')
  , f_apEV_event_Eszkoztar = ifelse(apEV_event_Eszkoztar <   3.5 , '1', '2')
  , f_wEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(wEV_action_Mobilelhelyezkedes_Linkethasznalt <   0.5 , '1', '2')
  , f_apEV_event_M1X = ifelse(apEV_event_M1X <   1.5 , '1', '2')
  , f_everEV_trafficSource_google_organic = ifelse(everEV_trafficSource_google_organic <   523.5 , '1', '2')
  , f_everEV_action_Eszkoztar_Kattintas = ifelse(everEV_action_Eszkoztar_Kattintas <   3.5 , '1', '2')
  , f_everEV_event_Erdeklodes = ifelse(everEV_event_Erdeklodes <   0.5 , '1', '2')
  , f_everEV_action_Mentes_Sikeres = ifelse(everEV_action_Mentes_Sikeres <   2.5 , '1', '2')
  , f_apEV_event_FrontendBelepes = ifelse(apEV_event_FrontendBelepes <   0.5 , '1', '2')
  , f_everEV_event_FrontendBelepes = ifelse(everEV_event_FrontendBelepes <   1.5 , '1', '2')
  , f_everEV_event_Mentes = ifelse(everEV_event_Mentes <   2.5 , '1', '2')
  , f_apEV_referer_google = ifelse(apEV_referer_google <   6.5 , '1', '2')
  , f_everEV_socialNetwork_Facebook = ifelse(everEV_socialNetwork_Facebook <   313.5 , '1', '2')
  , f_apEV_trafficSource_google_cpc = ifelse(apEV_trafficSource_google_cpc <   136.5 , '1', '2')
  , f_everEV_SocialSourceReferral = ifelse(everEV_SocialSourceReferral <   313.5 , '1', '2')
  , f_everEV_channel_Social = ifelse(everEV_channel_Social <   313.5 , '1', '2')
  , f_apEV_channel_Paid_Search = ifelse(apEV_channel_Paid_Search <   136.5 , '1', '2')
  , f_everEV_action_FrontendBelepes_Megjelenes = ifelse(everEV_action_FrontendBelepes_Megjelenes <   4.5 , '1', '2')
  , f_everEV_trafficSource_facebook_referral = ifelse(everEV_trafficSource_facebook_referral <   313.5 , '1', '2')
  , f_wEV_trafficSource_facebook_referral = ifelse(wEV_trafficSource_facebook_referral <   29.5 , '1', '2')
  , f_wEV_socialNetwork_Facebook = ifelse(wEV_socialNetwork_Facebook <   29.5 , '1', '2')
  , f_wEV_SocialSourceReferral = ifelse(wEV_SocialSourceReferral <   29.5 , '1', '2')
  , f_wEV_channel_Social = ifelse(wEV_channel_Social <   29.5 , '1', '2')
  , f_apEV_action_Erdeklodes_OfficeUtcapontos = ifelse(apEV_action_Erdeklodes_OfficeUtcapontos <   0.5 , '1', '2')
  , f_wEV_channel_Organic_Search = ifelse(wEV_channel_Organic_Search <   55.5 , '1', '2')
  , f_apEV_action_FrontendBelepes_Megjelenes = ifelse(apEV_action_FrontendBelepes_Megjelenes <   0.5 , '1', '2')
  , f_everEV_Budapestrol_Budapestre = ifelse(everEV_Budapestrol_Budapestre <   469.5 , '1', '2')
  , f_wEV_trafficSource_google_organic = ifelse(wEV_trafficSource_google_organic <   55.5 , '1', '2')
  , f_wEV_referer_google = ifelse(wEV_referer_google <   7.5 , '1', '2')
  , f_apEV_trafficSource_facebook_referral = ifelse(apEV_trafficSource_facebook_referral <   55.5 , '1', '2')
  , f_apEV_socialNetwork_Facebook = ifelse(apEV_socialNetwork_Facebook <   55.5 , '1', '2')
  , f_everEV_label_FrontendBelepes_Megjelenes_fejlec = ifelse(everEV_label_FrontendBelepes_Megjelenes_fejlec <   2.5 , '1', '2')
  , f_apEV_SocialSourceReferral = ifelse(apEV_SocialSourceReferral <   55.5 , '1', '2')
  , f_apEV_channel_Social = ifelse(apEV_channel_Social <   55.5 , '1', '2')
  , f_wEV_event_FrontendBelepes = ifelse(wEV_event_FrontendBelepes <   0.5 , '1', '2')
  , f_everEV_channel_Paid_Search = ifelse(everEV_channel_Paid_Search <   301.5 , '1', '2')
  , f_everEV_trafficSource_google_cpc = ifelse(everEV_trafficSource_google_cpc <   301.5 , '1', '2')
  , f_wEV_channel_Email = ifelse(wEV_channel_Email <   13.5 , '1', '2')
  , f_apEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(apEV_action_M1X_talalkozottDuplikaltHirdetessel <   4.5 , '1', '2')
  , f_everEV_event_OfficeLogo = ifelse(everEV_event_OfficeLogo <   0.5 , '1', '2')
  , f_wEV_trafficSource_google_cpc = ifelse(wEV_trafficSource_google_cpc <   69.5 , '1', '2')
  , f_everEV_channel_Email = ifelse(everEV_channel_Email <   280.5 , '1', '2')
  , f_everEV_event_AI = ifelse(everEV_event_AI <   0.5 , '1', '2')
  , f_everEV_event_lakoparkListaKattintas = ifelse(everEV_event_lakoparkListaKattintas <   0.5 , '1', '2')
  , f_wEV_event_M1X = ifelse(wEV_event_M1X <   2.5 , '1', '2')
  , f_everEV_event_lakoparkMegjelenes = ifelse(everEV_event_lakoparkMegjelenes <   1.5 , '1', '2')
  , f_everEV_referer_creativecdn = ifelse(everEV_referer_creativecdn <   2.5 , '1', '2')
  , f_wEV_channel_Paid_Search = ifelse(wEV_channel_Paid_Search <   68.5 , '1', '2')
  , f_everEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(everEV_action_M1X_talalkozottDuplikaltHirdetessel <   4.5 , '1', '2')
  , f_everEV_action_Erdeklodes_OfficeUtcapontos = ifelse(everEV_action_Erdeklodes_OfficeUtcapontos <   1.5 , '1', '2')
  , f_everEV_action_FrontendBelepes_Sikeres = ifelse(everEV_action_FrontendBelepes_Sikeres <   2.5 , '1', '2')
  , f_apEV_channel_Email = ifelse(apEV_channel_Email <   47.5 , '1', '2')
  , f_everEV_label_FrontendBelepes_Sikeres_fejlec = ifelse(everEV_label_FrontendBelepes_Sikeres_fejlec <   1.5 , '1', '2')
  , f_wEV_trafficSource_Hirdetesfigyelo_email = ifelse(wEV_trafficSource_Hirdetesfigyelo_email <   5.5 , '1', '2')
  , f_wEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(wEV_action_M1X_talalkozottDuplikaltHirdetessel <   2.5 , '1', '2')
  , f_apEV_action_Erdeklodes_OfficeNemUtcapontos = ifelse(apEV_action_Erdeklodes_OfficeNemUtcapontos <   0.5 , '1', '2')
  , f_apEV_referer_creativecdn = ifelse(apEV_referer_creativecdn <   0.5 , '1', '2')
  , f_apEV_trafficSource_Hirdetesfigyelo_email = ifelse(apEV_trafficSource_Hirdetesfigyelo_email <   18.5 , '1', '2')
  , f_everEV_action_Erdeklodes_MaganFizetos = ifelse(everEV_action_Erdeklodes_MaganFizetos <   0.5 , '1', '2')
  , f_everEV_referer_icom = ifelse(everEV_referer_icom <   8.5 , '1', '2')
  , f_everEV_action_Erdeklodes_OfficeNemUtcapontos = ifelse(everEV_action_Erdeklodes_OfficeNemUtcapontos <   3.5 , '1', '2')
  , f_everEV_trafficSource_Hirdetesfigyelo_email = ifelse(everEV_trafficSource_Hirdetesfigyelo_email <   233.5 , '1', '2')
  , f_everEV_referer_google = ifelse(everEV_referer_google <   40.5 , '1', '2')
  , f_wEV_action_Mobilreszletek_Kepgaleria = ifelse(wEV_action_Mobilreszletek_Kepgaleria <   1.5 , '1', '2')
  , f_everEV_Visits = ifelse(everEV_Visits <   49.5 , '1', '2')
  , f_everEV_Felhasznalo_Magan = ifelse(everEV_Felhasznalo_Magan <   0.5 , '1', '2')
  , f_everEV_event_Mobilreszletek = ifelse(everEV_event_Mobilreszletek <   33.5 , '1', '2')
  , f_wEV_event_Mobilreszletek = ifelse(wEV_event_Mobilreszletek <   3.5 , '1', '2')
  , f_everEV_action_Mobilreszletek_Kepgaleria = ifelse(everEV_action_Mobilreszletek_Kepgaleria <   25.5 , '1', '2')
  , f_everRE_AVG_Kepgaleria = ifelse(everRE_AVG_Kepgaleria <   0.06639 , '1', '2')
  , f_everEV_event_Piac = ifelse(everEV_event_Piac <   0.5 , '1', '2')
  , f_wRE_AVG_Kepgaleria = ifelse(wRE_AVG_Kepgaleria <   0.1235 , '1', '2')
  , f_wEV_trafficSource_rtbhouse_retargeting = ifelse(wEV_trafficSource_rtbhouse_retargeting <   0.5 , '1', '2')
  , f_everEV_channel_Direct = ifelse(everEV_channel_Direct <   92.5 , '1', '2')
  , f_wRE_SUM_Kepgaleria = ifelse(wRE_SUM_Kepgaleria <   2.5 , '1', '2')
  , f_everEV_action_Piac_undefined = ifelse(everEV_action_Piac_undefined <   0.5 , '1', '2')
  , f_wRE_MAX_Kepgaleria = ifelse(wRE_MAX_Kepgaleria <   0.5 , '1', '2')
  , f_apEV_trafficSource_rtbhouse_retargeting = ifelse(apEV_trafficSource_rtbhouse_retargeting <   0.5 , '1', '2')
  , f_apEV_action_Mobilreszletek_Kepgaleria = ifelse(apEV_action_Mobilreszletek_Kepgaleria <   1.5 , '1', '2')
  , f_apEV_event_Mobilreszletek = ifelse(apEV_event_Mobilreszletek <   12.5 , '1', '2')
  , f_apRE_AVG_Kepgaleria = ifelse(apRE_AVG_Kepgaleria <   0.2153 , '1', '2')
  , f_everEV_event_MaganhirdetoInfok = ifelse(everEV_event_MaganhirdetoInfok <   0.5 , '1', '2')
  , f_everRE_MAX_Kepgaleria = ifelse(everRE_MAX_Kepgaleria <   0.5 , '1', '2')
  , f_everEV_trafficSource_rtbhouse_retargeting = ifelse(everEV_trafficSource_rtbhouse_retargeting <   2.5 , '1', '2')
  , f_apRE_SUM_Kepgaleria = ifelse(apRE_SUM_Kepgaleria <   4.5 , '1', '2')
  , f_everRE_SUM_Kepgaleria = ifelse(everRE_SUM_Kepgaleria <   25.5 , '1', '2')
  , f_apRE_MAX_Kepgaleria = ifelse(apRE_MAX_Kepgaleria <   0.5 , '1', '2')
  , f_everEV_action_Erdeklodes_MaganIngyenes = ifelse(everEV_action_Erdeklodes_MaganIngyenes <   0.5 , '1', '2')
  , f_everRE_ikerhaz = ifelse(everRE_ikerhaz <   2.5 , '1', '2')
  , f_everEV_Videkrol_Budapestre = ifelse(everEV_Videkrol_Budapestre <   140.5 , '1', '2')
),]



```

External
```{r}
JZ_external <- JZ_external[, `:=`(
  f_apEV_event_autosuggest = ifelse(apEV_event_autosuggest <   0.5 , '1', ifelse(apEV_event_autosuggest <  2.5 , '2', '3'))
  , f_everEV_event_Mobilelhelyezkedes = ifelse(everEV_event_Mobilelhelyezkedes <   3.5 , '1', ifelse(everEV_event_Mobilelhelyezkedes <  14.5 , '2', '3'))
  , f_everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   2.5 , '1', ifelse(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <  11.5 , '2', '3'))
  , f_apEV_event_Mobilelhelyezkedes = ifelse(apEV_event_Mobilelhelyezkedes <   4.5 , '1', '2')
  , f_everEV_event_autosuggest = ifelse(everEV_event_autosuggest <   2.5 , '1', ifelse(everEV_event_autosuggest <  31.5 , '2', '3'))
  , f_wEV_event_autosuggest = ifelse(wEV_event_autosuggest <   0.5 , '1', ifelse(wEV_event_autosuggest <  2.5 , '2', '3'))
  , f_wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   2.5 , '1', '2')
  , f_apEV_timeOnSite = ifelse(apEV_timeOnSite <   81110 , '1', ifelse(apEV_timeOnSite <  877100 , '2', '3'))
  , f_apEV_pageviews = ifelse(apEV_pageviews <   3288 , '1', ifelse(apEV_pageviews <  14920 , '2', '3'))
  , f_apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   4.5 , '1', '2')
  , f_wEV_event_Mobilelhelyezkedes = ifelse(wEV_event_Mobilelhelyezkedes <   2.5 , '1', '2')
  , f_apEV_HirdetesTipus_Lakossag = ifelse(apEV_HirdetesTipus_Lakossag <   127.5 , '1', ifelse(apEV_HirdetesTipus_Lakossag <  486.5 , '2', '3'))
  , f_apEV_HirdetesEladasmodja_Elado = ifelse(apEV_HirdetesEladasmodja_Elado <   127.5 , '1', ifelse(apEV_HirdetesEladasmodja_Elado <  486.5 , '2', '3'))
  , f_apEV_device_desktop = ifelse(apEV_device_desktop <   164.5 , '1', ifelse(apEV_device_desktop <  444.5 , '2', '3'))
  , f_apEV_hits = ifelse(apEV_hits <   4267 , '1', ifelse(apEV_hits <  27120 , '2', '3'))
  , f_wEV_timeOnSite = ifelse(wEV_timeOnSite <   91650 , '1', ifelse(wEV_timeOnSite <  864100 , '2', '3'))
  , f_wEV_HirdetesTipus_Lakossag = ifelse(wEV_HirdetesTipus_Lakossag <   83.5 , '1', ifelse(wEV_HirdetesTipus_Lakossag <  348.5 , '2', '3'))
  , f_apEV_Honnan_Hungary = ifelse(apEV_Honnan_Hungary <   157.5 , '1', ifelse(apEV_Honnan_Hungary <  707.5 , '2', '3'))
  , f_wEV_HirdetesEladasmodja_Elado = ifelse(wEV_HirdetesEladasmodja_Elado <   80.5 , '1', ifelse(wEV_HirdetesEladasmodja_Elado <  338.5 , '2', '3'))
  , f_apEV_event_ReszletekOldalMegjelenes = ifelse(apEV_event_ReszletekOldalMegjelenes <   51.5 , '1', '2')
  , f_wEV_device_desktop = ifelse(wEV_device_desktop <   71.5 , '1', ifelse(wEV_device_desktop <  289.5 , '2', '3'))
  , f_wEV_pageviews = ifelse(wEV_pageviews <   2060 , '1', ifelse(wEV_pageviews <  11980 , '2', '3'))
  , f_wEV_Honnan_Hungary = ifelse(wEV_Honnan_Hungary <   122.5 , '1', ifelse(wEV_Honnan_Hungary <  366.5 , '2', '3'))
  , f_everEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(everEV_action_Kereso_ReszletesSzukitoLenyit <   1.5 , '1', ifelse(everEV_action_Kereso_ReszletesSzukitoLenyit <  10.5 , '2', '3'))
  , f_apEV_event_Kereso = ifelse(apEV_event_Kereso <   2.5 , '1', '2')
  , f_everEV_event_Kereso = ifelse(everEV_event_Kereso <   1.5 , '1', ifelse(everEV_event_Kereso <  10.5 , '2', '3'))
  , f_wEV_hits = ifelse(wEV_hits <   5662 , '1', '2')
  , f_apEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(apEV_action_Kereso_ReszletesSzukitoLenyit <   2.5 , '1', '2')
  , f_wEV_event_Kepgaleria = ifelse(wEV_event_Kepgaleria <   19.5 , '1', ifelse(wEV_event_Kepgaleria <  144.5 , '2', '3'))
  , f_wEV_event_ReszletekOldalMegjelenes = ifelse(wEV_event_ReszletekOldalMegjelenes <   52.5 , '1', '2')
  , f_apEV_event_Kepgaleria = ifelse(apEV_event_Kepgaleria <   47.5 , '1', ifelse(apEV_event_Kepgaleria <  229.5 , '2', '3'))
  , f_apEV_Visits = ifelse(apEV_Visits <   5.5 , '1', ifelse(apEV_Visits <  24.5 , '2', '3'))
  , f_everEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(everEV_action_Mobilelhelyezkedes_Linkethasznalt <   1.5 , '1', '2')
  , f_everEV_hits = ifelse(everEV_hits <   16800 , '1', ifelse(everEV_hits <  259900 , '2', '3'))
  , f_everEV_HirdetesEladasmodja_Elado = ifelse(everEV_HirdetesEladasmodja_Elado <   584.5 , '1', ifelse(everEV_HirdetesEladasmodja_Elado <  2830 , '2', '3'))
  , f_apEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(apEV_action_Mobilelhelyezkedes_Linkethasznalt <   0.5 , '1', '2')
  , f_everEV_event_Kereses = ifelse(everEV_event_Kereses <   1.5 , '1', '2')
  , f_wEV_Visits = ifelse(wEV_Visits <   9.5 , '1', '2')
  , f_everEV_HirdetesTipus_Lakossag = ifelse(everEV_HirdetesTipus_Lakossag <   584.5 , '1', '2')
  , f_apEV_event_Nincstalalatfrontend = ifelse(apEV_event_Nincstalalatfrontend <   0.5 , '1', '2')
  , f_wEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(wEV_action_Kereso_ReszletesSzukitoLenyit <   1.5 , '1', '2')
  , f_wEV_event_Kereso = ifelse(wEV_event_Kereso <   1.5 , '1', '2')
  , f_apEV_event_Erdeklodes = ifelse(apEV_event_Erdeklodes <   0.5 , '1', '2')
  , f_apEV_Bejelentkezve_Magan = ifelse(apEV_Bejelentkezve_Magan <   97.5 , '1', '2')
  , f_wEV_Bejelentkezve_Magan = ifelse(wEV_Bejelentkezve_Magan <   23.5 , '1', '2')
  , f_everEV_pageviews = ifelse(everEV_pageviews <   49070 , '1', '2')
  , f_apEV_referer_icom = ifelse(apEV_referer_icom <   0.5 , '1', '2')
  , f_everEV_timeOnSite = ifelse(everEV_timeOnSite <   2665000 , '1', '2')
  , f_wEV_Budapestrol_Budapestre = ifelse(wEV_Budapestrol_Budapestre <   81.5 , '1', '2')
  , f_everEV_event_Kepgaleria = ifelse(everEV_event_Kepgaleria <   104.5 , '1', ifelse(everEV_event_Kepgaleria <  1703 , '2', '3'))
  , f_everEV_Bejelentkezve_Magan = ifelse(everEV_Bejelentkezve_Magan <   238.5 , '1', '2')
  , f_apEV_Budapestrol_Budapestre = ifelse(apEV_Budapestrol_Budapestre <   168.5 , '1', '2')
  , f_apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   1.5 , '1', '2')
  , f_everEV_event_Nincstalalatfrontend = ifelse(everEV_event_Nincstalalatfrontend <   4.5 , '1', '2')
  , f_everEV_event_Eszkoztar = ifelse(everEV_event_Eszkoztar <   0.5 , '1', ifelse(everEV_event_Eszkoztar <  12.5 , '2', '3'))
  , f_apEV_channel_Organic_Search = ifelse(apEV_channel_Organic_Search <   104.5 , '1', '2')
  , f_wEV_action_Eszkoztar_Kattintas = ifelse(wEV_action_Eszkoztar_Kattintas <   0.5 , '1', '2')
  , f_apEV_action_Eszkoztar_Kattintas = ifelse(apEV_action_Eszkoztar_Kattintas <   1.5 , '1', '2')
  , f_wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   0.5 , '1', '2')
  , f_everEV_device_desktop = ifelse(everEV_device_desktop <   729.5 , '1', '2')
  , f_apEV_trafficSource_google_organic = ifelse(apEV_trafficSource_google_organic <   104.5 , '1', '2')
  , f_everEV_Honnan_Hungary = ifelse(everEV_Honnan_Hungary <   734.5 , '1', '2')
  , f_wEV_event_Erdeklodes = ifelse(wEV_event_Erdeklodes <   0.5 , '1', '2')
  , f_wEV_referer_icom = ifelse(wEV_referer_icom <   0.5 , '1', '2')
  , f_wEV_event_Eszkoztar = ifelse(wEV_event_Eszkoztar <   0.5 , '1', '2')
  , f_everEV_ReszletekOldalMegjelenes = ifelse(everEV_ReszletekOldalMegjelenes <   216.5 , '1', '2')
  , f_everEV_event_ReszletekOldalMegjelenes = ifelse(everEV_event_ReszletekOldalMegjelenes <   216.5 , '1', '2')
  , f_everEV_action_Kereses_ujEpitesuSzuro = ifelse(everEV_action_Kereses_ujEpitesuSzuro <   1.5 , '1', '2')
  , f_everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   2.5 , '1', '2')
  , f_everEV_action_Kereses_mobil = ifelse(everEV_action_Kereses_mobil <   0.5 , '1', '2')
  , f_everEV_channel_Organic_Search = ifelse(everEV_channel_Organic_Search <   292.5 , '1', '2')
  , f_everEV_event_M1X = ifelse(everEV_event_M1X <   1.5 , '1', '2')
  , f_apEV_event_Eszkoztar = ifelse(apEV_event_Eszkoztar <   3.5 , '1', '2')
  , f_wEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(wEV_action_Mobilelhelyezkedes_Linkethasznalt <   0.5 , '1', '2')
  , f_apEV_event_M1X = ifelse(apEV_event_M1X <   1.5 , '1', '2')
  , f_everEV_trafficSource_google_organic = ifelse(everEV_trafficSource_google_organic <   523.5 , '1', '2')
  , f_everEV_action_Eszkoztar_Kattintas = ifelse(everEV_action_Eszkoztar_Kattintas <   3.5 , '1', '2')
  , f_everEV_event_Erdeklodes = ifelse(everEV_event_Erdeklodes <   0.5 , '1', '2')
  , f_everEV_action_Mentes_Sikeres = ifelse(everEV_action_Mentes_Sikeres <   2.5 , '1', '2')
  , f_apEV_event_FrontendBelepes = ifelse(apEV_event_FrontendBelepes <   0.5 , '1', '2')
  , f_everEV_event_FrontendBelepes = ifelse(everEV_event_FrontendBelepes <   1.5 , '1', '2')
  , f_everEV_event_Mentes = ifelse(everEV_event_Mentes <   2.5 , '1', '2')
  , f_apEV_referer_google = ifelse(apEV_referer_google <   6.5 , '1', '2')
  , f_everEV_socialNetwork_Facebook = ifelse(everEV_socialNetwork_Facebook <   313.5 , '1', '2')
  , f_apEV_trafficSource_google_cpc = ifelse(apEV_trafficSource_google_cpc <   136.5 , '1', '2')
  , f_everEV_SocialSourceReferral = ifelse(everEV_SocialSourceReferral <   313.5 , '1', '2')
  , f_everEV_channel_Social = ifelse(everEV_channel_Social <   313.5 , '1', '2')
  , f_apEV_channel_Paid_Search = ifelse(apEV_channel_Paid_Search <   136.5 , '1', '2')
  , f_everEV_action_FrontendBelepes_Megjelenes = ifelse(everEV_action_FrontendBelepes_Megjelenes <   4.5 , '1', '2')
  , f_everEV_trafficSource_facebook_referral = ifelse(everEV_trafficSource_facebook_referral <   313.5 , '1', '2')
  , f_wEV_trafficSource_facebook_referral = ifelse(wEV_trafficSource_facebook_referral <   29.5 , '1', '2')
  , f_wEV_socialNetwork_Facebook = ifelse(wEV_socialNetwork_Facebook <   29.5 , '1', '2')
  , f_wEV_SocialSourceReferral = ifelse(wEV_SocialSourceReferral <   29.5 , '1', '2')
  , f_wEV_channel_Social = ifelse(wEV_channel_Social <   29.5 , '1', '2')
  , f_apEV_action_Erdeklodes_OfficeUtcapontos = ifelse(apEV_action_Erdeklodes_OfficeUtcapontos <   0.5 , '1', '2')
  , f_wEV_channel_Organic_Search = ifelse(wEV_channel_Organic_Search <   55.5 , '1', '2')
  , f_apEV_action_FrontendBelepes_Megjelenes = ifelse(apEV_action_FrontendBelepes_Megjelenes <   0.5 , '1', '2')
  , f_everEV_Budapestrol_Budapestre = ifelse(everEV_Budapestrol_Budapestre <   469.5 , '1', '2')
  , f_wEV_trafficSource_google_organic = ifelse(wEV_trafficSource_google_organic <   55.5 , '1', '2')
  , f_wEV_referer_google = ifelse(wEV_referer_google <   7.5 , '1', '2')
  , f_apEV_trafficSource_facebook_referral = ifelse(apEV_trafficSource_facebook_referral <   55.5 , '1', '2')
  , f_apEV_socialNetwork_Facebook = ifelse(apEV_socialNetwork_Facebook <   55.5 , '1', '2')
  , f_everEV_label_FrontendBelepes_Megjelenes_fejlec = ifelse(everEV_label_FrontendBelepes_Megjelenes_fejlec <   2.5 , '1', '2')
  , f_apEV_SocialSourceReferral = ifelse(apEV_SocialSourceReferral <   55.5 , '1', '2')
  , f_apEV_channel_Social = ifelse(apEV_channel_Social <   55.5 , '1', '2')
  , f_wEV_event_FrontendBelepes = ifelse(wEV_event_FrontendBelepes <   0.5 , '1', '2')
  , f_everEV_channel_Paid_Search = ifelse(everEV_channel_Paid_Search <   301.5 , '1', '2')
  , f_everEV_trafficSource_google_cpc = ifelse(everEV_trafficSource_google_cpc <   301.5 , '1', '2')
  , f_wEV_channel_Email = ifelse(wEV_channel_Email <   13.5 , '1', '2')
  , f_apEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(apEV_action_M1X_talalkozottDuplikaltHirdetessel <   4.5 , '1', '2')
  , f_everEV_event_OfficeLogo = ifelse(everEV_event_OfficeLogo <   0.5 , '1', '2')
  , f_wEV_trafficSource_google_cpc = ifelse(wEV_trafficSource_google_cpc <   69.5 , '1', '2')
  , f_everEV_channel_Email = ifelse(everEV_channel_Email <   280.5 , '1', '2')
  , f_everEV_event_AI = ifelse(everEV_event_AI <   0.5 , '1', '2')
  , f_everEV_event_lakoparkListaKattintas = ifelse(everEV_event_lakoparkListaKattintas <   0.5 , '1', '2')
  , f_wEV_event_M1X = ifelse(wEV_event_M1X <   2.5 , '1', '2')
  , f_everEV_event_lakoparkMegjelenes = ifelse(everEV_event_lakoparkMegjelenes <   1.5 , '1', '2')
  , f_everEV_referer_creativecdn = ifelse(everEV_referer_creativecdn <   2.5 , '1', '2')
  , f_wEV_channel_Paid_Search = ifelse(wEV_channel_Paid_Search <   68.5 , '1', '2')
  , f_everEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(everEV_action_M1X_talalkozottDuplikaltHirdetessel <   4.5 , '1', '2')
  , f_everEV_action_Erdeklodes_OfficeUtcapontos = ifelse(everEV_action_Erdeklodes_OfficeUtcapontos <   1.5 , '1', '2')
  , f_everEV_action_FrontendBelepes_Sikeres = ifelse(everEV_action_FrontendBelepes_Sikeres <   2.5 , '1', '2')
  , f_apEV_channel_Email = ifelse(apEV_channel_Email <   47.5 , '1', '2')
  , f_everEV_label_FrontendBelepes_Sikeres_fejlec = ifelse(everEV_label_FrontendBelepes_Sikeres_fejlec <   1.5 , '1', '2')
  , f_wEV_trafficSource_Hirdetesfigyelo_email = ifelse(wEV_trafficSource_Hirdetesfigyelo_email <   5.5 , '1', '2')
  , f_wEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(wEV_action_M1X_talalkozottDuplikaltHirdetessel <   2.5 , '1', '2')
  , f_apEV_action_Erdeklodes_OfficeNemUtcapontos = ifelse(apEV_action_Erdeklodes_OfficeNemUtcapontos <   0.5 , '1', '2')
  , f_apEV_referer_creativecdn = ifelse(apEV_referer_creativecdn <   0.5 , '1', '2')
  , f_apEV_trafficSource_Hirdetesfigyelo_email = ifelse(apEV_trafficSource_Hirdetesfigyelo_email <   18.5 , '1', '2')
  , f_everEV_action_Erdeklodes_MaganFizetos = ifelse(everEV_action_Erdeklodes_MaganFizetos <   0.5 , '1', '2')
  , f_everEV_referer_icom = ifelse(everEV_referer_icom <   8.5 , '1', '2')
  , f_everEV_action_Erdeklodes_OfficeNemUtcapontos = ifelse(everEV_action_Erdeklodes_OfficeNemUtcapontos <   3.5 , '1', '2')
  , f_everEV_trafficSource_Hirdetesfigyelo_email = ifelse(everEV_trafficSource_Hirdetesfigyelo_email <   233.5 , '1', '2')
  , f_everEV_referer_google = ifelse(everEV_referer_google <   40.5 , '1', '2')
  , f_wEV_action_Mobilreszletek_Kepgaleria = ifelse(wEV_action_Mobilreszletek_Kepgaleria <   1.5 , '1', '2')
  , f_everEV_Visits = ifelse(everEV_Visits <   49.5 , '1', '2')
  , f_everEV_Felhasznalo_Magan = ifelse(everEV_Felhasznalo_Magan <   0.5 , '1', '2')
  , f_everEV_event_Mobilreszletek = ifelse(everEV_event_Mobilreszletek <   33.5 , '1', '2')
  , f_wEV_event_Mobilreszletek = ifelse(wEV_event_Mobilreszletek <   3.5 , '1', '2')
  , f_everEV_action_Mobilreszletek_Kepgaleria = ifelse(everEV_action_Mobilreszletek_Kepgaleria <   25.5 , '1', '2')
  , f_everRE_AVG_Kepgaleria = ifelse(everRE_AVG_Kepgaleria <   0.06639 , '1', '2')
  , f_everEV_event_Piac = ifelse(everEV_event_Piac <   0.5 , '1', '2')
  , f_wRE_AVG_Kepgaleria = ifelse(wRE_AVG_Kepgaleria <   0.1235 , '1', '2')
  , f_wEV_trafficSource_rtbhouse_retargeting = ifelse(wEV_trafficSource_rtbhouse_retargeting <   0.5 , '1', '2')
  , f_everEV_channel_Direct = ifelse(everEV_channel_Direct <   92.5 , '1', '2')
  , f_wRE_SUM_Kepgaleria = ifelse(wRE_SUM_Kepgaleria <   2.5 , '1', '2')
  , f_everEV_action_Piac_undefined = ifelse(everEV_action_Piac_undefined <   0.5 , '1', '2')
  , f_wRE_MAX_Kepgaleria = ifelse(wRE_MAX_Kepgaleria <   0.5 , '1', '2')
  , f_apEV_trafficSource_rtbhouse_retargeting = ifelse(apEV_trafficSource_rtbhouse_retargeting <   0.5 , '1', '2')
  , f_apEV_action_Mobilreszletek_Kepgaleria = ifelse(apEV_action_Mobilreszletek_Kepgaleria <   1.5 , '1', '2')
  , f_apEV_event_Mobilreszletek = ifelse(apEV_event_Mobilreszletek <   12.5 , '1', '2')
  , f_apRE_AVG_Kepgaleria = ifelse(apRE_AVG_Kepgaleria <   0.2153 , '1', '2')
  , f_everEV_event_MaganhirdetoInfok = ifelse(everEV_event_MaganhirdetoInfok <   0.5 , '1', '2')
  , f_everRE_MAX_Kepgaleria = ifelse(everRE_MAX_Kepgaleria <   0.5 , '1', '2')
  , f_everEV_trafficSource_rtbhouse_retargeting = ifelse(everEV_trafficSource_rtbhouse_retargeting <   2.5 , '1', '2')
  , f_apRE_SUM_Kepgaleria = ifelse(apRE_SUM_Kepgaleria <   4.5 , '1', '2')
  , f_everRE_SUM_Kepgaleria = ifelse(everRE_SUM_Kepgaleria <   25.5 , '1', '2')
  , f_apRE_MAX_Kepgaleria = ifelse(apRE_MAX_Kepgaleria <   0.5 , '1', '2')
  , f_everEV_action_Erdeklodes_MaganIngyenes = ifelse(everEV_action_Erdeklodes_MaganIngyenes <   0.5 , '1', '2')
  , f_everRE_ikerhaz = ifelse(everRE_ikerhaz <   2.5 , '1', '2')
  , f_everEV_Videkrol_Budapestre = ifelse(everEV_Videkrol_Budapestre <   140.5 , '1', '2')
),]

```


#2 weeks performance period (f_Target)

Train
```{r}
JZ_train <- JZ_train[, `:=`(
f2w_apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   0.5 , '1', ifelse(apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <  5.5 , '2', '3'))
, f2w_everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   0.5 , '1', ifelse(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <  8.5 , '2', '3'))
, f2w_everEV_event_Mobilelhelyezkedes = ifelse(everEV_event_Mobilelhelyezkedes <   2.5 , '1', ifelse(everEV_event_Mobilelhelyezkedes <  11.5 , '2', '3'))
, f2w_apEV_pageviews = ifelse(apEV_pageviews <   668.5 , '1', ifelse(apEV_pageviews <  15270 , '2', '3'))
, f2w_apEV_HirdetesTipus_Lakossag = ifelse(apEV_HirdetesTipus_Lakossag <   73.5 , '1', ifelse(apEV_HirdetesTipus_Lakossag <  657.5 , '2', '3'))
, f2w_apEV_HirdetesEladasmodja_Elado = ifelse(apEV_HirdetesEladasmodja_Elado <   73.5 , '1', ifelse(apEV_HirdetesEladasmodja_Elado <  657.5 , '2', '3'))
, f2w_apEV_event_Mobilelhelyezkedes = ifelse(apEV_event_Mobilelhelyezkedes <   4.5 , '1', '2')
, f2w_apEV_timeOnSite = ifelse(apEV_timeOnSite <   81110 , '1', ifelse(apEV_timeOnSite <  1224000 , '2', '3'))
, f2w_everEV_event_autosuggest = ifelse(everEV_event_autosuggest <   2.5 , '1', ifelse(everEV_event_autosuggest <  18.5 , '2', '3'))
, f2w_apEV_event_autosuggest = ifelse(apEV_event_autosuggest <   2.5 , '1', '2')
, f2w_apEV_hits = ifelse(apEV_hits <   2977 , '1', ifelse(apEV_hits <  75480 , '2', '3'))
, f2w_wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   2.5 , '1', '2')
, f2w_apEV_device_desktop = ifelse(apEV_device_desktop <   96.5 , '1', ifelse(apEV_device_desktop <  464.5 , '2', '3'))
, f2w_wEV_event_Mobilelhelyezkedes = ifelse(wEV_event_Mobilelhelyezkedes <   2.5 , '1', '2')
, f2w_wEV_pageviews = ifelse(wEV_pageviews <   593 , '1', ifelse(wEV_pageviews <  11560 , '2', '3'))
, f2w_wEV_event_autosuggest = ifelse(wEV_event_autosuggest <   0.5 , '1', ifelse(wEV_event_autosuggest <  2.5 , '2', '3'))
, f2w_wEV_timeOnSite = ifelse(wEV_timeOnSite <   91650 , '1', ifelse(wEV_timeOnSite <  909400 , '2', '3'))
, f2w_wEV_event_ReszletekOldalMegjelenes = ifelse(wEV_event_ReszletekOldalMegjelenes <   9.5 , '1', ifelse(wEV_event_ReszletekOldalMegjelenes <  50.5 , '2', '3'))
, f2w_apEV_event_ReszletekOldalMegjelenes = ifelse(apEV_event_ReszletekOldalMegjelenes <   51.5 , '1', ifelse(apEV_event_ReszletekOldalMegjelenes <  105.5 , '2', '3'))
, f2w_everEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(everEV_action_Kereso_ReszletesSzukitoLenyit <   1.5 , '1', ifelse(everEV_action_Kereso_ReszletesSzukitoLenyit <  10.5 , '2', '3'))
, f2w_wEV_HirdetesEladasmodja_Elado = ifelse(wEV_HirdetesEladasmodja_Elado <   78.5 , '1', ifelse(wEV_HirdetesEladasmodja_Elado <  348.5 , '2', '3'))
, f2w_wEV_HirdetesTipus_Lakossag = ifelse(wEV_HirdetesTipus_Lakossag <   108.5 , '1', ifelse(wEV_HirdetesTipus_Lakossag <  348.5 , '2', '3'))
, f2w_everEV_event_Kereso = ifelse(everEV_event_Kereso <   1.5 , '1', ifelse(everEV_event_Kereso <  10.5 , '2', '3'))
, f2w_wEV_hits = ifelse(wEV_hits <   3444 , '1', ifelse(wEV_hits <  29580 , '2', '3'))
, f2w_apEV_Honnan_Hungary = ifelse(apEV_Honnan_Hungary <   186.5 , '1', ifelse(apEV_Honnan_Hungary <  707.5 , '2', '3'))
, f2w_wEV_device_desktop = ifelse(wEV_device_desktop <   118.5 , '1', ifelse(wEV_device_desktop <  339.5 , '2', '3'))
, f2w_apEV_event_Kereso = ifelse(apEV_event_Kereso <   2.5 , '1', '2')
, f2w_wEV_Honnan_Hungary = ifelse(wEV_Honnan_Hungary <   122.5 , '1', ifelse(wEV_Honnan_Hungary <  366.5 , '2', '3'))
, f2w_apEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(apEV_action_Kereso_ReszletesSzukitoLenyit <   2.5 , '1', '2')
, f2w_everEV_pageviews = ifelse(everEV_pageviews <   8147 , '1', ifelse(everEV_pageviews <  86420 , '2', '3'))
, f2w_apEV_event_Kepgaleria = ifelse(apEV_event_Kepgaleria <   19.5 , '1', ifelse(apEV_event_Kepgaleria <  108.5 , '2', '3'))
, f2w_wEV_Visits = ifelse(wEV_Visits <   3.5 , '1', ifelse(wEV_Visits <  20.5 , '2', '3'))
, f2w_everEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(everEV_action_Mobilelhelyezkedes_Linkethasznalt <   1.5 , '1', '2')
, f2w_everEV_HirdetesEladasmodja_Elado = ifelse(everEV_HirdetesEladasmodja_Elado <   375.5 , '1', ifelse(everEV_HirdetesEladasmodja_Elado <  1008 , '2', '3'))
, f2w_everEV_HirdetesTipus_Lakossag = ifelse(everEV_HirdetesTipus_Lakossag <   376.5 , '1', ifelse(everEV_HirdetesTipus_Lakossag <  1016 , '2', '3'))
, f2w_everEV_timeOnSite = ifelse(everEV_timeOnSite <   535900 , '1', ifelse(everEV_timeOnSite <  4351000 , '2', '3'))
, f2w_apEV_Visits = ifelse(apEV_Visits <   5.5 , '1', ifelse(apEV_Visits <  38.5 , '2', '3'))
, f2w_wEV_event_Kepgaleria = ifelse(wEV_event_Kepgaleria <   17.5 , '1', ifelse(wEV_event_Kepgaleria <  100.5 , '2', '3'))
, f2w_everEV_hits = ifelse(everEV_hits <   29230 , '1', ifelse(everEV_hits <  337100 , '2', '3'))
, f2w_apEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(apEV_action_Mobilelhelyezkedes_Linkethasznalt <   0.5 , '1', '2')
, f2w_apEV_Bejelentkezve_Magan = ifelse(apEV_Bejelentkezve_Magan <   69.5 , '1', '2')
, f2w_everEV_event_Eszkoztar = ifelse(everEV_event_Eszkoztar <   0.5 , '1', ifelse(everEV_event_Eszkoztar <  12.5 , '2', '3'))
, f2w_wEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(wEV_action_Kereso_ReszletesSzukitoLenyit <   1.5 , '1', '2')
, f2w_wEV_event_Kereso = ifelse(wEV_event_Kereso <   1.5 , '1', '2')
, f2w_apEV_referer_icom = ifelse(apEV_referer_icom <   0.5 , '1', '2')
, f2w_everEV_event_Kereses = ifelse(everEV_event_Kereses <   1.5 , '1', '2')
, f2w_everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   0.5 , '1', ifelse(everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <  3.5 , '2', '3'))
, f2w_apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   0.5 , '1', ifelse(apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <  2.5 , '2', '3'))
, f2w_everEV_device_desktop = ifelse(everEV_device_desktop <   425.5 , '1', ifelse(everEV_device_desktop <  2124 , '2', '3'))
, f2w_everEV_ReszletekOldalMegjelenes = ifelse(everEV_ReszletekOldalMegjelenes <   125.5 , '1', ifelse(everEV_ReszletekOldalMegjelenes <  419.5 , '2', '3'))
, f2w_everEV_event_ReszletekOldalMegjelenes = ifelse(everEV_event_ReszletekOldalMegjelenes <   125.5 , '1', ifelse(everEV_event_ReszletekOldalMegjelenes <  419.5 , '2', '3'))
, f2w_everEV_action_Eszkoztar_Kattintas = ifelse(everEV_action_Eszkoztar_Kattintas <   0.5 , '1', ifelse(everEV_action_Eszkoztar_Kattintas <  3.5 , '2', '3'))
, f2w_wEV_Bejelentkezve_Magan = ifelse(wEV_Bejelentkezve_Magan <   39.5 , '1', '2')
, f2w_apEV_event_Erdeklodes = ifelse(apEV_event_Erdeklodes <   0.5 , '1', '2')
, f2w_apEV_event_Nincstalalatfrontend = ifelse(apEV_event_Nincstalalatfrontend <   0.5 , '1', '2')
, f2w_everEV_Bejelentkezve_Magan = ifelse(everEV_Bejelentkezve_Magan <   558.5 , '1', '2')
, f2w_everEV_Honnan_Hungary = ifelse(everEV_Honnan_Hungary <   734.5 , '1', ifelse(everEV_Honnan_Hungary <  3696 , '2', '3'))
, f2w_everEV_event_Kepgaleria = ifelse(everEV_event_Kepgaleria <   130.5 , '1', ifelse(everEV_event_Kepgaleria <  1408 , '2', '3'))
, f2w_everEV_action_Mentes_Sikeres = ifelse(everEV_action_Mentes_Sikeres <   2.5 , '1', '2')
, f2w_wEV_event_Erdeklodes = ifelse(wEV_event_Erdeklodes <   0.5 , '1', '2')
, f2w_everEV_event_Mentes = ifelse(everEV_event_Mentes <   2.5 , '1', '2')
, f2w_wEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(wEV_action_Mobilelhelyezkedes_Linkethasznalt <   0.5 , '1', '2')
, f2w_apEV_event_Eszkoztar = ifelse(apEV_event_Eszkoztar <   1.5 , '1', '2')
, f2w_apEV_Budapestrol_Budapestre = ifelse(apEV_Budapestrol_Budapestre <   87.5 , '1', '2')
, f2w_wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   0.5 , '1', '2')
, f2w_wEV_referer_icom = ifelse(wEV_referer_icom <   0.5 , '1', '2')
, f2w_apEV_event_FrontendBelepes = ifelse(apEV_event_FrontendBelepes <   0.5 , '1', '2')
, f2w_wEV_action_Eszkoztar_Kattintas = ifelse(wEV_action_Eszkoztar_Kattintas <   0.5 , '1', '2')
, f2w_everEV_action_Kereses_ujEpitesuSzuro = ifelse(everEV_action_Kereses_ujEpitesuSzuro <   1.5 , '1', '2')
, f2w_wEV_event_Eszkoztar = ifelse(wEV_event_Eszkoztar <   0.5 , '1', '2')
, f2w_wEV_Budapestrol_Budapestre = ifelse(wEV_Budapestrol_Budapestre <   85.5 , '1', '2')
, f2w_apEV_channel_Organic_Search = ifelse(apEV_channel_Organic_Search <   74.5 , '1', '2')
, f2w_apEV_action_Eszkoztar_Kattintas = ifelse(apEV_action_Eszkoztar_Kattintas <   1.5 , '1', '2')
, f2w_apEV_action_FrontendBelepes_Megjelenes = ifelse(apEV_action_FrontendBelepes_Megjelenes <   0.5 , '1', '2')
, f2w_apEV_trafficSource_google_organic = ifelse(apEV_trafficSource_google_organic <   74.5 , '1', '2')
, f2w_apEV_event_M1X = ifelse(apEV_event_M1X <   1.5 , '1', '2')
, f2w_everEV_socialNetwork_Facebook = ifelse(everEV_socialNetwork_Facebook <   343.5 , '1', '2')
, f2w_apEV_referer_google = ifelse(apEV_referer_google <   1.5 , '1', ifelse(apEV_referer_google <  14.5 , '2', '3'))
, f2w_everEV_SocialSourceReferral = ifelse(everEV_SocialSourceReferral <   343.5 , '1', '2')
, f2w_everEV_channel_Social = ifelse(everEV_channel_Social <   343.5 , '1', '2')
, f2w_everEV_action_FrontendBelepes_Megjelenes = ifelse(everEV_action_FrontendBelepes_Megjelenes <   4.5 , '1', '2')
, f2w_everEV_event_FrontendBelepes = ifelse(everEV_event_FrontendBelepes <   7.5 , '1', '2')
, f2w_everEV_trafficSource_facebook_referral = ifelse(everEV_trafficSource_facebook_referral <   343.5 , '1', '2')
, f2w_everEV_event_Nincstalalatfrontend = ifelse(everEV_event_Nincstalalatfrontend <   3.5 , '1', '2')
, f2w_wEV_event_FrontendBelepes = ifelse(wEV_event_FrontendBelepes <   0.5 , '1', '2')
, f2w_everEV_channel_Organic_Search = ifelse(everEV_channel_Organic_Search <   485.5 , '1', '2')
, f2w_everEV_trafficSource_google_organic = ifelse(everEV_trafficSource_google_organic <   523.5 , '1', '2')
, f2w_everEV_Budapestrol_Budapestre = ifelse(everEV_Budapestrol_Budapestre <   401.5 , '1', '2')
, f2w_everEV_event_lakoparkMegjelenes = ifelse(everEV_event_lakoparkMegjelenes <   1.5 , '1', '2')
, f2w_wEV_trafficSource_facebook_referral = ifelse(wEV_trafficSource_facebook_referral <   37.5 , '1', '2')
, f2w_wEV_socialNetwork_Facebook = ifelse(wEV_socialNetwork_Facebook <   37.5 , '1', '2')
, f2w_wEV_SocialSourceReferral = ifelse(wEV_SocialSourceReferral <   37.5 , '1', '2')
, f2w_wEV_channel_Social = ifelse(wEV_channel_Social <   37.5 , '1', '2')
, f2w_everEV_action_Kereses_mobil = ifelse(everEV_action_Kereses_mobil <   0.5 , '1', '2')
, f2w_everEV_label_FrontendBelepes_Megjelenes_fejlec = ifelse(everEV_label_FrontendBelepes_Megjelenes_fejlec <   2.5 , '1', '2')
, f2w_everEV_event_Erdeklodes = ifelse(everEV_event_Erdeklodes <   0.5 , '1', '2')
, f2w_everEV_event_M1X = ifelse(everEV_event_M1X <   6.5 , '1', '2')
, f2w_everEV_channel_Email = ifelse(everEV_channel_Email <   382.5 , '1', '2')
, f2w_apEV_trafficSource_google_cpc = ifelse(apEV_trafficSource_google_cpc <   63.5 , '1', '2')
, f2w_apEV_trafficSource_facebook_referral = ifelse(apEV_trafficSource_facebook_referral <   62.5 , '1', '2')
, f2w_apEV_socialNetwork_Facebook = ifelse(apEV_socialNetwork_Facebook <   62.5 , '1', '2')
, f2w_apEV_SocialSourceReferral = ifelse(apEV_SocialSourceReferral <   62.5 , '1', '2')
, f2w_apEV_channel_Social = ifelse(apEV_channel_Social <   62.5 , '1', '2')
, f2w_wEV_event_M1X = ifelse(wEV_event_M1X <   1.5 , '1', '2')
, f2w_apEV_channel_Paid_Search = ifelse(apEV_channel_Paid_Search <   63.5 , '1', '2')
, f2w_apEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(apEV_action_M1X_talalkozottDuplikaltHirdetessel <   1.5 , '1', '2')
, f2w_apEV_action_Erdeklodes_OfficeUtcapontos = ifelse(apEV_action_Erdeklodes_OfficeUtcapontos <   0.5 , '1', '2')
, f2w_everEV_trafficSource_Hirdetesfigyelo_email = ifelse(everEV_trafficSource_Hirdetesfigyelo_email <   232.5 , '1', '2')
, f2w_wEV_channel_Email = ifelse(wEV_channel_Email <   11.5 , '1', '2')
, f2w_wEV_channel_Organic_Search = ifelse(wEV_channel_Organic_Search <   60.5 , '1', '2')
, f2w_apEV_channel_Email = ifelse(apEV_channel_Email <   46.5 , '1', '2')
, f2w_wEV_referer_google = ifelse(wEV_referer_google <   7.5 , '1', '2')
, f2w_wEV_trafficSource_google_organic = ifelse(wEV_trafficSource_google_organic <   55.5 , '1', '2')
, f2w_everEV_event_AI = ifelse(everEV_event_AI <   0.5 , '1', '2')
, f2w_apEV_action_Erdeklodes_OfficeNemUtcapontos = ifelse(apEV_action_Erdeklodes_OfficeNemUtcapontos <   0.5 , '1', '2')
, f2w_everEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(everEV_action_M1X_talalkozottDuplikaltHirdetessel <   7.5 , '1', '2')
, f2w_everEV_action_FrontendBelepes_Sikeres = ifelse(everEV_action_FrontendBelepes_Sikeres <   2.5 , '1', '2')
, f2w_everEV_label_FrontendBelepes_Sikeres_fejlec = ifelse(everEV_label_FrontendBelepes_Sikeres_fejlec <   1.5 , '1', '2')
, f2w_everEV_trafficSource_google_cpc = ifelse(everEV_trafficSource_google_cpc <   301.5 , '1', '2')
, f2w_wEV_trafficSource_Hirdetesfigyelo_email = ifelse(wEV_trafficSource_Hirdetesfigyelo_email <   5.5 , '1', '2')
, f2w_everEV_channel_Paid_Search = ifelse(everEV_channel_Paid_Search <   301.5 , '1', '2')
, f2w_everEV_event_OfficeLogo = ifelse(everEV_event_OfficeLogo <   0.5 , '1', '2')
, f2w_everEV_event_lakoparkListaKattintas = ifelse(everEV_event_lakoparkListaKattintas <   0.5 , '1', '2')
, f2w_wEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(wEV_action_M1X_talalkozottDuplikaltHirdetessel <   1.5 , '1', '2')
, f2w_apEV_trafficSource_Hirdetesfigyelo_email = ifelse(apEV_trafficSource_Hirdetesfigyelo_email <   13.5 , '1', '2')
, f2w_everEV_referer_icom = ifelse(everEV_referer_icom <   9.5 , '1', '2')
, f2w_wEV_trafficSource_google_cpc = ifelse(wEV_trafficSource_google_cpc <   69.5 , '1', '2')
, f2w_everEV_Felhasznalo_Magan = ifelse(everEV_Felhasznalo_Magan <   0.5 , '1', '2')
, f2w_everEV_event_Piac = ifelse(everEV_event_Piac <   0.5 , '1', '2')
, f2w_everEV_action_Piac_undefined = ifelse(everEV_action_Piac_undefined <   0.5 , '1', '2')
, f2w_everEV_referer_google = ifelse(everEV_referer_google <   10.5 , '1', '2')
, f2w_everEV_channel_Direct = ifelse(everEV_channel_Direct <   97.5 , '1', '2')
, f2w_everRE_AVG_Kepgaleria = ifelse(everRE_AVG_Kepgaleria <   0.06057 , '1', '2')
, f2w_wEV_channel_Paid_Search = ifelse(wEV_channel_Paid_Search <   68.5 , '1', '2')
, f2w_everRE_MAX_Kepgaleria = ifelse(everRE_MAX_Kepgaleria <   0.5 , '1', '2')
, f2w_everRE_SUM_Kepgaleria = ifelse(everRE_SUM_Kepgaleria <   0.5 , '1', '2')
, f2w_everRE_megye_Budapest = ifelse(everRE_megye_Budapest <   5.5 , '1', '2')
, f2w_everEV_event_Mobilreszletek = ifelse(everEV_event_Mobilreszletek <   35.5 , '1', '2')
, f2w_apRE_AVG_Kepgaleria = ifelse(apRE_AVG_Kepgaleria <   0.04447 , '1', '2')
, f2w_apRE_MAX_Kepgaleria = ifelse(apRE_MAX_Kepgaleria <   0.5 , '1', '2')
, f2w_apRE_SUM_Kepgaleria = ifelse(apRE_SUM_Kepgaleria <   0.5 , '1', '2')
, f2w_apEV_action_Mobilreszletek_Kepgaleria = ifelse(apEV_action_Mobilreszletek_Kepgaleria <   0.5 , '1', '2')
, f2w_everEV_action_Erdeklodes_OfficeUtcapontos = ifelse(everEV_action_Erdeklodes_OfficeUtcapontos <   1.5 , '1', '2')
, f2w_apEV_trafficSource_rtbhouse_retargeting = ifelse(apEV_trafficSource_rtbhouse_retargeting <   0.5 , '1', '2')
, f2w_everRE_ikerhaz = ifelse(everRE_ikerhaz <   2.5 , '1', '2')
, f2w_wEV_action_Mobilreszletek_Kepgaleria = ifelse(wEV_action_Mobilreszletek_Kepgaleria <   0.5 , '1', '2')
, f2w_wEV_trafficSource_rtbhouse_retargeting = ifelse(wEV_trafficSource_rtbhouse_retargeting <   0.5 , '1', '2')
, f2w_everEV_Visits = ifelse(everEV_Visits <   49.5 , '1', '2')
, f2w_everEV_action_Mobilreszletek_Kepgaleria = ifelse(everEV_action_Mobilreszletek_Kepgaleria <   31.5 , '1', '2')
, f2w_apEV_event_Mobilreszletek = ifelse(apEV_event_Mobilreszletek <   1.5 , '1', '2')
, f2w_everEV_action_Erdeklodes_OfficeNemUtcapontos = ifelse(everEV_action_Erdeklodes_OfficeNemUtcapontos <   3.5 , '1', '2')
, f2w_wEV_event_Mobilreszletek = ifelse(wEV_event_Mobilreszletek <   1.5 , '1', '2')
, f2w_everEV_action_Erdeklodes_MaganFizetos = ifelse(everEV_action_Erdeklodes_MaganFizetos <   1.5 , '1', '2')
, f2w_wRE_AVG_Kepgaleria = ifelse(wRE_AVG_Kepgaleria <   0.07353 , '1', '2')
, f2w_wEV_referer_creativecdn = ifelse(wEV_referer_creativecdn <   0.5 , '1', '2')
, f2w_wRE_AVG_Hirdetesar = ifelse(wRE_AVG_Hirdetesar <   16440000 , '1', '2')
, f2w_wRE_MAX_Kepgaleria = ifelse(wRE_MAX_Kepgaleria <   0.5 , '1', '2')
, f2w_wRE_SUM_Kepgaleria = ifelse(wRE_SUM_Kepgaleria <   0.5 , '1', '2')
, f2w_apRE_megye_Budapest = ifelse(apRE_megye_Budapest <   2.5 , '1', '2')
, f2w_everRE_megye_Pest = ifelse(everRE_megye_Pest <   4.5 , '1', '2')
, f2w_everEV_event_MaganhirdetoInfok = ifelse(everEV_event_MaganhirdetoInfok <   1.5 , '1', '2')
, f2w_wRE_nyaralo = ifelse(wRE_nyaralo <   0.5 , '1', '2')
, f2w_everRE_ujszeru = ifelse(everRE_ujszeru <   6.5 , '1', '2')
, f2w_apRE_AVG_Hirdetesar = ifelse(apRE_AVG_Hirdetesar <   16600000 , '1', '2')
, f2w_apEV_referer_creativecdn = ifelse(apEV_referer_creativecdn <   0.5 , '1', '2')
, f2w_wRE_megye_Budapest = ifelse(wRE_megye_Budapest <   1.5 , '1', '2')
, f2w_everEV_referer_creativecdn = ifelse(everEV_referer_creativecdn <   2.5 , '1', '2')
, f2w_everEV_Videkrol_Budapestre = ifelse(everEV_Videkrol_Budapestre <   140.5 , '1', '2')
, f2w_apEV_Videkrol_Budapestre = ifelse(apEV_Videkrol_Budapestre <   19.5 , '1', '2')
, f2w_wRE_MAX_Hirdetesar = ifelse(wRE_MAX_Hirdetesar <   18500000 , '1', '2')
, f2w_everRE_ker_11 = ifelse(everRE_ker_11 <   0.5 , '1', '2')
, f2w_everEV_channel_Display = ifelse(everEV_channel_Display <   0.5 , '1', '2')
, f2w_everEV_MAX_visitNumber = ifelse(everEV_MAX_visitNumber <   50.5 , '1', '2')
, f2w_everEV_MIN_visitNumber = ifelse(everEV_MIN_visitNumber <   1.5 , '1', '2')
, f2w_wRE_MAX_DIFF_Visit_per_RE = ifelse(wRE_MAX_DIFF_Visit_per_RE <   1.5 , '1', '2')
, f2w_everRE_AVG_Hirdetesar = ifelse(everRE_AVG_Hirdetesar <   20890000 , '1', '2')
, f2w_wRE_AVG_DIFF_Visit_per_RE = ifelse(wRE_AVG_DIFF_Visit_per_RE <   1.181 , '1', '2')
, f2w_everEV_channel_Referral = ifelse(everEV_channel_Referral <   117.5 , '1', '2')
, f2w_everEV_trafficSource_rtbhouse_retargeting = ifelse(everEV_trafficSource_rtbhouse_retargeting <   2.5 , '1', '2')
, f2w_everRE_tegla = ifelse(everRE_tegla <   6.5 , '1', '2')
, f2w_apEV_event_MaganhirdetoInfok = ifelse(apEV_event_MaganhirdetoInfok <   0.5 , '1', '2')
, f2w_everEV_action_Erdeklodes_MaganIngyenes = ifelse(everEV_action_Erdeklodes_MaganIngyenes <   0.5 , '1', '2')
, f2w_wEV_Videkrol_Budapestre = ifelse(wEV_Videkrol_Budapestre <   10.5 , '1', '2')
, f2w_apRE_tegla = ifelse(apRE_tegla <   5.5 , '1', '2')
, f2w_everRE_felujitott = ifelse(everRE_felujitott <   9.5 , '1', '2')
, f2w_everRE_lakas = ifelse(everRE_lakas <   8.5 , '1', '2')
, f2w_apRE_megye_Pest = ifelse(apRE_megye_Pest <   1.5 , '1', '2')
, f2w_apRE_lakas = ifelse(apRE_lakas <   7.5 , '1', '2')
, f2w_everEV_action_MaganToDo_Megjelenes = ifelse(everEV_action_MaganToDo_Megjelenes <   3.5 , '1', '2')
, f2w_everRE_SUM_DIFF_Visit_per_RE = ifelse(everRE_SUM_DIFF_Visit_per_RE <   97.5 , '1', '2')
  ),]
```

Test
```{r}
JZ_test <- JZ_test[, `:=`(
f2w_apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   0.5 , '1', ifelse(apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <  5.5 , '2', '3'))
, f2w_everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   0.5 , '1', ifelse(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <  8.5 , '2', '3'))
, f2w_everEV_event_Mobilelhelyezkedes = ifelse(everEV_event_Mobilelhelyezkedes <   2.5 , '1', ifelse(everEV_event_Mobilelhelyezkedes <  11.5 , '2', '3'))
, f2w_apEV_pageviews = ifelse(apEV_pageviews <   668.5 , '1', ifelse(apEV_pageviews <  15270 , '2', '3'))
, f2w_apEV_HirdetesTipus_Lakossag = ifelse(apEV_HirdetesTipus_Lakossag <   73.5 , '1', ifelse(apEV_HirdetesTipus_Lakossag <  657.5 , '2', '3'))
, f2w_apEV_HirdetesEladasmodja_Elado = ifelse(apEV_HirdetesEladasmodja_Elado <   73.5 , '1', ifelse(apEV_HirdetesEladasmodja_Elado <  657.5 , '2', '3'))
, f2w_apEV_event_Mobilelhelyezkedes = ifelse(apEV_event_Mobilelhelyezkedes <   4.5 , '1', '2')
, f2w_apEV_timeOnSite = ifelse(apEV_timeOnSite <   81110 , '1', ifelse(apEV_timeOnSite <  1224000 , '2', '3'))
, f2w_everEV_event_autosuggest = ifelse(everEV_event_autosuggest <   2.5 , '1', ifelse(everEV_event_autosuggest <  18.5 , '2', '3'))
, f2w_apEV_event_autosuggest = ifelse(apEV_event_autosuggest <   2.5 , '1', '2')
, f2w_apEV_hits = ifelse(apEV_hits <   2977 , '1', ifelse(apEV_hits <  75480 , '2', '3'))
, f2w_wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   2.5 , '1', '2')
, f2w_apEV_device_desktop = ifelse(apEV_device_desktop <   96.5 , '1', ifelse(apEV_device_desktop <  464.5 , '2', '3'))
, f2w_wEV_event_Mobilelhelyezkedes = ifelse(wEV_event_Mobilelhelyezkedes <   2.5 , '1', '2')
, f2w_wEV_pageviews = ifelse(wEV_pageviews <   593 , '1', ifelse(wEV_pageviews <  11560 , '2', '3'))
, f2w_wEV_event_autosuggest = ifelse(wEV_event_autosuggest <   0.5 , '1', ifelse(wEV_event_autosuggest <  2.5 , '2', '3'))
, f2w_wEV_timeOnSite = ifelse(wEV_timeOnSite <   91650 , '1', ifelse(wEV_timeOnSite <  909400 , '2', '3'))
, f2w_wEV_event_ReszletekOldalMegjelenes = ifelse(wEV_event_ReszletekOldalMegjelenes <   9.5 , '1', ifelse(wEV_event_ReszletekOldalMegjelenes <  50.5 , '2', '3'))
, f2w_apEV_event_ReszletekOldalMegjelenes = ifelse(apEV_event_ReszletekOldalMegjelenes <   51.5 , '1', ifelse(apEV_event_ReszletekOldalMegjelenes <  105.5 , '2', '3'))
, f2w_everEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(everEV_action_Kereso_ReszletesSzukitoLenyit <   1.5 , '1', ifelse(everEV_action_Kereso_ReszletesSzukitoLenyit <  10.5 , '2', '3'))
, f2w_wEV_HirdetesEladasmodja_Elado = ifelse(wEV_HirdetesEladasmodja_Elado <   78.5 , '1', ifelse(wEV_HirdetesEladasmodja_Elado <  348.5 , '2', '3'))
, f2w_wEV_HirdetesTipus_Lakossag = ifelse(wEV_HirdetesTipus_Lakossag <   108.5 , '1', ifelse(wEV_HirdetesTipus_Lakossag <  348.5 , '2', '3'))
, f2w_everEV_event_Kereso = ifelse(everEV_event_Kereso <   1.5 , '1', ifelse(everEV_event_Kereso <  10.5 , '2', '3'))
, f2w_wEV_hits = ifelse(wEV_hits <   3444 , '1', ifelse(wEV_hits <  29580 , '2', '3'))
, f2w_apEV_Honnan_Hungary = ifelse(apEV_Honnan_Hungary <   186.5 , '1', ifelse(apEV_Honnan_Hungary <  707.5 , '2', '3'))
, f2w_wEV_device_desktop = ifelse(wEV_device_desktop <   118.5 , '1', ifelse(wEV_device_desktop <  339.5 , '2', '3'))
, f2w_apEV_event_Kereso = ifelse(apEV_event_Kereso <   2.5 , '1', '2')
, f2w_wEV_Honnan_Hungary = ifelse(wEV_Honnan_Hungary <   122.5 , '1', ifelse(wEV_Honnan_Hungary <  366.5 , '2', '3'))
, f2w_apEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(apEV_action_Kereso_ReszletesSzukitoLenyit <   2.5 , '1', '2')
, f2w_everEV_pageviews = ifelse(everEV_pageviews <   8147 , '1', ifelse(everEV_pageviews <  86420 , '2', '3'))
, f2w_apEV_event_Kepgaleria = ifelse(apEV_event_Kepgaleria <   19.5 , '1', ifelse(apEV_event_Kepgaleria <  108.5 , '2', '3'))
, f2w_wEV_Visits = ifelse(wEV_Visits <   3.5 , '1', ifelse(wEV_Visits <  20.5 , '2', '3'))
, f2w_everEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(everEV_action_Mobilelhelyezkedes_Linkethasznalt <   1.5 , '1', '2')
, f2w_everEV_HirdetesEladasmodja_Elado = ifelse(everEV_HirdetesEladasmodja_Elado <   375.5 , '1', ifelse(everEV_HirdetesEladasmodja_Elado <  1008 , '2', '3'))
, f2w_everEV_HirdetesTipus_Lakossag = ifelse(everEV_HirdetesTipus_Lakossag <   376.5 , '1', ifelse(everEV_HirdetesTipus_Lakossag <  1016 , '2', '3'))
, f2w_everEV_timeOnSite = ifelse(everEV_timeOnSite <   535900 , '1', ifelse(everEV_timeOnSite <  4351000 , '2', '3'))
, f2w_apEV_Visits = ifelse(apEV_Visits <   5.5 , '1', ifelse(apEV_Visits <  38.5 , '2', '3'))
, f2w_wEV_event_Kepgaleria = ifelse(wEV_event_Kepgaleria <   17.5 , '1', ifelse(wEV_event_Kepgaleria <  100.5 , '2', '3'))
, f2w_everEV_hits = ifelse(everEV_hits <   29230 , '1', ifelse(everEV_hits <  337100 , '2', '3'))
, f2w_apEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(apEV_action_Mobilelhelyezkedes_Linkethasznalt <   0.5 , '1', '2')
, f2w_apEV_Bejelentkezve_Magan = ifelse(apEV_Bejelentkezve_Magan <   69.5 , '1', '2')
, f2w_everEV_event_Eszkoztar = ifelse(everEV_event_Eszkoztar <   0.5 , '1', ifelse(everEV_event_Eszkoztar <  12.5 , '2', '3'))
, f2w_wEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(wEV_action_Kereso_ReszletesSzukitoLenyit <   1.5 , '1', '2')
, f2w_wEV_event_Kereso = ifelse(wEV_event_Kereso <   1.5 , '1', '2')
, f2w_apEV_referer_icom = ifelse(apEV_referer_icom <   0.5 , '1', '2')
, f2w_everEV_event_Kereses = ifelse(everEV_event_Kereses <   1.5 , '1', '2')
, f2w_everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   0.5 , '1', ifelse(everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <  3.5 , '2', '3'))
, f2w_apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   0.5 , '1', ifelse(apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <  2.5 , '2', '3'))
, f2w_everEV_device_desktop = ifelse(everEV_device_desktop <   425.5 , '1', ifelse(everEV_device_desktop <  2124 , '2', '3'))
, f2w_everEV_ReszletekOldalMegjelenes = ifelse(everEV_ReszletekOldalMegjelenes <   125.5 , '1', ifelse(everEV_ReszletekOldalMegjelenes <  419.5 , '2', '3'))
, f2w_everEV_event_ReszletekOldalMegjelenes = ifelse(everEV_event_ReszletekOldalMegjelenes <   125.5 , '1', ifelse(everEV_event_ReszletekOldalMegjelenes <  419.5 , '2', '3'))
, f2w_everEV_action_Eszkoztar_Kattintas = ifelse(everEV_action_Eszkoztar_Kattintas <   0.5 , '1', ifelse(everEV_action_Eszkoztar_Kattintas <  3.5 , '2', '3'))
, f2w_wEV_Bejelentkezve_Magan = ifelse(wEV_Bejelentkezve_Magan <   39.5 , '1', '2')
, f2w_apEV_event_Erdeklodes = ifelse(apEV_event_Erdeklodes <   0.5 , '1', '2')
, f2w_apEV_event_Nincstalalatfrontend = ifelse(apEV_event_Nincstalalatfrontend <   0.5 , '1', '2')
, f2w_everEV_Bejelentkezve_Magan = ifelse(everEV_Bejelentkezve_Magan <   558.5 , '1', '2')
, f2w_everEV_Honnan_Hungary = ifelse(everEV_Honnan_Hungary <   734.5 , '1', ifelse(everEV_Honnan_Hungary <  3696 , '2', '3'))
, f2w_everEV_event_Kepgaleria = ifelse(everEV_event_Kepgaleria <   130.5 , '1', ifelse(everEV_event_Kepgaleria <  1408 , '2', '3'))
, f2w_everEV_action_Mentes_Sikeres = ifelse(everEV_action_Mentes_Sikeres <   2.5 , '1', '2')
, f2w_wEV_event_Erdeklodes = ifelse(wEV_event_Erdeklodes <   0.5 , '1', '2')
, f2w_everEV_event_Mentes = ifelse(everEV_event_Mentes <   2.5 , '1', '2')
, f2w_wEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(wEV_action_Mobilelhelyezkedes_Linkethasznalt <   0.5 , '1', '2')
, f2w_apEV_event_Eszkoztar = ifelse(apEV_event_Eszkoztar <   1.5 , '1', '2')
, f2w_apEV_Budapestrol_Budapestre = ifelse(apEV_Budapestrol_Budapestre <   87.5 , '1', '2')
, f2w_wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   0.5 , '1', '2')
, f2w_wEV_referer_icom = ifelse(wEV_referer_icom <   0.5 , '1', '2')
, f2w_apEV_event_FrontendBelepes = ifelse(apEV_event_FrontendBelepes <   0.5 , '1', '2')
, f2w_wEV_action_Eszkoztar_Kattintas = ifelse(wEV_action_Eszkoztar_Kattintas <   0.5 , '1', '2')
, f2w_everEV_action_Kereses_ujEpitesuSzuro = ifelse(everEV_action_Kereses_ujEpitesuSzuro <   1.5 , '1', '2')
, f2w_wEV_event_Eszkoztar = ifelse(wEV_event_Eszkoztar <   0.5 , '1', '2')
, f2w_wEV_Budapestrol_Budapestre = ifelse(wEV_Budapestrol_Budapestre <   85.5 , '1', '2')
, f2w_apEV_channel_Organic_Search = ifelse(apEV_channel_Organic_Search <   74.5 , '1', '2')
, f2w_apEV_action_Eszkoztar_Kattintas = ifelse(apEV_action_Eszkoztar_Kattintas <   1.5 , '1', '2')
, f2w_apEV_action_FrontendBelepes_Megjelenes = ifelse(apEV_action_FrontendBelepes_Megjelenes <   0.5 , '1', '2')
, f2w_apEV_trafficSource_google_organic = ifelse(apEV_trafficSource_google_organic <   74.5 , '1', '2')
, f2w_apEV_event_M1X = ifelse(apEV_event_M1X <   1.5 , '1', '2')
, f2w_everEV_socialNetwork_Facebook = ifelse(everEV_socialNetwork_Facebook <   343.5 , '1', '2')
, f2w_apEV_referer_google = ifelse(apEV_referer_google <   1.5 , '1', ifelse(apEV_referer_google <  14.5 , '2', '3'))
, f2w_everEV_SocialSourceReferral = ifelse(everEV_SocialSourceReferral <   343.5 , '1', '2')
, f2w_everEV_channel_Social = ifelse(everEV_channel_Social <   343.5 , '1', '2')
, f2w_everEV_action_FrontendBelepes_Megjelenes = ifelse(everEV_action_FrontendBelepes_Megjelenes <   4.5 , '1', '2')
, f2w_everEV_event_FrontendBelepes = ifelse(everEV_event_FrontendBelepes <   7.5 , '1', '2')
, f2w_everEV_trafficSource_facebook_referral = ifelse(everEV_trafficSource_facebook_referral <   343.5 , '1', '2')
, f2w_everEV_event_Nincstalalatfrontend = ifelse(everEV_event_Nincstalalatfrontend <   3.5 , '1', '2')
, f2w_wEV_event_FrontendBelepes = ifelse(wEV_event_FrontendBelepes <   0.5 , '1', '2')
, f2w_everEV_channel_Organic_Search = ifelse(everEV_channel_Organic_Search <   485.5 , '1', '2')
, f2w_everEV_trafficSource_google_organic = ifelse(everEV_trafficSource_google_organic <   523.5 , '1', '2')
, f2w_everEV_Budapestrol_Budapestre = ifelse(everEV_Budapestrol_Budapestre <   401.5 , '1', '2')
, f2w_everEV_event_lakoparkMegjelenes = ifelse(everEV_event_lakoparkMegjelenes <   1.5 , '1', '2')
, f2w_wEV_trafficSource_facebook_referral = ifelse(wEV_trafficSource_facebook_referral <   37.5 , '1', '2')
, f2w_wEV_socialNetwork_Facebook = ifelse(wEV_socialNetwork_Facebook <   37.5 , '1', '2')
, f2w_wEV_SocialSourceReferral = ifelse(wEV_SocialSourceReferral <   37.5 , '1', '2')
, f2w_wEV_channel_Social = ifelse(wEV_channel_Social <   37.5 , '1', '2')
, f2w_everEV_action_Kereses_mobil = ifelse(everEV_action_Kereses_mobil <   0.5 , '1', '2')
, f2w_everEV_label_FrontendBelepes_Megjelenes_fejlec = ifelse(everEV_label_FrontendBelepes_Megjelenes_fejlec <   2.5 , '1', '2')
, f2w_everEV_event_Erdeklodes = ifelse(everEV_event_Erdeklodes <   0.5 , '1', '2')
, f2w_everEV_event_M1X = ifelse(everEV_event_M1X <   6.5 , '1', '2')
, f2w_everEV_channel_Email = ifelse(everEV_channel_Email <   382.5 , '1', '2')
, f2w_apEV_trafficSource_google_cpc = ifelse(apEV_trafficSource_google_cpc <   63.5 , '1', '2')
, f2w_apEV_trafficSource_facebook_referral = ifelse(apEV_trafficSource_facebook_referral <   62.5 , '1', '2')
, f2w_apEV_socialNetwork_Facebook = ifelse(apEV_socialNetwork_Facebook <   62.5 , '1', '2')
, f2w_apEV_SocialSourceReferral = ifelse(apEV_SocialSourceReferral <   62.5 , '1', '2')
, f2w_apEV_channel_Social = ifelse(apEV_channel_Social <   62.5 , '1', '2')
, f2w_wEV_event_M1X = ifelse(wEV_event_M1X <   1.5 , '1', '2')
, f2w_apEV_channel_Paid_Search = ifelse(apEV_channel_Paid_Search <   63.5 , '1', '2')
, f2w_apEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(apEV_action_M1X_talalkozottDuplikaltHirdetessel <   1.5 , '1', '2')
, f2w_apEV_action_Erdeklodes_OfficeUtcapontos = ifelse(apEV_action_Erdeklodes_OfficeUtcapontos <   0.5 , '1', '2')
, f2w_everEV_trafficSource_Hirdetesfigyelo_email = ifelse(everEV_trafficSource_Hirdetesfigyelo_email <   232.5 , '1', '2')
, f2w_wEV_channel_Email = ifelse(wEV_channel_Email <   11.5 , '1', '2')
, f2w_wEV_channel_Organic_Search = ifelse(wEV_channel_Organic_Search <   60.5 , '1', '2')
, f2w_apEV_channel_Email = ifelse(apEV_channel_Email <   46.5 , '1', '2')
, f2w_wEV_referer_google = ifelse(wEV_referer_google <   7.5 , '1', '2')
, f2w_wEV_trafficSource_google_organic = ifelse(wEV_trafficSource_google_organic <   55.5 , '1', '2')
, f2w_everEV_event_AI = ifelse(everEV_event_AI <   0.5 , '1', '2')
, f2w_apEV_action_Erdeklodes_OfficeNemUtcapontos = ifelse(apEV_action_Erdeklodes_OfficeNemUtcapontos <   0.5 , '1', '2')
, f2w_everEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(everEV_action_M1X_talalkozottDuplikaltHirdetessel <   7.5 , '1', '2')
, f2w_everEV_action_FrontendBelepes_Sikeres = ifelse(everEV_action_FrontendBelepes_Sikeres <   2.5 , '1', '2')
, f2w_everEV_label_FrontendBelepes_Sikeres_fejlec = ifelse(everEV_label_FrontendBelepes_Sikeres_fejlec <   1.5 , '1', '2')
, f2w_everEV_trafficSource_google_cpc = ifelse(everEV_trafficSource_google_cpc <   301.5 , '1', '2')
, f2w_wEV_trafficSource_Hirdetesfigyelo_email = ifelse(wEV_trafficSource_Hirdetesfigyelo_email <   5.5 , '1', '2')
, f2w_everEV_channel_Paid_Search = ifelse(everEV_channel_Paid_Search <   301.5 , '1', '2')
, f2w_everEV_event_OfficeLogo = ifelse(everEV_event_OfficeLogo <   0.5 , '1', '2')
, f2w_everEV_event_lakoparkListaKattintas = ifelse(everEV_event_lakoparkListaKattintas <   0.5 , '1', '2')
, f2w_wEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(wEV_action_M1X_talalkozottDuplikaltHirdetessel <   1.5 , '1', '2')
, f2w_apEV_trafficSource_Hirdetesfigyelo_email = ifelse(apEV_trafficSource_Hirdetesfigyelo_email <   13.5 , '1', '2')
, f2w_everEV_referer_icom = ifelse(everEV_referer_icom <   9.5 , '1', '2')
, f2w_wEV_trafficSource_google_cpc = ifelse(wEV_trafficSource_google_cpc <   69.5 , '1', '2')
, f2w_everEV_Felhasznalo_Magan = ifelse(everEV_Felhasznalo_Magan <   0.5 , '1', '2')
, f2w_everEV_event_Piac = ifelse(everEV_event_Piac <   0.5 , '1', '2')
, f2w_everEV_action_Piac_undefined = ifelse(everEV_action_Piac_undefined <   0.5 , '1', '2')
, f2w_everEV_referer_google = ifelse(everEV_referer_google <   10.5 , '1', '2')
, f2w_everEV_channel_Direct = ifelse(everEV_channel_Direct <   97.5 , '1', '2')
, f2w_everRE_AVG_Kepgaleria = ifelse(everRE_AVG_Kepgaleria <   0.06057 , '1', '2')
, f2w_wEV_channel_Paid_Search = ifelse(wEV_channel_Paid_Search <   68.5 , '1', '2')
, f2w_everRE_MAX_Kepgaleria = ifelse(everRE_MAX_Kepgaleria <   0.5 , '1', '2')
, f2w_everRE_SUM_Kepgaleria = ifelse(everRE_SUM_Kepgaleria <   0.5 , '1', '2')
, f2w_everRE_megye_Budapest = ifelse(everRE_megye_Budapest <   5.5 , '1', '2')
, f2w_everEV_event_Mobilreszletek = ifelse(everEV_event_Mobilreszletek <   35.5 , '1', '2')
, f2w_apRE_AVG_Kepgaleria = ifelse(apRE_AVG_Kepgaleria <   0.04447 , '1', '2')
, f2w_apRE_MAX_Kepgaleria = ifelse(apRE_MAX_Kepgaleria <   0.5 , '1', '2')
, f2w_apRE_SUM_Kepgaleria = ifelse(apRE_SUM_Kepgaleria <   0.5 , '1', '2')
, f2w_apEV_action_Mobilreszletek_Kepgaleria = ifelse(apEV_action_Mobilreszletek_Kepgaleria <   0.5 , '1', '2')
, f2w_everEV_action_Erdeklodes_OfficeUtcapontos = ifelse(everEV_action_Erdeklodes_OfficeUtcapontos <   1.5 , '1', '2')
, f2w_apEV_trafficSource_rtbhouse_retargeting = ifelse(apEV_trafficSource_rtbhouse_retargeting <   0.5 , '1', '2')
, f2w_everRE_ikerhaz = ifelse(everRE_ikerhaz <   2.5 , '1', '2')
, f2w_wEV_action_Mobilreszletek_Kepgaleria = ifelse(wEV_action_Mobilreszletek_Kepgaleria <   0.5 , '1', '2')
, f2w_wEV_trafficSource_rtbhouse_retargeting = ifelse(wEV_trafficSource_rtbhouse_retargeting <   0.5 , '1', '2')
, f2w_everEV_Visits = ifelse(everEV_Visits <   49.5 , '1', '2')
, f2w_everEV_action_Mobilreszletek_Kepgaleria = ifelse(everEV_action_Mobilreszletek_Kepgaleria <   31.5 , '1', '2')
, f2w_apEV_event_Mobilreszletek = ifelse(apEV_event_Mobilreszletek <   1.5 , '1', '2')
, f2w_everEV_action_Erdeklodes_OfficeNemUtcapontos = ifelse(everEV_action_Erdeklodes_OfficeNemUtcapontos <   3.5 , '1', '2')
, f2w_wEV_event_Mobilreszletek = ifelse(wEV_event_Mobilreszletek <   1.5 , '1', '2')
, f2w_everEV_action_Erdeklodes_MaganFizetos = ifelse(everEV_action_Erdeklodes_MaganFizetos <   1.5 , '1', '2')
, f2w_wRE_AVG_Kepgaleria = ifelse(wRE_AVG_Kepgaleria <   0.07353 , '1', '2')
, f2w_wEV_referer_creativecdn = ifelse(wEV_referer_creativecdn <   0.5 , '1', '2')
, f2w_wRE_AVG_Hirdetesar = ifelse(wRE_AVG_Hirdetesar <   16440000 , '1', '2')
, f2w_wRE_MAX_Kepgaleria = ifelse(wRE_MAX_Kepgaleria <   0.5 , '1', '2')
, f2w_wRE_SUM_Kepgaleria = ifelse(wRE_SUM_Kepgaleria <   0.5 , '1', '2')
, f2w_apRE_megye_Budapest = ifelse(apRE_megye_Budapest <   2.5 , '1', '2')
, f2w_everRE_megye_Pest = ifelse(everRE_megye_Pest <   4.5 , '1', '2')
, f2w_everEV_event_MaganhirdetoInfok = ifelse(everEV_event_MaganhirdetoInfok <   1.5 , '1', '2')
, f2w_wRE_nyaralo = ifelse(wRE_nyaralo <   0.5 , '1', '2')
, f2w_everRE_ujszeru = ifelse(everRE_ujszeru <   6.5 , '1', '2')
, f2w_apRE_AVG_Hirdetesar = ifelse(apRE_AVG_Hirdetesar <   16600000 , '1', '2')
, f2w_apEV_referer_creativecdn = ifelse(apEV_referer_creativecdn <   0.5 , '1', '2')
, f2w_wRE_megye_Budapest = ifelse(wRE_megye_Budapest <   1.5 , '1', '2')
, f2w_everEV_referer_creativecdn = ifelse(everEV_referer_creativecdn <   2.5 , '1', '2')
, f2w_everEV_Videkrol_Budapestre = ifelse(everEV_Videkrol_Budapestre <   140.5 , '1', '2')
, f2w_apEV_Videkrol_Budapestre = ifelse(apEV_Videkrol_Budapestre <   19.5 , '1', '2')
, f2w_wRE_MAX_Hirdetesar = ifelse(wRE_MAX_Hirdetesar <   18500000 , '1', '2')
, f2w_everRE_ker_11 = ifelse(everRE_ker_11 <   0.5 , '1', '2')
, f2w_everEV_channel_Display = ifelse(everEV_channel_Display <   0.5 , '1', '2')
, f2w_everEV_MAX_visitNumber = ifelse(everEV_MAX_visitNumber <   50.5 , '1', '2')
, f2w_everEV_MIN_visitNumber = ifelse(everEV_MIN_visitNumber <   1.5 , '1', '2')
, f2w_wRE_MAX_DIFF_Visit_per_RE = ifelse(wRE_MAX_DIFF_Visit_per_RE <   1.5 , '1', '2')
, f2w_everRE_AVG_Hirdetesar = ifelse(everRE_AVG_Hirdetesar <   20890000 , '1', '2')
, f2w_wRE_AVG_DIFF_Visit_per_RE = ifelse(wRE_AVG_DIFF_Visit_per_RE <   1.181 , '1', '2')
, f2w_everEV_channel_Referral = ifelse(everEV_channel_Referral <   117.5 , '1', '2')
, f2w_everEV_trafficSource_rtbhouse_retargeting = ifelse(everEV_trafficSource_rtbhouse_retargeting <   2.5 , '1', '2')
, f2w_everRE_tegla = ifelse(everRE_tegla <   6.5 , '1', '2')
, f2w_apEV_event_MaganhirdetoInfok = ifelse(apEV_event_MaganhirdetoInfok <   0.5 , '1', '2')
, f2w_everEV_action_Erdeklodes_MaganIngyenes = ifelse(everEV_action_Erdeklodes_MaganIngyenes <   0.5 , '1', '2')
, f2w_wEV_Videkrol_Budapestre = ifelse(wEV_Videkrol_Budapestre <   10.5 , '1', '2')
, f2w_apRE_tegla = ifelse(apRE_tegla <   5.5 , '1', '2')
, f2w_everRE_felujitott = ifelse(everRE_felujitott <   9.5 , '1', '2')
, f2w_everRE_lakas = ifelse(everRE_lakas <   8.5 , '1', '2')
, f2w_apRE_megye_Pest = ifelse(apRE_megye_Pest <   1.5 , '1', '2')
, f2w_apRE_lakas = ifelse(apRE_lakas <   7.5 , '1', '2')
, f2w_everEV_action_MaganToDo_Megjelenes = ifelse(everEV_action_MaganToDo_Megjelenes <   3.5 , '1', '2')
, f2w_everRE_SUM_DIFF_Visit_per_RE = ifelse(everRE_SUM_DIFF_Visit_per_RE <   97.5 , '1', '2')
  ),]
```

External
```{r}
JZ_external <- JZ_external[, `:=`(
f2w_apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   0.5 , '1', ifelse(apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <  5.5 , '2', '3'))
, f2w_everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   0.5 , '1', ifelse(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <  8.5 , '2', '3'))
, f2w_everEV_event_Mobilelhelyezkedes = ifelse(everEV_event_Mobilelhelyezkedes <   2.5 , '1', ifelse(everEV_event_Mobilelhelyezkedes <  11.5 , '2', '3'))
, f2w_apEV_pageviews = ifelse(apEV_pageviews <   668.5 , '1', ifelse(apEV_pageviews <  15270 , '2', '3'))
, f2w_apEV_HirdetesTipus_Lakossag = ifelse(apEV_HirdetesTipus_Lakossag <   73.5 , '1', ifelse(apEV_HirdetesTipus_Lakossag <  657.5 , '2', '3'))
, f2w_apEV_HirdetesEladasmodja_Elado = ifelse(apEV_HirdetesEladasmodja_Elado <   73.5 , '1', ifelse(apEV_HirdetesEladasmodja_Elado <  657.5 , '2', '3'))
, f2w_apEV_event_Mobilelhelyezkedes = ifelse(apEV_event_Mobilelhelyezkedes <   4.5 , '1', '2')
, f2w_apEV_timeOnSite = ifelse(apEV_timeOnSite <   81110 , '1', ifelse(apEV_timeOnSite <  1224000 , '2', '3'))
, f2w_everEV_event_autosuggest = ifelse(everEV_event_autosuggest <   2.5 , '1', ifelse(everEV_event_autosuggest <  18.5 , '2', '3'))
, f2w_apEV_event_autosuggest = ifelse(apEV_event_autosuggest <   2.5 , '1', '2')
, f2w_apEV_hits = ifelse(apEV_hits <   2977 , '1', ifelse(apEV_hits <  75480 , '2', '3'))
, f2w_wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = ifelse(wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik <   2.5 , '1', '2')
, f2w_apEV_device_desktop = ifelse(apEV_device_desktop <   96.5 , '1', ifelse(apEV_device_desktop <  464.5 , '2', '3'))
, f2w_wEV_event_Mobilelhelyezkedes = ifelse(wEV_event_Mobilelhelyezkedes <   2.5 , '1', '2')
, f2w_wEV_pageviews = ifelse(wEV_pageviews <   593 , '1', ifelse(wEV_pageviews <  11560 , '2', '3'))
, f2w_wEV_event_autosuggest = ifelse(wEV_event_autosuggest <   0.5 , '1', ifelse(wEV_event_autosuggest <  2.5 , '2', '3'))
, f2w_wEV_timeOnSite = ifelse(wEV_timeOnSite <   91650 , '1', ifelse(wEV_timeOnSite <  909400 , '2', '3'))
, f2w_wEV_event_ReszletekOldalMegjelenes = ifelse(wEV_event_ReszletekOldalMegjelenes <   9.5 , '1', ifelse(wEV_event_ReszletekOldalMegjelenes <  50.5 , '2', '3'))
, f2w_apEV_event_ReszletekOldalMegjelenes = ifelse(apEV_event_ReszletekOldalMegjelenes <   51.5 , '1', ifelse(apEV_event_ReszletekOldalMegjelenes <  105.5 , '2', '3'))
, f2w_everEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(everEV_action_Kereso_ReszletesSzukitoLenyit <   1.5 , '1', ifelse(everEV_action_Kereso_ReszletesSzukitoLenyit <  10.5 , '2', '3'))
, f2w_wEV_HirdetesEladasmodja_Elado = ifelse(wEV_HirdetesEladasmodja_Elado <   78.5 , '1', ifelse(wEV_HirdetesEladasmodja_Elado <  348.5 , '2', '3'))
, f2w_wEV_HirdetesTipus_Lakossag = ifelse(wEV_HirdetesTipus_Lakossag <   108.5 , '1', ifelse(wEV_HirdetesTipus_Lakossag <  348.5 , '2', '3'))
, f2w_everEV_event_Kereso = ifelse(everEV_event_Kereso <   1.5 , '1', ifelse(everEV_event_Kereso <  10.5 , '2', '3'))
, f2w_wEV_hits = ifelse(wEV_hits <   3444 , '1', ifelse(wEV_hits <  29580 , '2', '3'))
, f2w_apEV_Honnan_Hungary = ifelse(apEV_Honnan_Hungary <   186.5 , '1', ifelse(apEV_Honnan_Hungary <  707.5 , '2', '3'))
, f2w_wEV_device_desktop = ifelse(wEV_device_desktop <   118.5 , '1', ifelse(wEV_device_desktop <  339.5 , '2', '3'))
, f2w_apEV_event_Kereso = ifelse(apEV_event_Kereso <   2.5 , '1', '2')
, f2w_wEV_Honnan_Hungary = ifelse(wEV_Honnan_Hungary <   122.5 , '1', ifelse(wEV_Honnan_Hungary <  366.5 , '2', '3'))
, f2w_apEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(apEV_action_Kereso_ReszletesSzukitoLenyit <   2.5 , '1', '2')
, f2w_everEV_pageviews = ifelse(everEV_pageviews <   8147 , '1', ifelse(everEV_pageviews <  86420 , '2', '3'))
, f2w_apEV_event_Kepgaleria = ifelse(apEV_event_Kepgaleria <   19.5 , '1', ifelse(apEV_event_Kepgaleria <  108.5 , '2', '3'))
, f2w_wEV_Visits = ifelse(wEV_Visits <   3.5 , '1', ifelse(wEV_Visits <  20.5 , '2', '3'))
, f2w_everEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(everEV_action_Mobilelhelyezkedes_Linkethasznalt <   1.5 , '1', '2')
, f2w_everEV_HirdetesEladasmodja_Elado = ifelse(everEV_HirdetesEladasmodja_Elado <   375.5 , '1', ifelse(everEV_HirdetesEladasmodja_Elado <  1008 , '2', '3'))
, f2w_everEV_HirdetesTipus_Lakossag = ifelse(everEV_HirdetesTipus_Lakossag <   376.5 , '1', ifelse(everEV_HirdetesTipus_Lakossag <  1016 , '2', '3'))
, f2w_everEV_timeOnSite = ifelse(everEV_timeOnSite <   535900 , '1', ifelse(everEV_timeOnSite <  4351000 , '2', '3'))
, f2w_apEV_Visits = ifelse(apEV_Visits <   5.5 , '1', ifelse(apEV_Visits <  38.5 , '2', '3'))
, f2w_wEV_event_Kepgaleria = ifelse(wEV_event_Kepgaleria <   17.5 , '1', ifelse(wEV_event_Kepgaleria <  100.5 , '2', '3'))
, f2w_everEV_hits = ifelse(everEV_hits <   29230 , '1', ifelse(everEV_hits <  337100 , '2', '3'))
, f2w_apEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(apEV_action_Mobilelhelyezkedes_Linkethasznalt <   0.5 , '1', '2')
, f2w_apEV_Bejelentkezve_Magan = ifelse(apEV_Bejelentkezve_Magan <   69.5 , '1', '2')
, f2w_everEV_event_Eszkoztar = ifelse(everEV_event_Eszkoztar <   0.5 , '1', ifelse(everEV_event_Eszkoztar <  12.5 , '2', '3'))
, f2w_wEV_action_Kereso_ReszletesSzukitoLenyit = ifelse(wEV_action_Kereso_ReszletesSzukitoLenyit <   1.5 , '1', '2')
, f2w_wEV_event_Kereso = ifelse(wEV_event_Kereso <   1.5 , '1', '2')
, f2w_apEV_referer_icom = ifelse(apEV_referer_icom <   0.5 , '1', '2')
, f2w_everEV_event_Kereses = ifelse(everEV_event_Kereses <   1.5 , '1', '2')
, f2w_everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   0.5 , '1', ifelse(everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <  3.5 , '2', '3'))
, f2w_apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   0.5 , '1', ifelse(apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <  2.5 , '2', '3'))
, f2w_everEV_device_desktop = ifelse(everEV_device_desktop <   425.5 , '1', ifelse(everEV_device_desktop <  2124 , '2', '3'))
, f2w_everEV_ReszletekOldalMegjelenes = ifelse(everEV_ReszletekOldalMegjelenes <   125.5 , '1', ifelse(everEV_ReszletekOldalMegjelenes <  419.5 , '2', '3'))
, f2w_everEV_event_ReszletekOldalMegjelenes = ifelse(everEV_event_ReszletekOldalMegjelenes <   125.5 , '1', ifelse(everEV_event_ReszletekOldalMegjelenes <  419.5 , '2', '3'))
, f2w_everEV_action_Eszkoztar_Kattintas = ifelse(everEV_action_Eszkoztar_Kattintas <   0.5 , '1', ifelse(everEV_action_Eszkoztar_Kattintas <  3.5 , '2', '3'))
, f2w_wEV_Bejelentkezve_Magan = ifelse(wEV_Bejelentkezve_Magan <   39.5 , '1', '2')
, f2w_apEV_event_Erdeklodes = ifelse(apEV_event_Erdeklodes <   0.5 , '1', '2')
, f2w_apEV_event_Nincstalalatfrontend = ifelse(apEV_event_Nincstalalatfrontend <   0.5 , '1', '2')
, f2w_everEV_Bejelentkezve_Magan = ifelse(everEV_Bejelentkezve_Magan <   558.5 , '1', '2')
, f2w_everEV_Honnan_Hungary = ifelse(everEV_Honnan_Hungary <   734.5 , '1', ifelse(everEV_Honnan_Hungary <  3696 , '2', '3'))
, f2w_everEV_event_Kepgaleria = ifelse(everEV_event_Kepgaleria <   130.5 , '1', ifelse(everEV_event_Kepgaleria <  1408 , '2', '3'))
, f2w_everEV_action_Mentes_Sikeres = ifelse(everEV_action_Mentes_Sikeres <   2.5 , '1', '2')
, f2w_wEV_event_Erdeklodes = ifelse(wEV_event_Erdeklodes <   0.5 , '1', '2')
, f2w_everEV_event_Mentes = ifelse(everEV_event_Mentes <   2.5 , '1', '2')
, f2w_wEV_action_Mobilelhelyezkedes_Linkethasznalt = ifelse(wEV_action_Mobilelhelyezkedes_Linkethasznalt <   0.5 , '1', '2')
, f2w_apEV_event_Eszkoztar = ifelse(apEV_event_Eszkoztar <   1.5 , '1', '2')
, f2w_apEV_Budapestrol_Budapestre = ifelse(apEV_Budapestrol_Budapestre <   87.5 , '1', '2')
, f2w_wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = ifelse(wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest <   0.5 , '1', '2')
, f2w_wEV_referer_icom = ifelse(wEV_referer_icom <   0.5 , '1', '2')
, f2w_apEV_event_FrontendBelepes = ifelse(apEV_event_FrontendBelepes <   0.5 , '1', '2')
, f2w_wEV_action_Eszkoztar_Kattintas = ifelse(wEV_action_Eszkoztar_Kattintas <   0.5 , '1', '2')
, f2w_everEV_action_Kereses_ujEpitesuSzuro = ifelse(everEV_action_Kereses_ujEpitesuSzuro <   1.5 , '1', '2')
, f2w_wEV_event_Eszkoztar = ifelse(wEV_event_Eszkoztar <   0.5 , '1', '2')
, f2w_wEV_Budapestrol_Budapestre = ifelse(wEV_Budapestrol_Budapestre <   85.5 , '1', '2')
, f2w_apEV_channel_Organic_Search = ifelse(apEV_channel_Organic_Search <   74.5 , '1', '2')
, f2w_apEV_action_Eszkoztar_Kattintas = ifelse(apEV_action_Eszkoztar_Kattintas <   1.5 , '1', '2')
, f2w_apEV_action_FrontendBelepes_Megjelenes = ifelse(apEV_action_FrontendBelepes_Megjelenes <   0.5 , '1', '2')
, f2w_apEV_trafficSource_google_organic = ifelse(apEV_trafficSource_google_organic <   74.5 , '1', '2')
, f2w_apEV_event_M1X = ifelse(apEV_event_M1X <   1.5 , '1', '2')
, f2w_everEV_socialNetwork_Facebook = ifelse(everEV_socialNetwork_Facebook <   343.5 , '1', '2')
, f2w_apEV_referer_google = ifelse(apEV_referer_google <   1.5 , '1', ifelse(apEV_referer_google <  14.5 , '2', '3'))
, f2w_everEV_SocialSourceReferral = ifelse(everEV_SocialSourceReferral <   343.5 , '1', '2')
, f2w_everEV_channel_Social = ifelse(everEV_channel_Social <   343.5 , '1', '2')
, f2w_everEV_action_FrontendBelepes_Megjelenes = ifelse(everEV_action_FrontendBelepes_Megjelenes <   4.5 , '1', '2')
, f2w_everEV_event_FrontendBelepes = ifelse(everEV_event_FrontendBelepes <   7.5 , '1', '2')
, f2w_everEV_trafficSource_facebook_referral = ifelse(everEV_trafficSource_facebook_referral <   343.5 , '1', '2')
, f2w_everEV_event_Nincstalalatfrontend = ifelse(everEV_event_Nincstalalatfrontend <   3.5 , '1', '2')
, f2w_wEV_event_FrontendBelepes = ifelse(wEV_event_FrontendBelepes <   0.5 , '1', '2')
, f2w_everEV_channel_Organic_Search = ifelse(everEV_channel_Organic_Search <   485.5 , '1', '2')
, f2w_everEV_trafficSource_google_organic = ifelse(everEV_trafficSource_google_organic <   523.5 , '1', '2')
, f2w_everEV_Budapestrol_Budapestre = ifelse(everEV_Budapestrol_Budapestre <   401.5 , '1', '2')
, f2w_everEV_event_lakoparkMegjelenes = ifelse(everEV_event_lakoparkMegjelenes <   1.5 , '1', '2')
, f2w_wEV_trafficSource_facebook_referral = ifelse(wEV_trafficSource_facebook_referral <   37.5 , '1', '2')
, f2w_wEV_socialNetwork_Facebook = ifelse(wEV_socialNetwork_Facebook <   37.5 , '1', '2')
, f2w_wEV_SocialSourceReferral = ifelse(wEV_SocialSourceReferral <   37.5 , '1', '2')
, f2w_wEV_channel_Social = ifelse(wEV_channel_Social <   37.5 , '1', '2')
, f2w_everEV_action_Kereses_mobil = ifelse(everEV_action_Kereses_mobil <   0.5 , '1', '2')
, f2w_everEV_label_FrontendBelepes_Megjelenes_fejlec = ifelse(everEV_label_FrontendBelepes_Megjelenes_fejlec <   2.5 , '1', '2')
, f2w_everEV_event_Erdeklodes = ifelse(everEV_event_Erdeklodes <   0.5 , '1', '2')
, f2w_everEV_event_M1X = ifelse(everEV_event_M1X <   6.5 , '1', '2')
, f2w_everEV_channel_Email = ifelse(everEV_channel_Email <   382.5 , '1', '2')
, f2w_apEV_trafficSource_google_cpc = ifelse(apEV_trafficSource_google_cpc <   63.5 , '1', '2')
, f2w_apEV_trafficSource_facebook_referral = ifelse(apEV_trafficSource_facebook_referral <   62.5 , '1', '2')
, f2w_apEV_socialNetwork_Facebook = ifelse(apEV_socialNetwork_Facebook <   62.5 , '1', '2')
, f2w_apEV_SocialSourceReferral = ifelse(apEV_SocialSourceReferral <   62.5 , '1', '2')
, f2w_apEV_channel_Social = ifelse(apEV_channel_Social <   62.5 , '1', '2')
, f2w_wEV_event_M1X = ifelse(wEV_event_M1X <   1.5 , '1', '2')
, f2w_apEV_channel_Paid_Search = ifelse(apEV_channel_Paid_Search <   63.5 , '1', '2')
, f2w_apEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(apEV_action_M1X_talalkozottDuplikaltHirdetessel <   1.5 , '1', '2')
, f2w_apEV_action_Erdeklodes_OfficeUtcapontos = ifelse(apEV_action_Erdeklodes_OfficeUtcapontos <   0.5 , '1', '2')
, f2w_everEV_trafficSource_Hirdetesfigyelo_email = ifelse(everEV_trafficSource_Hirdetesfigyelo_email <   232.5 , '1', '2')
, f2w_wEV_channel_Email = ifelse(wEV_channel_Email <   11.5 , '1', '2')
, f2w_wEV_channel_Organic_Search = ifelse(wEV_channel_Organic_Search <   60.5 , '1', '2')
, f2w_apEV_channel_Email = ifelse(apEV_channel_Email <   46.5 , '1', '2')
, f2w_wEV_referer_google = ifelse(wEV_referer_google <   7.5 , '1', '2')
, f2w_wEV_trafficSource_google_organic = ifelse(wEV_trafficSource_google_organic <   55.5 , '1', '2')
, f2w_everEV_event_AI = ifelse(everEV_event_AI <   0.5 , '1', '2')
, f2w_apEV_action_Erdeklodes_OfficeNemUtcapontos = ifelse(apEV_action_Erdeklodes_OfficeNemUtcapontos <   0.5 , '1', '2')
, f2w_everEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(everEV_action_M1X_talalkozottDuplikaltHirdetessel <   7.5 , '1', '2')
, f2w_everEV_action_FrontendBelepes_Sikeres = ifelse(everEV_action_FrontendBelepes_Sikeres <   2.5 , '1', '2')
, f2w_everEV_label_FrontendBelepes_Sikeres_fejlec = ifelse(everEV_label_FrontendBelepes_Sikeres_fejlec <   1.5 , '1', '2')
, f2w_everEV_trafficSource_google_cpc = ifelse(everEV_trafficSource_google_cpc <   301.5 , '1', '2')
, f2w_wEV_trafficSource_Hirdetesfigyelo_email = ifelse(wEV_trafficSource_Hirdetesfigyelo_email <   5.5 , '1', '2')
, f2w_everEV_channel_Paid_Search = ifelse(everEV_channel_Paid_Search <   301.5 , '1', '2')
, f2w_everEV_event_OfficeLogo = ifelse(everEV_event_OfficeLogo <   0.5 , '1', '2')
, f2w_everEV_event_lakoparkListaKattintas = ifelse(everEV_event_lakoparkListaKattintas <   0.5 , '1', '2')
, f2w_wEV_action_M1X_talalkozottDuplikaltHirdetessel = ifelse(wEV_action_M1X_talalkozottDuplikaltHirdetessel <   1.5 , '1', '2')
, f2w_apEV_trafficSource_Hirdetesfigyelo_email = ifelse(apEV_trafficSource_Hirdetesfigyelo_email <   13.5 , '1', '2')
, f2w_everEV_referer_icom = ifelse(everEV_referer_icom <   9.5 , '1', '2')
, f2w_wEV_trafficSource_google_cpc = ifelse(wEV_trafficSource_google_cpc <   69.5 , '1', '2')
, f2w_everEV_Felhasznalo_Magan = ifelse(everEV_Felhasznalo_Magan <   0.5 , '1', '2')
, f2w_everEV_event_Piac = ifelse(everEV_event_Piac <   0.5 , '1', '2')
, f2w_everEV_action_Piac_undefined = ifelse(everEV_action_Piac_undefined <   0.5 , '1', '2')
, f2w_everEV_referer_google = ifelse(everEV_referer_google <   10.5 , '1', '2')
, f2w_everEV_channel_Direct = ifelse(everEV_channel_Direct <   97.5 , '1', '2')
, f2w_everRE_AVG_Kepgaleria = ifelse(everRE_AVG_Kepgaleria <   0.06057 , '1', '2')
, f2w_wEV_channel_Paid_Search = ifelse(wEV_channel_Paid_Search <   68.5 , '1', '2')
, f2w_everRE_MAX_Kepgaleria = ifelse(everRE_MAX_Kepgaleria <   0.5 , '1', '2')
, f2w_everRE_SUM_Kepgaleria = ifelse(everRE_SUM_Kepgaleria <   0.5 , '1', '2')
, f2w_everRE_megye_Budapest = ifelse(everRE_megye_Budapest <   5.5 , '1', '2')
, f2w_everEV_event_Mobilreszletek = ifelse(everEV_event_Mobilreszletek <   35.5 , '1', '2')
, f2w_apRE_AVG_Kepgaleria = ifelse(apRE_AVG_Kepgaleria <   0.04447 , '1', '2')
, f2w_apRE_MAX_Kepgaleria = ifelse(apRE_MAX_Kepgaleria <   0.5 , '1', '2')
, f2w_apRE_SUM_Kepgaleria = ifelse(apRE_SUM_Kepgaleria <   0.5 , '1', '2')
, f2w_apEV_action_Mobilreszletek_Kepgaleria = ifelse(apEV_action_Mobilreszletek_Kepgaleria <   0.5 , '1', '2')
, f2w_everEV_action_Erdeklodes_OfficeUtcapontos = ifelse(everEV_action_Erdeklodes_OfficeUtcapontos <   1.5 , '1', '2')
, f2w_apEV_trafficSource_rtbhouse_retargeting = ifelse(apEV_trafficSource_rtbhouse_retargeting <   0.5 , '1', '2')
, f2w_everRE_ikerhaz = ifelse(everRE_ikerhaz <   2.5 , '1', '2')
, f2w_wEV_action_Mobilreszletek_Kepgaleria = ifelse(wEV_action_Mobilreszletek_Kepgaleria <   0.5 , '1', '2')
, f2w_wEV_trafficSource_rtbhouse_retargeting = ifelse(wEV_trafficSource_rtbhouse_retargeting <   0.5 , '1', '2')
, f2w_everEV_Visits = ifelse(everEV_Visits <   49.5 , '1', '2')
, f2w_everEV_action_Mobilreszletek_Kepgaleria = ifelse(everEV_action_Mobilreszletek_Kepgaleria <   31.5 , '1', '2')
, f2w_apEV_event_Mobilreszletek = ifelse(apEV_event_Mobilreszletek <   1.5 , '1', '2')
, f2w_everEV_action_Erdeklodes_OfficeNemUtcapontos = ifelse(everEV_action_Erdeklodes_OfficeNemUtcapontos <   3.5 , '1', '2')
, f2w_wEV_event_Mobilreszletek = ifelse(wEV_event_Mobilreszletek <   1.5 , '1', '2')
, f2w_everEV_action_Erdeklodes_MaganFizetos = ifelse(everEV_action_Erdeklodes_MaganFizetos <   1.5 , '1', '2')
, f2w_wRE_AVG_Kepgaleria = ifelse(wRE_AVG_Kepgaleria <   0.07353 , '1', '2')
, f2w_wEV_referer_creativecdn = ifelse(wEV_referer_creativecdn <   0.5 , '1', '2')
, f2w_wRE_AVG_Hirdetesar = ifelse(wRE_AVG_Hirdetesar <   16440000 , '1', '2')
, f2w_wRE_MAX_Kepgaleria = ifelse(wRE_MAX_Kepgaleria <   0.5 , '1', '2')
, f2w_wRE_SUM_Kepgaleria = ifelse(wRE_SUM_Kepgaleria <   0.5 , '1', '2')
, f2w_apRE_megye_Budapest = ifelse(apRE_megye_Budapest <   2.5 , '1', '2')
, f2w_everRE_megye_Pest = ifelse(everRE_megye_Pest <   4.5 , '1', '2')
, f2w_everEV_event_MaganhirdetoInfok = ifelse(everEV_event_MaganhirdetoInfok <   1.5 , '1', '2')
, f2w_wRE_nyaralo = ifelse(wRE_nyaralo <   0.5 , '1', '2')
, f2w_everRE_ujszeru = ifelse(everRE_ujszeru <   6.5 , '1', '2')
, f2w_apRE_AVG_Hirdetesar = ifelse(apRE_AVG_Hirdetesar <   16600000 , '1', '2')
, f2w_apEV_referer_creativecdn = ifelse(apEV_referer_creativecdn <   0.5 , '1', '2')
, f2w_wRE_megye_Budapest = ifelse(wRE_megye_Budapest <   1.5 , '1', '2')
, f2w_everEV_referer_creativecdn = ifelse(everEV_referer_creativecdn <   2.5 , '1', '2')
, f2w_everEV_Videkrol_Budapestre = ifelse(everEV_Videkrol_Budapestre <   140.5 , '1', '2')
, f2w_apEV_Videkrol_Budapestre = ifelse(apEV_Videkrol_Budapestre <   19.5 , '1', '2')
, f2w_wRE_MAX_Hirdetesar = ifelse(wRE_MAX_Hirdetesar <   18500000 , '1', '2')
, f2w_everRE_ker_11 = ifelse(everRE_ker_11 <   0.5 , '1', '2')
, f2w_everEV_channel_Display = ifelse(everEV_channel_Display <   0.5 , '1', '2')
, f2w_everEV_MAX_visitNumber = ifelse(everEV_MAX_visitNumber <   50.5 , '1', '2')
, f2w_everEV_MIN_visitNumber = ifelse(everEV_MIN_visitNumber <   1.5 , '1', '2')
, f2w_wRE_MAX_DIFF_Visit_per_RE = ifelse(wRE_MAX_DIFF_Visit_per_RE <   1.5 , '1', '2')
, f2w_everRE_AVG_Hirdetesar = ifelse(everRE_AVG_Hirdetesar <   20890000 , '1', '2')
, f2w_wRE_AVG_DIFF_Visit_per_RE = ifelse(wRE_AVG_DIFF_Visit_per_RE <   1.181 , '1', '2')
, f2w_everEV_channel_Referral = ifelse(everEV_channel_Referral <   117.5 , '1', '2')
, f2w_everEV_trafficSource_rtbhouse_retargeting = ifelse(everEV_trafficSource_rtbhouse_retargeting <   2.5 , '1', '2')
, f2w_everRE_tegla = ifelse(everRE_tegla <   6.5 , '1', '2')
, f2w_apEV_event_MaganhirdetoInfok = ifelse(apEV_event_MaganhirdetoInfok <   0.5 , '1', '2')
, f2w_everEV_action_Erdeklodes_MaganIngyenes = ifelse(everEV_action_Erdeklodes_MaganIngyenes <   0.5 , '1', '2')
, f2w_wEV_Videkrol_Budapestre = ifelse(wEV_Videkrol_Budapestre <   10.5 , '1', '2')
, f2w_apRE_tegla = ifelse(apRE_tegla <   5.5 , '1', '2')
, f2w_everRE_felujitott = ifelse(everRE_felujitott <   9.5 , '1', '2')
, f2w_everRE_lakas = ifelse(everRE_lakas <   8.5 , '1', '2')
, f2w_apRE_megye_Pest = ifelse(apRE_megye_Pest <   1.5 , '1', '2')
, f2w_apRE_lakas = ifelse(apRE_lakas <   7.5 , '1', '2')
, f2w_everEV_action_MaganToDo_Megjelenes = ifelse(everEV_action_MaganToDo_Megjelenes <   3.5 , '1', '2')
, f2w_everRE_SUM_DIFF_Visit_per_RE = ifelse(everRE_SUM_DIFF_Visit_per_RE <   97.5 , '1', '2')
  ),]
```



# Variables 1 week

```{r}
var_10_factor_simple <- c("f_apEV_event_autosuggest", "f_everEV_event_Mobilelhelyezkedes", "f_everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik", "f_apEV_event_Mobilelhelyezkedes", "f_everEV_event_autosuggest", "f_wEV_event_autosuggest", "f_wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik", "f_apEV_timeOnSite", "f_apEV_pageviews")

var_11_factor_all <- c(names(JZ_train)[grep("^f_.*", names(JZ_train))])
var_11_dropped <- c("f_Target", "f_Target_2w")
var_11_factor_all <- var_11_factor_all[! var_11_factor_all %in% var_11_dropped]

```

# Variables 2 weeks

```{r}
var_10_factor_simple_2w <- c("f2w_apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik", "f2w_everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik", "f2w_everEV_event_Mobilelhelyezkedes", "f2w_apEV_pageviews", "f2w_apEV_HirdetesTipus_Lakossag", "f2w_apEV_HirdetesEladasmodja_Elado", "f2w_apEV_event_Mobilelhelyezkedes", "f2w_apEV_timeOnSite", "f2w_everEV_event_autosuggest", "f2w_apEV_event_autosuggest", "f2w_apEV_hits")

var_11_factor_all_2w <- c(names(JZ_train)[grep("^f2w_.*", names(JZ_train))])
var_11_dropped_2w <- c("f_Target", "f_Target_2w")
var_11_factor_all_2w <- var_11_factor_all_2w[! var_11_factor_all_2w %in% var_11_dropped_2w]

```
