==> 4_The_Final_Model/03.3_Feature_Engineering_Log.Rmd <==
#1 week performance period (f_Target)

Train
```{r}
JZ_train <- JZ_train[, `:=`(
  log_apEV_event_autosuggest = log(apEV_event_autosuggest + 1)
  , log_everEV_event_Mobilelhelyezkedes = log(everEV_event_Mobilelhelyezkedes + 1)
  , log_everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
  , log_apEV_event_Mobilelhelyezkedes = log(apEV_event_Mobilelhelyezkedes + 1)
  , log_everEV_event_autosuggest = log(everEV_event_autosuggest + 1)
  , log_wEV_event_autosuggest = log(wEV_event_autosuggest + 1)
  , log_wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
  , log_apEV_timeOnSite = log(apEV_timeOnSite + 1)
  , log_apEV_pageviews = log(apEV_pageviews + 1)
  , log_apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
  , log_wEV_event_Mobilelhelyezkedes = log(wEV_event_Mobilelhelyezkedes + 1)
  , log_apEV_HirdetesTipus_Lakossag = log(apEV_HirdetesTipus_Lakossag + 1)
  , log_apEV_HirdetesEladasmodja_Elado = log(apEV_HirdetesEladasmodja_Elado + 1)
  , log_apEV_device_desktop = log(apEV_device_desktop + 1)
  , log_apEV_hits = log(apEV_hits + 1)
  , log_wEV_timeOnSite = log(wEV_timeOnSite + 1)
  , log_wEV_HirdetesTipus_Lakossag = log(wEV_HirdetesTipus_Lakossag + 1)
  , log_apEV_Honnan_Hungary = log(apEV_Honnan_Hungary + 1)
  , log_wEV_HirdetesEladasmodja_Elado = log(wEV_HirdetesEladasmodja_Elado + 1)
  , log_apEV_event_ReszletekOldalMegjelenes = log(apEV_event_ReszletekOldalMegjelenes + 1)
  , log_wEV_device_desktop = log(wEV_device_desktop + 1)
  , log_wEV_pageviews = log(wEV_pageviews + 1)
  , log_wEV_Honnan_Hungary = log(wEV_Honnan_Hungary + 1)
  , log_everEV_action_Kereso_ReszletesSzukitoLenyit = log(everEV_action_Kereso_ReszletesSzukitoLenyit + 1)
  , log_apEV_event_Kereso = log(apEV_event_Kereso + 1)
  , log_everEV_event_Kereso = log(everEV_event_Kereso + 1)
  , log_wEV_hits = log(wEV_hits + 1)
  , log_apEV_action_Kereso_ReszletesSzukitoLenyit = log(apEV_action_Kereso_ReszletesSzukitoLenyit + 1)
  , log_wEV_event_Kepgaleria = log(wEV_event_Kepgaleria + 1)
  , log_wEV_event_ReszletekOldalMegjelenes = log(wEV_event_ReszletekOldalMegjelenes + 1)
  , log_apEV_event_Kepgaleria = log(apEV_event_Kepgaleria + 1)
  , log_apEV_Visits = log(apEV_Visits + 1)
  , log_everEV_action_Mobilelhelyezkedes_Linkethasznalt = log(everEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
  , log_everEV_hits = log(everEV_hits + 1)
  , log_everEV_HirdetesEladasmodja_Elado = log(everEV_HirdetesEladasmodja_Elado + 1)
  , log_apEV_action_Mobilelhelyezkedes_Linkethasznalt = log(apEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
  , log_everEV_event_Kereses = log(everEV_event_Kereses + 1)
  , log_wEV_Visits = log(wEV_Visits + 1)
  , log_everEV_HirdetesTipus_Lakossag = log(everEV_HirdetesTipus_Lakossag + 1)
  , log_apEV_event_Nincstalalatfrontend = log(apEV_event_Nincstalalatfrontend + 1)
  , log_wEV_action_Kereso_ReszletesSzukitoLenyit = log(wEV_action_Kereso_ReszletesSzukitoLenyit + 1)
  , log_wEV_event_Kereso = log(wEV_event_Kereso + 1)
  , log_apEV_event_Erdeklodes = log(apEV_event_Erdeklodes + 1)
  , log_apEV_Bejelentkezve_Magan = log(apEV_Bejelentkezve_Magan + 1)
  , log_wEV_Bejelentkezve_Magan = log(wEV_Bejelentkezve_Magan + 1)
  , log_everEV_pageviews = log(everEV_pageviews + 1)
  , log_apEV_referer_icom = log(apEV_referer_icom + 1)
  , log_everEV_timeOnSite = log(everEV_timeOnSite + 1)
  , log_wEV_Budapestrol_Budapestre = log(wEV_Budapestrol_Budapestre + 1)
  , log_everEV_event_Kepgaleria = log(everEV_event_Kepgaleria + 1)
  , log_everEV_Bejelentkezve_Magan = log(everEV_Bejelentkezve_Magan + 1)
  , log_apEV_Budapestrol_Budapestre = log(apEV_Budapestrol_Budapestre + 1)
  , log_apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
  , log_everEV_event_Nincstalalatfrontend = log(everEV_event_Nincstalalatfrontend + 1)
  , log_everEV_event_Eszkoztar = log(everEV_event_Eszkoztar + 1)
  , log_apEV_channel_Organic_Search = log(apEV_channel_Organic_Search + 1)
  , log_wEV_action_Eszkoztar_Kattintas = log(wEV_action_Eszkoztar_Kattintas + 1)
  , log_apEV_action_Eszkoztar_Kattintas = log(apEV_action_Eszkoztar_Kattintas + 1)
  , log_wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
  , log_everEV_device_desktop = log(everEV_device_desktop + 1)
  , log_apEV_trafficSource_google_organic = log(apEV_trafficSource_google_organic + 1)
  , log_everEV_Honnan_Hungary = log(everEV_Honnan_Hungary + 1)
  , log_wEV_event_Erdeklodes = log(wEV_event_Erdeklodes + 1)
  , log_wEV_referer_icom = log(wEV_referer_icom + 1)
  , log_wEV_event_Eszkoztar = log(wEV_event_Eszkoztar + 1)
  , log_everEV_ReszletekOldalMegjelenes = log(everEV_ReszletekOldalMegjelenes + 1)
  , log_everEV_event_ReszletekOldalMegjelenes = log(everEV_event_ReszletekOldalMegjelenes + 1)
  , log_everEV_action_Kereses_ujEpitesuSzuro = log(everEV_action_Kereses_ujEpitesuSzuro + 1)
  , log_everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
  , log_everEV_action_Kereses_mobil = log(everEV_action_Kereses_mobil + 1)
  , log_everEV_channel_Organic_Search = log(everEV_channel_Organic_Search + 1)
  , log_everEV_event_M1X = log(everEV_event_M1X + 1)
  , log_apEV_event_Eszkoztar = log(apEV_event_Eszkoztar + 1)
  , log_wEV_action_Mobilelhelyezkedes_Linkethasznalt = log(wEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
  , log_apEV_event_M1X = log(apEV_event_M1X + 1)
  , log_everEV_trafficSource_google_organic = log(everEV_trafficSource_google_organic + 1)
  , log_everEV_action_Eszkoztar_Kattintas = log(everEV_action_Eszkoztar_Kattintas + 1)
  , log_everEV_event_Erdeklodes = log(everEV_event_Erdeklodes + 1)
  , log_everEV_action_Mentes_Sikeres = log(everEV_action_Mentes_Sikeres + 1)
  , log_apEV_event_FrontendBelepes = log(apEV_event_FrontendBelepes + 1)
  , log_everEV_event_FrontendBelepes = log(everEV_event_FrontendBelepes + 1)
  , log_everEV_event_Mentes = log(everEV_event_Mentes + 1)
  , log_apEV_referer_google = log(apEV_referer_google + 1)
  , log_everEV_socialNetwork_Facebook = log(everEV_socialNetwork_Facebook + 1)
  , log_apEV_trafficSource_google_cpc = log(apEV_trafficSource_google_cpc + 1)
  , log_everEV_SocialSourceReferral = log(everEV_SocialSourceReferral + 1)
  , log_everEV_channel_Social = log(everEV_channel_Social + 1)
  , log_apEV_channel_Paid_Search = log(apEV_channel_Paid_Search + 1)
  , log_everEV_action_FrontendBelepes_Megjelenes = log(everEV_action_FrontendBelepes_Megjelenes + 1)
  , log_everEV_trafficSource_facebook_referral = log(everEV_trafficSource_facebook_referral + 1)
  , log_wEV_trafficSource_facebook_referral = log(wEV_trafficSource_facebook_referral + 1)
  , log_wEV_socialNetwork_Facebook = log(wEV_socialNetwork_Facebook + 1)
  , log_wEV_SocialSourceReferral = log(wEV_SocialSourceReferral + 1)
  , log_wEV_channel_Social = log(wEV_channel_Social + 1)
  , log_apEV_action_Erdeklodes_OfficeUtcapontos = log(apEV_action_Erdeklodes_OfficeUtcapontos + 1)
  , log_wEV_channel_Organic_Search = log(wEV_channel_Organic_Search + 1)
  , log_apEV_action_FrontendBelepes_Megjelenes = log(apEV_action_FrontendBelepes_Megjelenes + 1)
  , log_everEV_Budapestrol_Budapestre = log(everEV_Budapestrol_Budapestre + 1)
  , log_wEV_trafficSource_google_organic = log(wEV_trafficSource_google_organic + 1)
  , log_wEV_referer_google = log(wEV_referer_google + 1)
  , log_apEV_trafficSource_facebook_referral = log(apEV_trafficSource_facebook_referral + 1)
  , log_apEV_socialNetwork_Facebook = log(apEV_socialNetwork_Facebook + 1)
  , log_everEV_label_FrontendBelepes_Megjelenes_fejlec = log(everEV_label_FrontendBelepes_Megjelenes_fejlec + 1)
  , log_apEV_SocialSourceReferral = log(apEV_SocialSourceReferral + 1)
  , log_apEV_channel_Social = log(apEV_channel_Social + 1)
  , log_wEV_event_FrontendBelepes = log(wEV_event_FrontendBelepes + 1)
  , log_everEV_channel_Paid_Search = log(everEV_channel_Paid_Search + 1)
  , log_everEV_trafficSource_google_cpc = log(everEV_trafficSource_google_cpc + 1)
  , log_wEV_channel_Email = log(wEV_channel_Email + 1)
  , log_apEV_action_M1X_talalkozottDuplikaltHirdetessel = log(apEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
  , log_everEV_event_OfficeLogo = log(everEV_event_OfficeLogo + 1)
  , log_wEV_trafficSource_google_cpc = log(wEV_trafficSource_google_cpc + 1)
  , log_everEV_channel_Email = log(everEV_channel_Email + 1)
  , log_everEV_event_AI = log(everEV_event_AI + 1)
  , log_everEV_event_lakoparkListaKattintas = log(everEV_event_lakoparkListaKattintas + 1)
  , log_wEV_event_M1X = log(wEV_event_M1X + 1)
  , log_everEV_event_lakoparkMegjelenes = log(everEV_event_lakoparkMegjelenes + 1)
  , log_everEV_referer_creativecdn = log(everEV_referer_creativecdn + 1)
  , log_wEV_channel_Paid_Search = log(wEV_channel_Paid_Search + 1)
  , log_everEV_action_M1X_talalkozottDuplikaltHirdetessel = log(everEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
  , log_everEV_action_Erdeklodes_OfficeUtcapontos = log(everEV_action_Erdeklodes_OfficeUtcapontos + 1)
  , log_everEV_action_FrontendBelepes_Sikeres = log(everEV_action_FrontendBelepes_Sikeres + 1)
  , log_apEV_channel_Email = log(apEV_channel_Email + 1)
  , log_everEV_label_FrontendBelepes_Sikeres_fejlec = log(everEV_label_FrontendBelepes_Sikeres_fejlec + 1)
  , log_wEV_trafficSource_Hirdetesfigyelo_email = log(wEV_trafficSource_Hirdetesfigyelo_email + 1)
  , log_wEV_action_M1X_talalkozottDuplikaltHirdetessel = log(wEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
  , log_apEV_action_Erdeklodes_OfficeNemUtcapontos = log(apEV_action_Erdeklodes_OfficeNemUtcapontos + 1)
  , log_apEV_referer_creativecdn = log(apEV_referer_creativecdn + 1)
  , log_apEV_trafficSource_Hirdetesfigyelo_email = log(apEV_trafficSource_Hirdetesfigyelo_email + 1)
  , log_everEV_action_Erdeklodes_MaganFizetos = log(everEV_action_Erdeklodes_MaganFizetos + 1)
  , log_everEV_referer_icom = log(everEV_referer_icom + 1)
  , log_everEV_action_Erdeklodes_OfficeNemUtcapontos = log(everEV_action_Erdeklodes_OfficeNemUtcapontos + 1)
  , log_everEV_trafficSource_Hirdetesfigyelo_email = log(everEV_trafficSource_Hirdetesfigyelo_email + 1)
  , log_everEV_referer_google = log(everEV_referer_google + 1)
  , log_wEV_action_Mobilreszletek_Kepgaleria = log(wEV_action_Mobilreszletek_Kepgaleria + 1)
  , log_everEV_Visits = log(everEV_Visits + 1)
  , log_everEV_Felhasznalo_Magan = log(everEV_Felhasznalo_Magan + 1)
  , log_everEV_event_Mobilreszletek = log(everEV_event_Mobilreszletek + 1)
  , log_wEV_event_Mobilreszletek = log(wEV_event_Mobilreszletek + 1)
  , log_everEV_action_Mobilreszletek_Kepgaleria = log(everEV_action_Mobilreszletek_Kepgaleria + 1)
  , log_everRE_AVG_Kepgaleria = log(everRE_AVG_Kepgaleria + 1)
  , log_everEV_event_Piac = log(everEV_event_Piac + 1)
  , log_wRE_AVG_Kepgaleria = log(wRE_AVG_Kepgaleria + 1)
  , log_wEV_trafficSource_rtbhouse_retargeting = log(wEV_trafficSource_rtbhouse_retargeting + 1)
  , log_everEV_channel_Direct = log(everEV_channel_Direct + 1)
  , log_wRE_SUM_Kepgaleria = log(wRE_SUM_Kepgaleria + 1)
  , log_everEV_action_Piac_undefined = log(everEV_action_Piac_undefined + 1)
  , log_wRE_MAX_Kepgaleria = log(wRE_MAX_Kepgaleria + 1)
  , log_apEV_trafficSource_rtbhouse_retargeting = log(apEV_trafficSource_rtbhouse_retargeting + 1)
  , log_apEV_action_Mobilreszletek_Kepgaleria = log(apEV_action_Mobilreszletek_Kepgaleria + 1)
  , log_apEV_event_Mobilreszletek = log(apEV_event_Mobilreszletek + 1)
  , log_apRE_AVG_Kepgaleria = log(apRE_AVG_Kepgaleria + 1)
  , log_everEV_event_MaganhirdetoInfok = log(everEV_event_MaganhirdetoInfok + 1)
  , log_everRE_MAX_Kepgaleria = log(everRE_MAX_Kepgaleria + 1)
  , log_everEV_trafficSource_rtbhouse_retargeting = log(everEV_trafficSource_rtbhouse_retargeting + 1)
  , log_apRE_SUM_Kepgaleria = log(apRE_SUM_Kepgaleria + 1)
  , log_everRE_SUM_Kepgaleria = log(everRE_SUM_Kepgaleria + 1)
  , log_apRE_MAX_Kepgaleria = log(apRE_MAX_Kepgaleria + 1)
  , log_everEV_action_Erdeklodes_MaganIngyenes = log(everEV_action_Erdeklodes_MaganIngyenes + 1)
  , log_everRE_ikerhaz = log(everRE_ikerhaz + 1)
  , log_everEV_Videkrol_Budapestre = log(everEV_Videkrol_Budapestre + 1)

),]

```

Test
```{r}
JZ_test <- JZ_test[, `:=`(
  log_apEV_event_autosuggest = log(apEV_event_autosuggest + 1)
  , log_everEV_event_Mobilelhelyezkedes = log(everEV_event_Mobilelhelyezkedes + 1)
  , log_everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
  , log_apEV_event_Mobilelhelyezkedes = log(apEV_event_Mobilelhelyezkedes + 1)
  , log_everEV_event_autosuggest = log(everEV_event_autosuggest + 1)
  , log_wEV_event_autosuggest = log(wEV_event_autosuggest + 1)
  , log_wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
  , log_apEV_timeOnSite = log(apEV_timeOnSite + 1)
  , log_apEV_pageviews = log(apEV_pageviews + 1)
  , log_apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
  , log_wEV_event_Mobilelhelyezkedes = log(wEV_event_Mobilelhelyezkedes + 1)
  , log_apEV_HirdetesTipus_Lakossag = log(apEV_HirdetesTipus_Lakossag + 1)
  , log_apEV_HirdetesEladasmodja_Elado = log(apEV_HirdetesEladasmodja_Elado + 1)
  , log_apEV_device_desktop = log(apEV_device_desktop + 1)
  , log_apEV_hits = log(apEV_hits + 1)
  , log_wEV_timeOnSite = log(wEV_timeOnSite + 1)
  , log_wEV_HirdetesTipus_Lakossag = log(wEV_HirdetesTipus_Lakossag + 1)
  , log_apEV_Honnan_Hungary = log(apEV_Honnan_Hungary + 1)
  , log_wEV_HirdetesEladasmodja_Elado = log(wEV_HirdetesEladasmodja_Elado + 1)
  , log_apEV_event_ReszletekOldalMegjelenes = log(apEV_event_ReszletekOldalMegjelenes + 1)
  , log_wEV_device_desktop = log(wEV_device_desktop + 1)
  , log_wEV_pageviews = log(wEV_pageviews + 1)
  , log_wEV_Honnan_Hungary = log(wEV_Honnan_Hungary + 1)
  , log_everEV_action_Kereso_ReszletesSzukitoLenyit = log(everEV_action_Kereso_ReszletesSzukitoLenyit + 1)
  , log_apEV_event_Kereso = log(apEV_event_Kereso + 1)
  , log_everEV_event_Kereso = log(everEV_event_Kereso + 1)
  , log_wEV_hits = log(wEV_hits + 1)
  , log_apEV_action_Kereso_ReszletesSzukitoLenyit = log(apEV_action_Kereso_ReszletesSzukitoLenyit + 1)
  , log_wEV_event_Kepgaleria = log(wEV_event_Kepgaleria + 1)
  , log_wEV_event_ReszletekOldalMegjelenes = log(wEV_event_ReszletekOldalMegjelenes + 1)
  , log_apEV_event_Kepgaleria = log(apEV_event_Kepgaleria + 1)
  , log_apEV_Visits = log(apEV_Visits + 1)
  , log_everEV_action_Mobilelhelyezkedes_Linkethasznalt = log(everEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
  , log_everEV_hits = log(everEV_hits + 1)
  , log_everEV_HirdetesEladasmodja_Elado = log(everEV_HirdetesEladasmodja_Elado + 1)
  , log_apEV_action_Mobilelhelyezkedes_Linkethasznalt = log(apEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
  , log_everEV_event_Kereses = log(everEV_event_Kereses + 1)
  , log_wEV_Visits = log(wEV_Visits + 1)
  , log_everEV_HirdetesTipus_Lakossag = log(everEV_HirdetesTipus_Lakossag + 1)
  , log_apEV_event_Nincstalalatfrontend = log(apEV_event_Nincstalalatfrontend + 1)
  , log_wEV_action_Kereso_ReszletesSzukitoLenyit = log(wEV_action_Kereso_ReszletesSzukitoLenyit + 1)
  , log_wEV_event_Kereso = log(wEV_event_Kereso + 1)
  , log_apEV_event_Erdeklodes = log(apEV_event_Erdeklodes + 1)
  , log_apEV_Bejelentkezve_Magan = log(apEV_Bejelentkezve_Magan + 1)
  , log_wEV_Bejelentkezve_Magan = log(wEV_Bejelentkezve_Magan + 1)
  , log_everEV_pageviews = log(everEV_pageviews + 1)
  , log_apEV_referer_icom = log(apEV_referer_icom + 1)
  , log_everEV_timeOnSite = log(everEV_timeOnSite + 1)
  , log_wEV_Budapestrol_Budapestre = log(wEV_Budapestrol_Budapestre + 1)
  , log_everEV_event_Kepgaleria = log(everEV_event_Kepgaleria + 1)
  , log_everEV_Bejelentkezve_Magan = log(everEV_Bejelentkezve_Magan + 1)
  , log_apEV_Budapestrol_Budapestre = log(apEV_Budapestrol_Budapestre + 1)
  , log_apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
  , log_everEV_event_Nincstalalatfrontend = log(everEV_event_Nincstalalatfrontend + 1)
  , log_everEV_event_Eszkoztar = log(everEV_event_Eszkoztar + 1)
  , log_apEV_channel_Organic_Search = log(apEV_channel_Organic_Search + 1)
  , log_wEV_action_Eszkoztar_Kattintas = log(wEV_action_Eszkoztar_Kattintas + 1)
  , log_apEV_action_Eszkoztar_Kattintas = log(apEV_action_Eszkoztar_Kattintas + 1)
  , log_wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
  , log_everEV_device_desktop = log(everEV_device_desktop + 1)
  , log_apEV_trafficSource_google_organic = log(apEV_trafficSource_google_organic + 1)
  , log_everEV_Honnan_Hungary = log(everEV_Honnan_Hungary + 1)
  , log_wEV_event_Erdeklodes = log(wEV_event_Erdeklodes + 1)
  , log_wEV_referer_icom = log(wEV_referer_icom + 1)
  , log_wEV_event_Eszkoztar = log(wEV_event_Eszkoztar + 1)
  , log_everEV_ReszletekOldalMegjelenes = log(everEV_ReszletekOldalMegjelenes + 1)
  , log_everEV_event_ReszletekOldalMegjelenes = log(everEV_event_ReszletekOldalMegjelenes + 1)
  , log_everEV_action_Kereses_ujEpitesuSzuro = log(everEV_action_Kereses_ujEpitesuSzuro + 1)
  , log_everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
  , log_everEV_action_Kereses_mobil = log(everEV_action_Kereses_mobil + 1)
  , log_everEV_channel_Organic_Search = log(everEV_channel_Organic_Search + 1)
  , log_everEV_event_M1X = log(everEV_event_M1X + 1)
  , log_apEV_event_Eszkoztar = log(apEV_event_Eszkoztar + 1)
  , log_wEV_action_Mobilelhelyezkedes_Linkethasznalt = log(wEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
  , log_apEV_event_M1X = log(apEV_event_M1X + 1)
  , log_everEV_trafficSource_google_organic = log(everEV_trafficSource_google_organic + 1)
  , log_everEV_action_Eszkoztar_Kattintas = log(everEV_action_Eszkoztar_Kattintas + 1)
  , log_everEV_event_Erdeklodes = log(everEV_event_Erdeklodes + 1)
  , log_everEV_action_Mentes_Sikeres = log(everEV_action_Mentes_Sikeres + 1)
  , log_apEV_event_FrontendBelepes = log(apEV_event_FrontendBelepes + 1)
  , log_everEV_event_FrontendBelepes = log(everEV_event_FrontendBelepes + 1)
  , log_everEV_event_Mentes = log(everEV_event_Mentes + 1)
  , log_apEV_referer_google = log(apEV_referer_google + 1)
  , log_everEV_socialNetwork_Facebook = log(everEV_socialNetwork_Facebook + 1)
  , log_apEV_trafficSource_google_cpc = log(apEV_trafficSource_google_cpc + 1)
  , log_everEV_SocialSourceReferral = log(everEV_SocialSourceReferral + 1)
  , log_everEV_channel_Social = log(everEV_channel_Social + 1)
  , log_apEV_channel_Paid_Search = log(apEV_channel_Paid_Search + 1)
  , log_everEV_action_FrontendBelepes_Megjelenes = log(everEV_action_FrontendBelepes_Megjelenes + 1)
  , log_everEV_trafficSource_facebook_referral = log(everEV_trafficSource_facebook_referral + 1)
  , log_wEV_trafficSource_facebook_referral = log(wEV_trafficSource_facebook_referral + 1)
  , log_wEV_socialNetwork_Facebook = log(wEV_socialNetwork_Facebook + 1)
  , log_wEV_SocialSourceReferral = log(wEV_SocialSourceReferral + 1)
  , log_wEV_channel_Social = log(wEV_channel_Social + 1)
  , log_apEV_action_Erdeklodes_OfficeUtcapontos = log(apEV_action_Erdeklodes_OfficeUtcapontos + 1)
  , log_wEV_channel_Organic_Search = log(wEV_channel_Organic_Search + 1)
  , log_apEV_action_FrontendBelepes_Megjelenes = log(apEV_action_FrontendBelepes_Megjelenes + 1)
  , log_everEV_Budapestrol_Budapestre = log(everEV_Budapestrol_Budapestre + 1)
  , log_wEV_trafficSource_google_organic = log(wEV_trafficSource_google_organic + 1)
  , log_wEV_referer_google = log(wEV_referer_google + 1)
  , log_apEV_trafficSource_facebook_referral = log(apEV_trafficSource_facebook_referral + 1)
  , log_apEV_socialNetwork_Facebook = log(apEV_socialNetwork_Facebook + 1)
  , log_everEV_label_FrontendBelepes_Megjelenes_fejlec = log(everEV_label_FrontendBelepes_Megjelenes_fejlec + 1)
  , log_apEV_SocialSourceReferral = log(apEV_SocialSourceReferral + 1)
  , log_apEV_channel_Social = log(apEV_channel_Social + 1)
  , log_wEV_event_FrontendBelepes = log(wEV_event_FrontendBelepes + 1)
  , log_everEV_channel_Paid_Search = log(everEV_channel_Paid_Search + 1)
  , log_everEV_trafficSource_google_cpc = log(everEV_trafficSource_google_cpc + 1)
  , log_wEV_channel_Email = log(wEV_channel_Email + 1)
  , log_apEV_action_M1X_talalkozottDuplikaltHirdetessel = log(apEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
  , log_everEV_event_OfficeLogo = log(everEV_event_OfficeLogo + 1)
  , log_wEV_trafficSource_google_cpc = log(wEV_trafficSource_google_cpc + 1)
  , log_everEV_channel_Email = log(everEV_channel_Email + 1)
  , log_everEV_event_AI = log(everEV_event_AI + 1)
  , log_everEV_event_lakoparkListaKattintas = log(everEV_event_lakoparkListaKattintas + 1)
  , log_wEV_event_M1X = log(wEV_event_M1X + 1)
  , log_everEV_event_lakoparkMegjelenes = log(everEV_event_lakoparkMegjelenes + 1)
  , log_everEV_referer_creativecdn = log(everEV_referer_creativecdn + 1)
  , log_wEV_channel_Paid_Search = log(wEV_channel_Paid_Search + 1)
  , log_everEV_action_M1X_talalkozottDuplikaltHirdetessel = log(everEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
  , log_everEV_action_Erdeklodes_OfficeUtcapontos = log(everEV_action_Erdeklodes_OfficeUtcapontos + 1)
  , log_everEV_action_FrontendBelepes_Sikeres = log(everEV_action_FrontendBelepes_Sikeres + 1)
  , log_apEV_channel_Email = log(apEV_channel_Email + 1)
  , log_everEV_label_FrontendBelepes_Sikeres_fejlec = log(everEV_label_FrontendBelepes_Sikeres_fejlec + 1)
  , log_wEV_trafficSource_Hirdetesfigyelo_email = log(wEV_trafficSource_Hirdetesfigyelo_email + 1)
  , log_wEV_action_M1X_talalkozottDuplikaltHirdetessel = log(wEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
  , log_apEV_action_Erdeklodes_OfficeNemUtcapontos = log(apEV_action_Erdeklodes_OfficeNemUtcapontos + 1)
  , log_apEV_referer_creativecdn = log(apEV_referer_creativecdn + 1)
  , log_apEV_trafficSource_Hirdetesfigyelo_email = log(apEV_trafficSource_Hirdetesfigyelo_email + 1)
  , log_everEV_action_Erdeklodes_MaganFizetos = log(everEV_action_Erdeklodes_MaganFizetos + 1)
  , log_everEV_referer_icom = log(everEV_referer_icom + 1)
  , log_everEV_action_Erdeklodes_OfficeNemUtcapontos = log(everEV_action_Erdeklodes_OfficeNemUtcapontos + 1)
  , log_everEV_trafficSource_Hirdetesfigyelo_email = log(everEV_trafficSource_Hirdetesfigyelo_email + 1)
  , log_everEV_referer_google = log(everEV_referer_google + 1)
  , log_wEV_action_Mobilreszletek_Kepgaleria = log(wEV_action_Mobilreszletek_Kepgaleria + 1)
  , log_everEV_Visits = log(everEV_Visits + 1)
  , log_everEV_Felhasznalo_Magan = log(everEV_Felhasznalo_Magan + 1)
  , log_everEV_event_Mobilreszletek = log(everEV_event_Mobilreszletek + 1)
  , log_wEV_event_Mobilreszletek = log(wEV_event_Mobilreszletek + 1)
  , log_everEV_action_Mobilreszletek_Kepgaleria = log(everEV_action_Mobilreszletek_Kepgaleria + 1)
  , log_everRE_AVG_Kepgaleria = log(everRE_AVG_Kepgaleria + 1)
  , log_everEV_event_Piac = log(everEV_event_Piac + 1)
  , log_wRE_AVG_Kepgaleria = log(wRE_AVG_Kepgaleria + 1)
  , log_wEV_trafficSource_rtbhouse_retargeting = log(wEV_trafficSource_rtbhouse_retargeting + 1)
  , log_everEV_channel_Direct = log(everEV_channel_Direct + 1)
  , log_wRE_SUM_Kepgaleria = log(wRE_SUM_Kepgaleria + 1)
  , log_everEV_action_Piac_undefined = log(everEV_action_Piac_undefined + 1)
  , log_wRE_MAX_Kepgaleria = log(wRE_MAX_Kepgaleria + 1)
  , log_apEV_trafficSource_rtbhouse_retargeting = log(apEV_trafficSource_rtbhouse_retargeting + 1)
  , log_apEV_action_Mobilreszletek_Kepgaleria = log(apEV_action_Mobilreszletek_Kepgaleria + 1)
  , log_apEV_event_Mobilreszletek = log(apEV_event_Mobilreszletek + 1)
  , log_apRE_AVG_Kepgaleria = log(apRE_AVG_Kepgaleria + 1)
  , log_everEV_event_MaganhirdetoInfok = log(everEV_event_MaganhirdetoInfok + 1)
  , log_everRE_MAX_Kepgaleria = log(everRE_MAX_Kepgaleria + 1)
  , log_everEV_trafficSource_rtbhouse_retargeting = log(everEV_trafficSource_rtbhouse_retargeting + 1)
  , log_apRE_SUM_Kepgaleria = log(apRE_SUM_Kepgaleria + 1)
  , log_everRE_SUM_Kepgaleria = log(everRE_SUM_Kepgaleria + 1)
  , log_apRE_MAX_Kepgaleria = log(apRE_MAX_Kepgaleria + 1)
  , log_everEV_action_Erdeklodes_MaganIngyenes = log(everEV_action_Erdeklodes_MaganIngyenes + 1)
  , log_everRE_ikerhaz = log(everRE_ikerhaz + 1)
  , log_everEV_Videkrol_Budapestre = log(everEV_Videkrol_Budapestre + 1)
  
),]
```

External
```{r}

JZ_external <- JZ_external[, `:=`(
  log_apEV_event_autosuggest = log(apEV_event_autosuggest + 1)
  , log_everEV_event_Mobilelhelyezkedes = log(everEV_event_Mobilelhelyezkedes + 1)
  , log_everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
  , log_apEV_event_Mobilelhelyezkedes = log(apEV_event_Mobilelhelyezkedes + 1)
  , log_everEV_event_autosuggest = log(everEV_event_autosuggest + 1)
  , log_wEV_event_autosuggest = log(wEV_event_autosuggest + 1)
  , log_wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
  , log_apEV_timeOnSite = log(apEV_timeOnSite + 1)
  , log_apEV_pageviews = log(apEV_pageviews + 1)
  , log_apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
  , log_wEV_event_Mobilelhelyezkedes = log(wEV_event_Mobilelhelyezkedes + 1)
  , log_apEV_HirdetesTipus_Lakossag = log(apEV_HirdetesTipus_Lakossag + 1)
  , log_apEV_HirdetesEladasmodja_Elado = log(apEV_HirdetesEladasmodja_Elado + 1)
  , log_apEV_device_desktop = log(apEV_device_desktop + 1)
  , log_apEV_hits = log(apEV_hits + 1)
  , log_wEV_timeOnSite = log(wEV_timeOnSite + 1)
  , log_wEV_HirdetesTipus_Lakossag = log(wEV_HirdetesTipus_Lakossag + 1)
  , log_apEV_Honnan_Hungary = log(apEV_Honnan_Hungary + 1)
  , log_wEV_HirdetesEladasmodja_Elado = log(wEV_HirdetesEladasmodja_Elado + 1)
  , log_apEV_event_ReszletekOldalMegjelenes = log(apEV_event_ReszletekOldalMegjelenes + 1)
  , log_wEV_device_desktop = log(wEV_device_desktop + 1)
  , log_wEV_pageviews = log(wEV_pageviews + 1)
  , log_wEV_Honnan_Hungary = log(wEV_Honnan_Hungary + 1)
  , log_everEV_action_Kereso_ReszletesSzukitoLenyit = log(everEV_action_Kereso_ReszletesSzukitoLenyit + 1)
  , log_apEV_event_Kereso = log(apEV_event_Kereso + 1)
  , log_everEV_event_Kereso = log(everEV_event_Kereso + 1)
  , log_wEV_hits = log(wEV_hits + 1)
  , log_apEV_action_Kereso_ReszletesSzukitoLenyit = log(apEV_action_Kereso_ReszletesSzukitoLenyit + 1)
  , log_wEV_event_Kepgaleria = log(wEV_event_Kepgaleria + 1)
  , log_wEV_event_ReszletekOldalMegjelenes = log(wEV_event_ReszletekOldalMegjelenes + 1)
  , log_apEV_event_Kepgaleria = log(apEV_event_Kepgaleria + 1)
  , log_apEV_Visits = log(apEV_Visits + 1)
  , log_everEV_action_Mobilelhelyezkedes_Linkethasznalt = log(everEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
  , log_everEV_hits = log(everEV_hits + 1)
  , log_everEV_HirdetesEladasmodja_Elado = log(everEV_HirdetesEladasmodja_Elado + 1)
  , log_apEV_action_Mobilelhelyezkedes_Linkethasznalt = log(apEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
  , log_everEV_event_Kereses = log(everEV_event_Kereses + 1)
  , log_wEV_Visits = log(wEV_Visits + 1)
  , log_everEV_HirdetesTipus_Lakossag = log(everEV_HirdetesTipus_Lakossag + 1)
  , log_apEV_event_Nincstalalatfrontend = log(apEV_event_Nincstalalatfrontend + 1)
  , log_wEV_action_Kereso_ReszletesSzukitoLenyit = log(wEV_action_Kereso_ReszletesSzukitoLenyit + 1)
  , log_wEV_event_Kereso = log(wEV_event_Kereso + 1)
  , log_apEV_event_Erdeklodes = log(apEV_event_Erdeklodes + 1)
  , log_apEV_Bejelentkezve_Magan = log(apEV_Bejelentkezve_Magan + 1)
  , log_wEV_Bejelentkezve_Magan = log(wEV_Bejelentkezve_Magan + 1)
  , log_everEV_pageviews = log(everEV_pageviews + 1)
  , log_apEV_referer_icom = log(apEV_referer_icom + 1)
  , log_everEV_timeOnSite = log(everEV_timeOnSite + 1)
  , log_wEV_Budapestrol_Budapestre = log(wEV_Budapestrol_Budapestre + 1)
  , log_everEV_event_Kepgaleria = log(everEV_event_Kepgaleria + 1)
  , log_everEV_Bejelentkezve_Magan = log(everEV_Bejelentkezve_Magan + 1)
  , log_apEV_Budapestrol_Budapestre = log(apEV_Budapestrol_Budapestre + 1)
  , log_apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
  , log_everEV_event_Nincstalalatfrontend = log(everEV_event_Nincstalalatfrontend + 1)
  , log_everEV_event_Eszkoztar = log(everEV_event_Eszkoztar + 1)
  , log_apEV_channel_Organic_Search = log(apEV_channel_Organic_Search + 1)
  , log_wEV_action_Eszkoztar_Kattintas = log(wEV_action_Eszkoztar_Kattintas + 1)
  , log_apEV_action_Eszkoztar_Kattintas = log(apEV_action_Eszkoztar_Kattintas + 1)
  , log_wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
  , log_everEV_device_desktop = log(everEV_device_desktop + 1)
  , log_apEV_trafficSource_google_organic = log(apEV_trafficSource_google_organic + 1)
  , log_everEV_Honnan_Hungary = log(everEV_Honnan_Hungary + 1)
  , log_wEV_event_Erdeklodes = log(wEV_event_Erdeklodes + 1)
  , log_wEV_referer_icom = log(wEV_referer_icom + 1)
  , log_wEV_event_Eszkoztar = log(wEV_event_Eszkoztar + 1)
  , log_everEV_ReszletekOldalMegjelenes = log(everEV_ReszletekOldalMegjelenes + 1)
  , log_everEV_event_ReszletekOldalMegjelenes = log(everEV_event_ReszletekOldalMegjelenes + 1)
  , log_everEV_action_Kereses_ujEpitesuSzuro = log(everEV_action_Kereses_ujEpitesuSzuro + 1)
  , log_everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
  , log_everEV_action_Kereses_mobil = log(everEV_action_Kereses_mobil + 1)
  , log_everEV_channel_Organic_Search = log(everEV_channel_Organic_Search + 1)
  , log_everEV_event_M1X = log(everEV_event_M1X + 1)
  , log_apEV_event_Eszkoztar = log(apEV_event_Eszkoztar + 1)
  , log_wEV_action_Mobilelhelyezkedes_Linkethasznalt = log(wEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
  , log_apEV_event_M1X = log(apEV_event_M1X + 1)
  , log_everEV_trafficSource_google_organic = log(everEV_trafficSource_google_organic + 1)
  , log_everEV_action_Eszkoztar_Kattintas = log(everEV_action_Eszkoztar_Kattintas + 1)
  , log_everEV_event_Erdeklodes = log(everEV_event_Erdeklodes + 1)
  , log_everEV_action_Mentes_Sikeres = log(everEV_action_Mentes_Sikeres + 1)
  , log_apEV_event_FrontendBelepes = log(apEV_event_FrontendBelepes + 1)
  , log_everEV_event_FrontendBelepes = log(everEV_event_FrontendBelepes + 1)
  , log_everEV_event_Mentes = log(everEV_event_Mentes + 1)
  , log_apEV_referer_google = log(apEV_referer_google + 1)
  , log_everEV_socialNetwork_Facebook = log(everEV_socialNetwork_Facebook + 1)
  , log_apEV_trafficSource_google_cpc = log(apEV_trafficSource_google_cpc + 1)
  , log_everEV_SocialSourceReferral = log(everEV_SocialSourceReferral + 1)
  , log_everEV_channel_Social = log(everEV_channel_Social + 1)
  , log_apEV_channel_Paid_Search = log(apEV_channel_Paid_Search + 1)
  , log_everEV_action_FrontendBelepes_Megjelenes = log(everEV_action_FrontendBelepes_Megjelenes + 1)
  , log_everEV_trafficSource_facebook_referral = log(everEV_trafficSource_facebook_referral + 1)
  , log_wEV_trafficSource_facebook_referral = log(wEV_trafficSource_facebook_referral + 1)
  , log_wEV_socialNetwork_Facebook = log(wEV_socialNetwork_Facebook + 1)
  , log_wEV_SocialSourceReferral = log(wEV_SocialSourceReferral + 1)
  , log_wEV_channel_Social = log(wEV_channel_Social + 1)
  , log_apEV_action_Erdeklodes_OfficeUtcapontos = log(apEV_action_Erdeklodes_OfficeUtcapontos + 1)
  , log_wEV_channel_Organic_Search = log(wEV_channel_Organic_Search + 1)
  , log_apEV_action_FrontendBelepes_Megjelenes = log(apEV_action_FrontendBelepes_Megjelenes + 1)
  , log_everEV_Budapestrol_Budapestre = log(everEV_Budapestrol_Budapestre + 1)
  , log_wEV_trafficSource_google_organic = log(wEV_trafficSource_google_organic + 1)
  , log_wEV_referer_google = log(wEV_referer_google + 1)
  , log_apEV_trafficSource_facebook_referral = log(apEV_trafficSource_facebook_referral + 1)
  , log_apEV_socialNetwork_Facebook = log(apEV_socialNetwork_Facebook + 1)
  , log_everEV_label_FrontendBelepes_Megjelenes_fejlec = log(everEV_label_FrontendBelepes_Megjelenes_fejlec + 1)
  , log_apEV_SocialSourceReferral = log(apEV_SocialSourceReferral + 1)
  , log_apEV_channel_Social = log(apEV_channel_Social + 1)
  , log_wEV_event_FrontendBelepes = log(wEV_event_FrontendBelepes + 1)
  , log_everEV_channel_Paid_Search = log(everEV_channel_Paid_Search + 1)
  , log_everEV_trafficSource_google_cpc = log(everEV_trafficSource_google_cpc + 1)
  , log_wEV_channel_Email = log(wEV_channel_Email + 1)
  , log_apEV_action_M1X_talalkozottDuplikaltHirdetessel = log(apEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
  , log_everEV_event_OfficeLogo = log(everEV_event_OfficeLogo + 1)
  , log_wEV_trafficSource_google_cpc = log(wEV_trafficSource_google_cpc + 1)
  , log_everEV_channel_Email = log(everEV_channel_Email + 1)
  , log_everEV_event_AI = log(everEV_event_AI + 1)
  , log_everEV_event_lakoparkListaKattintas = log(everEV_event_lakoparkListaKattintas + 1)
  , log_wEV_event_M1X = log(wEV_event_M1X + 1)
  , log_everEV_event_lakoparkMegjelenes = log(everEV_event_lakoparkMegjelenes + 1)
  , log_everEV_referer_creativecdn = log(everEV_referer_creativecdn + 1)
  , log_wEV_channel_Paid_Search = log(wEV_channel_Paid_Search + 1)
  , log_everEV_action_M1X_talalkozottDuplikaltHirdetessel = log(everEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
  , log_everEV_action_Erdeklodes_OfficeUtcapontos = log(everEV_action_Erdeklodes_OfficeUtcapontos + 1)
  , log_everEV_action_FrontendBelepes_Sikeres = log(everEV_action_FrontendBelepes_Sikeres + 1)
  , log_apEV_channel_Email = log(apEV_channel_Email + 1)
  , log_everEV_label_FrontendBelepes_Sikeres_fejlec = log(everEV_label_FrontendBelepes_Sikeres_fejlec + 1)
  , log_wEV_trafficSource_Hirdetesfigyelo_email = log(wEV_trafficSource_Hirdetesfigyelo_email + 1)
  , log_wEV_action_M1X_talalkozottDuplikaltHirdetessel = log(wEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
  , log_apEV_action_Erdeklodes_OfficeNemUtcapontos = log(apEV_action_Erdeklodes_OfficeNemUtcapontos + 1)
  , log_apEV_referer_creativecdn = log(apEV_referer_creativecdn + 1)
  , log_apEV_trafficSource_Hirdetesfigyelo_email = log(apEV_trafficSource_Hirdetesfigyelo_email + 1)
  , log_everEV_action_Erdeklodes_MaganFizetos = log(everEV_action_Erdeklodes_MaganFizetos + 1)
  , log_everEV_referer_icom = log(everEV_referer_icom + 1)
  , log_everEV_action_Erdeklodes_OfficeNemUtcapontos = log(everEV_action_Erdeklodes_OfficeNemUtcapontos + 1)
  , log_everEV_trafficSource_Hirdetesfigyelo_email = log(everEV_trafficSource_Hirdetesfigyelo_email + 1)
  , log_everEV_referer_google = log(everEV_referer_google + 1)
  , log_wEV_action_Mobilreszletek_Kepgaleria = log(wEV_action_Mobilreszletek_Kepgaleria + 1)
  , log_everEV_Visits = log(everEV_Visits + 1)
  , log_everEV_Felhasznalo_Magan = log(everEV_Felhasznalo_Magan + 1)
  , log_everEV_event_Mobilreszletek = log(everEV_event_Mobilreszletek + 1)
  , log_wEV_event_Mobilreszletek = log(wEV_event_Mobilreszletek + 1)
  , log_everEV_action_Mobilreszletek_Kepgaleria = log(everEV_action_Mobilreszletek_Kepgaleria + 1)
  , log_everRE_AVG_Kepgaleria = log(everRE_AVG_Kepgaleria + 1)
  , log_everEV_event_Piac = log(everEV_event_Piac + 1)
  , log_wRE_AVG_Kepgaleria = log(wRE_AVG_Kepgaleria + 1)
  , log_wEV_trafficSource_rtbhouse_retargeting = log(wEV_trafficSource_rtbhouse_retargeting + 1)
  , log_everEV_channel_Direct = log(everEV_channel_Direct + 1)
  , log_wRE_SUM_Kepgaleria = log(wRE_SUM_Kepgaleria + 1)
  , log_everEV_action_Piac_undefined = log(everEV_action_Piac_undefined + 1)
  , log_wRE_MAX_Kepgaleria = log(wRE_MAX_Kepgaleria + 1)
  , log_apEV_trafficSource_rtbhouse_retargeting = log(apEV_trafficSource_rtbhouse_retargeting + 1)
  , log_apEV_action_Mobilreszletek_Kepgaleria = log(apEV_action_Mobilreszletek_Kepgaleria + 1)
  , log_apEV_event_Mobilreszletek = log(apEV_event_Mobilreszletek + 1)
  , log_apRE_AVG_Kepgaleria = log(apRE_AVG_Kepgaleria + 1)
  , log_everEV_event_MaganhirdetoInfok = log(everEV_event_MaganhirdetoInfok + 1)
  , log_everRE_MAX_Kepgaleria = log(everRE_MAX_Kepgaleria + 1)
  , log_everEV_trafficSource_rtbhouse_retargeting = log(everEV_trafficSource_rtbhouse_retargeting + 1)
  , log_apRE_SUM_Kepgaleria = log(apRE_SUM_Kepgaleria + 1)
  , log_everRE_SUM_Kepgaleria = log(everRE_SUM_Kepgaleria + 1)
  , log_apRE_MAX_Kepgaleria = log(apRE_MAX_Kepgaleria + 1)
  , log_everEV_action_Erdeklodes_MaganIngyenes = log(everEV_action_Erdeklodes_MaganIngyenes + 1)
  , log_everRE_ikerhaz = log(everRE_ikerhaz + 1)
  , log_everEV_Videkrol_Budapestre = log(everEV_Videkrol_Budapestre + 1)
  
),]
```


#2 weeks performance period (f_Target)


Train
```{r}
JZ_train <- JZ_train[, `:=`(
log2w_apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
, log2w_everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
, log2w_everEV_event_Mobilelhelyezkedes = log(everEV_event_Mobilelhelyezkedes + 1)
, log2w_apEV_pageviews = log(apEV_pageviews + 1)
, log2w_apEV_HirdetesTipus_Lakossag = log(apEV_HirdetesTipus_Lakossag + 1)
, log2w_apEV_HirdetesEladasmodja_Elado = log(apEV_HirdetesEladasmodja_Elado + 1)
, log2w_apEV_event_Mobilelhelyezkedes = log(apEV_event_Mobilelhelyezkedes + 1)
, log2w_apEV_timeOnSite = log(apEV_timeOnSite + 1)
, log2w_everEV_event_autosuggest = log(everEV_event_autosuggest + 1)
, log2w_apEV_event_autosuggest = log(apEV_event_autosuggest + 1)
, log2w_apEV_hits = log(apEV_hits + 1)
, log2w_wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
, log2w_apEV_device_desktop = log(apEV_device_desktop + 1)
, log2w_wEV_event_Mobilelhelyezkedes = log(wEV_event_Mobilelhelyezkedes + 1)
, log2w_wEV_pageviews = log(wEV_pageviews + 1)
, log2w_wEV_event_autosuggest = log(wEV_event_autosuggest + 1)
, log2w_wEV_timeOnSite = log(wEV_timeOnSite + 1)
, log2w_wEV_event_ReszletekOldalMegjelenes = log(wEV_event_ReszletekOldalMegjelenes + 1)
, log2w_apEV_event_ReszletekOldalMegjelenes = log(apEV_event_ReszletekOldalMegjelenes + 1)
, log2w_everEV_action_Kereso_ReszletesSzukitoLenyit = log(everEV_action_Kereso_ReszletesSzukitoLenyit + 1)
, log2w_wEV_HirdetesEladasmodja_Elado = log(wEV_HirdetesEladasmodja_Elado + 1)
, log2w_wEV_HirdetesTipus_Lakossag = log(wEV_HirdetesTipus_Lakossag + 1)
, log2w_everEV_event_Kereso = log(everEV_event_Kereso + 1)
, log2w_wEV_hits = log(wEV_hits + 1)
, log2w_apEV_Honnan_Hungary = log(apEV_Honnan_Hungary + 1)
, log2w_wEV_device_desktop = log(wEV_device_desktop + 1)
, log2w_apEV_event_Kereso = log(apEV_event_Kereso + 1)
, log2w_wEV_Honnan_Hungary = log(wEV_Honnan_Hungary + 1)
, log2w_apEV_action_Kereso_ReszletesSzukitoLenyit = log(apEV_action_Kereso_ReszletesSzukitoLenyit + 1)
, log2w_everEV_pageviews = log(everEV_pageviews + 1)
, log2w_apEV_event_Kepgaleria = log(apEV_event_Kepgaleria + 1)
, log2w_wEV_Visits = log(wEV_Visits + 1)
, log2w_everEV_action_Mobilelhelyezkedes_Linkethasznalt = log(everEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
, log2w_everEV_HirdetesEladasmodja_Elado = log(everEV_HirdetesEladasmodja_Elado + 1)
, log2w_everEV_HirdetesTipus_Lakossag = log(everEV_HirdetesTipus_Lakossag + 1)
, log2w_everEV_timeOnSite = log(everEV_timeOnSite + 1)
, log2w_apEV_Visits = log(apEV_Visits + 1)
, log2w_wEV_event_Kepgaleria = log(wEV_event_Kepgaleria + 1)
, log2w_everEV_hits = log(everEV_hits + 1)
, log2w_apEV_action_Mobilelhelyezkedes_Linkethasznalt = log(apEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
, log2w_apEV_Bejelentkezve_Magan = log(apEV_Bejelentkezve_Magan + 1)
, log2w_everEV_event_Eszkoztar = log(everEV_event_Eszkoztar + 1)
, log2w_wEV_action_Kereso_ReszletesSzukitoLenyit = log(wEV_action_Kereso_ReszletesSzukitoLenyit + 1)
, log2w_wEV_event_Kereso = log(wEV_event_Kereso + 1)
, log2w_apEV_referer_icom = log(apEV_referer_icom + 1)
, log2w_everEV_event_Kereses = log(everEV_event_Kereses + 1)
, log2w_everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
, log2w_apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
, log2w_everEV_device_desktop = log(everEV_device_desktop + 1)
, log2w_everEV_ReszletekOldalMegjelenes = log(everEV_ReszletekOldalMegjelenes + 1)
, log2w_everEV_event_ReszletekOldalMegjelenes = log(everEV_event_ReszletekOldalMegjelenes + 1)
, log2w_everEV_action_Eszkoztar_Kattintas = log(everEV_action_Eszkoztar_Kattintas + 1)
, log2w_wEV_Bejelentkezve_Magan = log(wEV_Bejelentkezve_Magan + 1)
, log2w_apEV_event_Erdeklodes = log(apEV_event_Erdeklodes + 1)
, log2w_apEV_event_Nincstalalatfrontend = log(apEV_event_Nincstalalatfrontend + 1)
, log2w_everEV_Bejelentkezve_Magan = log(everEV_Bejelentkezve_Magan + 1)
, log2w_everEV_Honnan_Hungary = log(everEV_Honnan_Hungary + 1)
, log2w_everEV_event_Kepgaleria = log(everEV_event_Kepgaleria + 1)
, log2w_everEV_action_Mentes_Sikeres = log(everEV_action_Mentes_Sikeres + 1)
, log2w_wEV_event_Erdeklodes = log(wEV_event_Erdeklodes + 1)
, log2w_everEV_event_Mentes = log(everEV_event_Mentes + 1)
, log2w_wEV_action_Mobilelhelyezkedes_Linkethasznalt = log(wEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
, log2w_apEV_event_Eszkoztar = log(apEV_event_Eszkoztar + 1)
, log2w_apEV_Budapestrol_Budapestre = log(apEV_Budapestrol_Budapestre + 1)
, log2w_wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
, log2w_wEV_referer_icom = log(wEV_referer_icom + 1)
, log2w_apEV_event_FrontendBelepes = log(apEV_event_FrontendBelepes + 1)
, log2w_wEV_action_Eszkoztar_Kattintas = log(wEV_action_Eszkoztar_Kattintas + 1)
, log2w_everEV_action_Kereses_ujEpitesuSzuro = log(everEV_action_Kereses_ujEpitesuSzuro + 1)
, log2w_wEV_event_Eszkoztar = log(wEV_event_Eszkoztar + 1)
, log2w_wEV_Budapestrol_Budapestre = log(wEV_Budapestrol_Budapestre + 1)
, log2w_apEV_channel_Organic_Search = log(apEV_channel_Organic_Search + 1)
, log2w_apEV_action_Eszkoztar_Kattintas = log(apEV_action_Eszkoztar_Kattintas + 1)
, log2w_apEV_action_FrontendBelepes_Megjelenes = log(apEV_action_FrontendBelepes_Megjelenes + 1)
, log2w_apEV_trafficSource_google_organic = log(apEV_trafficSource_google_organic + 1)
, log2w_apEV_event_M1X = log(apEV_event_M1X + 1)
, log2w_everEV_socialNetwork_Facebook = log(everEV_socialNetwork_Facebook + 1)
, log2w_apEV_referer_google = log(apEV_referer_google + 1)
, log2w_everEV_SocialSourceReferral = log(everEV_SocialSourceReferral + 1)
, log2w_everEV_channel_Social = log(everEV_channel_Social + 1)
, log2w_everEV_action_FrontendBelepes_Megjelenes = log(everEV_action_FrontendBelepes_Megjelenes + 1)
, log2w_everEV_event_FrontendBelepes = log(everEV_event_FrontendBelepes + 1)
, log2w_everEV_trafficSource_facebook_referral = log(everEV_trafficSource_facebook_referral + 1)
, log2w_everEV_event_Nincstalalatfrontend = log(everEV_event_Nincstalalatfrontend + 1)
, log2w_wEV_event_FrontendBelepes = log(wEV_event_FrontendBelepes + 1)
, log2w_everEV_channel_Organic_Search = log(everEV_channel_Organic_Search + 1)
, log2w_everEV_trafficSource_google_organic = log(everEV_trafficSource_google_organic + 1)
, log2w_everEV_Budapestrol_Budapestre = log(everEV_Budapestrol_Budapestre + 1)
, log2w_everEV_event_lakoparkMegjelenes = log(everEV_event_lakoparkMegjelenes + 1)
, log2w_wEV_trafficSource_facebook_referral = log(wEV_trafficSource_facebook_referral + 1)
, log2w_wEV_socialNetwork_Facebook = log(wEV_socialNetwork_Facebook + 1)
, log2w_wEV_SocialSourceReferral = log(wEV_SocialSourceReferral + 1)
, log2w_wEV_channel_Social = log(wEV_channel_Social + 1)
, log2w_everEV_action_Kereses_mobil = log(everEV_action_Kereses_mobil + 1)
, log2w_everEV_label_FrontendBelepes_Megjelenes_fejlec = log(everEV_label_FrontendBelepes_Megjelenes_fejlec + 1)
, log2w_everEV_event_Erdeklodes = log(everEV_event_Erdeklodes + 1)
, log2w_everEV_event_M1X = log(everEV_event_M1X + 1)
, log2w_everEV_channel_Email = log(everEV_channel_Email + 1)
, log2w_apEV_trafficSource_google_cpc = log(apEV_trafficSource_google_cpc + 1)
, log2w_apEV_trafficSource_facebook_referral = log(apEV_trafficSource_facebook_referral + 1)
, log2w_apEV_socialNetwork_Facebook = log(apEV_socialNetwork_Facebook + 1)
, log2w_apEV_SocialSourceReferral = log(apEV_SocialSourceReferral + 1)
, log2w_apEV_channel_Social = log(apEV_channel_Social + 1)
, log2w_wEV_event_M1X = log(wEV_event_M1X + 1)
, log2w_apEV_channel_Paid_Search = log(apEV_channel_Paid_Search + 1)
, log2w_apEV_action_M1X_talalkozottDuplikaltHirdetessel = log(apEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
, log2w_apEV_action_Erdeklodes_OfficeUtcapontos = log(apEV_action_Erdeklodes_OfficeUtcapontos + 1)
, log2w_everEV_trafficSource_Hirdetesfigyelo_email = log(everEV_trafficSource_Hirdetesfigyelo_email + 1)
, log2w_wEV_channel_Email = log(wEV_channel_Email + 1)
, log2w_wEV_channel_Organic_Search = log(wEV_channel_Organic_Search + 1)
, log2w_apEV_channel_Email = log(apEV_channel_Email + 1)
, log2w_wEV_referer_google = log(wEV_referer_google + 1)
, log2w_wEV_trafficSource_google_organic = log(wEV_trafficSource_google_organic + 1)
, log2w_everEV_event_AI = log(everEV_event_AI + 1)
, log2w_apEV_action_Erdeklodes_OfficeNemUtcapontos = log(apEV_action_Erdeklodes_OfficeNemUtcapontos + 1)
, log2w_everEV_action_M1X_talalkozottDuplikaltHirdetessel = log(everEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
, log2w_everEV_action_FrontendBelepes_Sikeres = log(everEV_action_FrontendBelepes_Sikeres + 1)
, log2w_everEV_label_FrontendBelepes_Sikeres_fejlec = log(everEV_label_FrontendBelepes_Sikeres_fejlec + 1)
, log2w_everEV_trafficSource_google_cpc = log(everEV_trafficSource_google_cpc + 1)
, log2w_wEV_trafficSource_Hirdetesfigyelo_email = log(wEV_trafficSource_Hirdetesfigyelo_email + 1)
, log2w_everEV_channel_Paid_Search = log(everEV_channel_Paid_Search + 1)
, log2w_everEV_event_OfficeLogo = log(everEV_event_OfficeLogo + 1)
, log2w_everEV_event_lakoparkListaKattintas = log(everEV_event_lakoparkListaKattintas + 1)
, log2w_wEV_action_M1X_talalkozottDuplikaltHirdetessel = log(wEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
, log2w_apEV_trafficSource_Hirdetesfigyelo_email = log(apEV_trafficSource_Hirdetesfigyelo_email + 1)
, log2w_everEV_referer_icom = log(everEV_referer_icom + 1)
, log2w_wEV_trafficSource_google_cpc = log(wEV_trafficSource_google_cpc + 1)
, log2w_everEV_Felhasznalo_Magan = log(everEV_Felhasznalo_Magan + 1)
, log2w_everEV_event_Piac = log(everEV_event_Piac + 1)
, log2w_everEV_action_Piac_undefined = log(everEV_action_Piac_undefined + 1)
, log2w_everEV_referer_google = log(everEV_referer_google + 1)
, log2w_everEV_channel_Direct = log(everEV_channel_Direct + 1)
, log2w_everRE_AVG_Kepgaleria = log(everRE_AVG_Kepgaleria + 1)
, log2w_wEV_channel_Paid_Search = log(wEV_channel_Paid_Search + 1)
, log2w_everRE_MAX_Kepgaleria = log(everRE_MAX_Kepgaleria + 1)
, log2w_everRE_SUM_Kepgaleria = log(everRE_SUM_Kepgaleria + 1)
, log2w_everRE_megye_Budapest = log(everRE_megye_Budapest + 1)
, log2w_everEV_event_Mobilreszletek = log(everEV_event_Mobilreszletek + 1)
, log2w_apRE_AVG_Kepgaleria = log(apRE_AVG_Kepgaleria + 1)
, log2w_apRE_MAX_Kepgaleria = log(apRE_MAX_Kepgaleria + 1)
, log2w_apRE_SUM_Kepgaleria = log(apRE_SUM_Kepgaleria + 1)
, log2w_apEV_action_Mobilreszletek_Kepgaleria = log(apEV_action_Mobilreszletek_Kepgaleria + 1)
, log2w_everEV_action_Erdeklodes_OfficeUtcapontos = log(everEV_action_Erdeklodes_OfficeUtcapontos + 1)
, log2w_apEV_trafficSource_rtbhouse_retargeting = log(apEV_trafficSource_rtbhouse_retargeting + 1)
, log2w_everRE_ikerhaz = log(everRE_ikerhaz + 1)
, log2w_wEV_action_Mobilreszletek_Kepgaleria = log(wEV_action_Mobilreszletek_Kepgaleria + 1)
, log2w_wEV_trafficSource_rtbhouse_retargeting = log(wEV_trafficSource_rtbhouse_retargeting + 1)
, log2w_everEV_Visits = log(everEV_Visits + 1)
, log2w_everEV_action_Mobilreszletek_Kepgaleria = log(everEV_action_Mobilreszletek_Kepgaleria + 1)
, log2w_apEV_event_Mobilreszletek = log(apEV_event_Mobilreszletek + 1)
, log2w_everEV_action_Erdeklodes_OfficeNemUtcapontos = log(everEV_action_Erdeklodes_OfficeNemUtcapontos + 1)
, log2w_wEV_event_Mobilreszletek = log(wEV_event_Mobilreszletek + 1)
, log2w_everEV_action_Erdeklodes_MaganFizetos = log(everEV_action_Erdeklodes_MaganFizetos + 1)
, log2w_wRE_AVG_Kepgaleria = log(wRE_AVG_Kepgaleria + 1)
, log2w_wEV_referer_creativecdn = log(wEV_referer_creativecdn + 1)
, log2w_wRE_AVG_Hirdetesar = log(wRE_AVG_Hirdetesar + 1)
, log2w_wRE_MAX_Kepgaleria = log(wRE_MAX_Kepgaleria + 1)
, log2w_wRE_SUM_Kepgaleria = log(wRE_SUM_Kepgaleria + 1)
, log2w_apRE_megye_Budapest = log(apRE_megye_Budapest + 1)
, log2w_everRE_megye_Pest = log(everRE_megye_Pest + 1)
, log2w_everEV_event_MaganhirdetoInfok = log(everEV_event_MaganhirdetoInfok + 1)
, log2w_wRE_nyaralo = log(wRE_nyaralo + 1)
, log2w_everRE_ujszeru = log(everRE_ujszeru + 1)
, log2w_apRE_AVG_Hirdetesar = log(apRE_AVG_Hirdetesar + 1)
, log2w_apEV_referer_creativecdn = log(apEV_referer_creativecdn + 1)
, log2w_wRE_megye_Budapest = log(wRE_megye_Budapest + 1)
, log2w_everEV_referer_creativecdn = log(everEV_referer_creativecdn + 1)
, log2w_everEV_Videkrol_Budapestre = log(everEV_Videkrol_Budapestre + 1)
, log2w_apEV_Videkrol_Budapestre = log(apEV_Videkrol_Budapestre + 1)
, log2w_wRE_MAX_Hirdetesar = log(wRE_MAX_Hirdetesar + 1)
, log2w_everRE_ker_11 = log(everRE_ker_11 + 1)
, log2w_everEV_channel_Display = log(everEV_channel_Display + 1)
, log2w_everEV_MAX_visitNumber = log(everEV_MAX_visitNumber + 1)
, log2w_everEV_MIN_visitNumber = log(everEV_MIN_visitNumber + 1)
, log2w_wRE_MAX_DIFF_Visit_per_RE = log(wRE_MAX_DIFF_Visit_per_RE + 1)
, log2w_everRE_AVG_Hirdetesar = log(everRE_AVG_Hirdetesar + 1)
, log2w_wRE_AVG_DIFF_Visit_per_RE = log(wRE_AVG_DIFF_Visit_per_RE + 1)
, log2w_everEV_channel_Referral = log(everEV_channel_Referral + 1)
, log2w_everEV_trafficSource_rtbhouse_retargeting = log(everEV_trafficSource_rtbhouse_retargeting + 1)
, log2w_everRE_tegla = log(everRE_tegla + 1)
, log2w_apEV_event_MaganhirdetoInfok = log(apEV_event_MaganhirdetoInfok + 1)
, log2w_everEV_action_Erdeklodes_MaganIngyenes = log(everEV_action_Erdeklodes_MaganIngyenes + 1)
, log2w_wEV_Videkrol_Budapestre = log(wEV_Videkrol_Budapestre + 1)
, log2w_apRE_tegla = log(apRE_tegla + 1)
, log2w_everRE_felujitott = log(everRE_felujitott + 1)
, log2w_everRE_lakas = log(everRE_lakas + 1)
, log2w_apRE_megye_Pest = log(apRE_megye_Pest + 1)
, log2w_apRE_lakas = log(apRE_lakas + 1)
, log2w_everEV_action_MaganToDo_Megjelenes = log(everEV_action_MaganToDo_Megjelenes + 1)
, log2w_everRE_SUM_DIFF_Visit_per_RE = log(everRE_SUM_DIFF_Visit_per_RE + 1)

),]

```

Test
```{r}
JZ_test <- JZ_test[, `:=`(
log2w_apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
, log2w_everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
, log2w_everEV_event_Mobilelhelyezkedes = log(everEV_event_Mobilelhelyezkedes + 1)
, log2w_apEV_pageviews = log(apEV_pageviews + 1)
, log2w_apEV_HirdetesTipus_Lakossag = log(apEV_HirdetesTipus_Lakossag + 1)
, log2w_apEV_HirdetesEladasmodja_Elado = log(apEV_HirdetesEladasmodja_Elado + 1)
, log2w_apEV_event_Mobilelhelyezkedes = log(apEV_event_Mobilelhelyezkedes + 1)
, log2w_apEV_timeOnSite = log(apEV_timeOnSite + 1)
, log2w_everEV_event_autosuggest = log(everEV_event_autosuggest + 1)
, log2w_apEV_event_autosuggest = log(apEV_event_autosuggest + 1)
, log2w_apEV_hits = log(apEV_hits + 1)
, log2w_wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
, log2w_apEV_device_desktop = log(apEV_device_desktop + 1)
, log2w_wEV_event_Mobilelhelyezkedes = log(wEV_event_Mobilelhelyezkedes + 1)
, log2w_wEV_pageviews = log(wEV_pageviews + 1)
, log2w_wEV_event_autosuggest = log(wEV_event_autosuggest + 1)
, log2w_wEV_timeOnSite = log(wEV_timeOnSite + 1)
, log2w_wEV_event_ReszletekOldalMegjelenes = log(wEV_event_ReszletekOldalMegjelenes + 1)
, log2w_apEV_event_ReszletekOldalMegjelenes = log(apEV_event_ReszletekOldalMegjelenes + 1)
, log2w_everEV_action_Kereso_ReszletesSzukitoLenyit = log(everEV_action_Kereso_ReszletesSzukitoLenyit + 1)
, log2w_wEV_HirdetesEladasmodja_Elado = log(wEV_HirdetesEladasmodja_Elado + 1)
, log2w_wEV_HirdetesTipus_Lakossag = log(wEV_HirdetesTipus_Lakossag + 1)
, log2w_everEV_event_Kereso = log(everEV_event_Kereso + 1)
, log2w_wEV_hits = log(wEV_hits + 1)
, log2w_apEV_Honnan_Hungary = log(apEV_Honnan_Hungary + 1)
, log2w_wEV_device_desktop = log(wEV_device_desktop + 1)
, log2w_apEV_event_Kereso = log(apEV_event_Kereso + 1)
, log2w_wEV_Honnan_Hungary = log(wEV_Honnan_Hungary + 1)
, log2w_apEV_action_Kereso_ReszletesSzukitoLenyit = log(apEV_action_Kereso_ReszletesSzukitoLenyit + 1)
, log2w_everEV_pageviews = log(everEV_pageviews + 1)
, log2w_apEV_event_Kepgaleria = log(apEV_event_Kepgaleria + 1)
, log2w_wEV_Visits = log(wEV_Visits + 1)
, log2w_everEV_action_Mobilelhelyezkedes_Linkethasznalt = log(everEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
, log2w_everEV_HirdetesEladasmodja_Elado = log(everEV_HirdetesEladasmodja_Elado + 1)
, log2w_everEV_HirdetesTipus_Lakossag = log(everEV_HirdetesTipus_Lakossag + 1)
, log2w_everEV_timeOnSite = log(everEV_timeOnSite + 1)
, log2w_apEV_Visits = log(apEV_Visits + 1)
, log2w_wEV_event_Kepgaleria = log(wEV_event_Kepgaleria + 1)
, log2w_everEV_hits = log(everEV_hits + 1)
, log2w_apEV_action_Mobilelhelyezkedes_Linkethasznalt = log(apEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
, log2w_apEV_Bejelentkezve_Magan = log(apEV_Bejelentkezve_Magan + 1)
, log2w_everEV_event_Eszkoztar = log(everEV_event_Eszkoztar + 1)
, log2w_wEV_action_Kereso_ReszletesSzukitoLenyit = log(wEV_action_Kereso_ReszletesSzukitoLenyit + 1)
, log2w_wEV_event_Kereso = log(wEV_event_Kereso + 1)
, log2w_apEV_referer_icom = log(apEV_referer_icom + 1)
, log2w_everEV_event_Kereses = log(everEV_event_Kereses + 1)
, log2w_everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
, log2w_apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
, log2w_everEV_device_desktop = log(everEV_device_desktop + 1)
, log2w_everEV_ReszletekOldalMegjelenes = log(everEV_ReszletekOldalMegjelenes + 1)
, log2w_everEV_event_ReszletekOldalMegjelenes = log(everEV_event_ReszletekOldalMegjelenes + 1)
, log2w_everEV_action_Eszkoztar_Kattintas = log(everEV_action_Eszkoztar_Kattintas + 1)
, log2w_wEV_Bejelentkezve_Magan = log(wEV_Bejelentkezve_Magan + 1)
, log2w_apEV_event_Erdeklodes = log(apEV_event_Erdeklodes + 1)
, log2w_apEV_event_Nincstalalatfrontend = log(apEV_event_Nincstalalatfrontend + 1)
, log2w_everEV_Bejelentkezve_Magan = log(everEV_Bejelentkezve_Magan + 1)
, log2w_everEV_Honnan_Hungary = log(everEV_Honnan_Hungary + 1)
, log2w_everEV_event_Kepgaleria = log(everEV_event_Kepgaleria + 1)
, log2w_everEV_action_Mentes_Sikeres = log(everEV_action_Mentes_Sikeres + 1)
, log2w_wEV_event_Erdeklodes = log(wEV_event_Erdeklodes + 1)
, log2w_everEV_event_Mentes = log(everEV_event_Mentes + 1)
, log2w_wEV_action_Mobilelhelyezkedes_Linkethasznalt = log(wEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
, log2w_apEV_event_Eszkoztar = log(apEV_event_Eszkoztar + 1)
, log2w_apEV_Budapestrol_Budapestre = log(apEV_Budapestrol_Budapestre + 1)
, log2w_wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
, log2w_wEV_referer_icom = log(wEV_referer_icom + 1)
, log2w_apEV_event_FrontendBelepes = log(apEV_event_FrontendBelepes + 1)
, log2w_wEV_action_Eszkoztar_Kattintas = log(wEV_action_Eszkoztar_Kattintas + 1)
, log2w_everEV_action_Kereses_ujEpitesuSzuro = log(everEV_action_Kereses_ujEpitesuSzuro + 1)
, log2w_wEV_event_Eszkoztar = log(wEV_event_Eszkoztar + 1)
, log2w_wEV_Budapestrol_Budapestre = log(wEV_Budapestrol_Budapestre + 1)
, log2w_apEV_channel_Organic_Search = log(apEV_channel_Organic_Search + 1)
, log2w_apEV_action_Eszkoztar_Kattintas = log(apEV_action_Eszkoztar_Kattintas + 1)
, log2w_apEV_action_FrontendBelepes_Megjelenes = log(apEV_action_FrontendBelepes_Megjelenes + 1)
, log2w_apEV_trafficSource_google_organic = log(apEV_trafficSource_google_organic + 1)
, log2w_apEV_event_M1X = log(apEV_event_M1X + 1)
, log2w_everEV_socialNetwork_Facebook = log(everEV_socialNetwork_Facebook + 1)
, log2w_apEV_referer_google = log(apEV_referer_google + 1)
, log2w_everEV_SocialSourceReferral = log(everEV_SocialSourceReferral + 1)
, log2w_everEV_channel_Social = log(everEV_channel_Social + 1)
, log2w_everEV_action_FrontendBelepes_Megjelenes = log(everEV_action_FrontendBelepes_Megjelenes + 1)
, log2w_everEV_event_FrontendBelepes = log(everEV_event_FrontendBelepes + 1)
, log2w_everEV_trafficSource_facebook_referral = log(everEV_trafficSource_facebook_referral + 1)
, log2w_everEV_event_Nincstalalatfrontend = log(everEV_event_Nincstalalatfrontend + 1)
, log2w_wEV_event_FrontendBelepes = log(wEV_event_FrontendBelepes + 1)
, log2w_everEV_channel_Organic_Search = log(everEV_channel_Organic_Search + 1)
, log2w_everEV_trafficSource_google_organic = log(everEV_trafficSource_google_organic + 1)
, log2w_everEV_Budapestrol_Budapestre = log(everEV_Budapestrol_Budapestre + 1)
, log2w_everEV_event_lakoparkMegjelenes = log(everEV_event_lakoparkMegjelenes + 1)
, log2w_wEV_trafficSource_facebook_referral = log(wEV_trafficSource_facebook_referral + 1)
, log2w_wEV_socialNetwork_Facebook = log(wEV_socialNetwork_Facebook + 1)
, log2w_wEV_SocialSourceReferral = log(wEV_SocialSourceReferral + 1)
, log2w_wEV_channel_Social = log(wEV_channel_Social + 1)
, log2w_everEV_action_Kereses_mobil = log(everEV_action_Kereses_mobil + 1)
, log2w_everEV_label_FrontendBelepes_Megjelenes_fejlec = log(everEV_label_FrontendBelepes_Megjelenes_fejlec + 1)
, log2w_everEV_event_Erdeklodes = log(everEV_event_Erdeklodes + 1)
, log2w_everEV_event_M1X = log(everEV_event_M1X + 1)
, log2w_everEV_channel_Email = log(everEV_channel_Email + 1)
, log2w_apEV_trafficSource_google_cpc = log(apEV_trafficSource_google_cpc + 1)
, log2w_apEV_trafficSource_facebook_referral = log(apEV_trafficSource_facebook_referral + 1)
, log2w_apEV_socialNetwork_Facebook = log(apEV_socialNetwork_Facebook + 1)
, log2w_apEV_SocialSourceReferral = log(apEV_SocialSourceReferral + 1)
, log2w_apEV_channel_Social = log(apEV_channel_Social + 1)
, log2w_wEV_event_M1X = log(wEV_event_M1X + 1)
, log2w_apEV_channel_Paid_Search = log(apEV_channel_Paid_Search + 1)
, log2w_apEV_action_M1X_talalkozottDuplikaltHirdetessel = log(apEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
, log2w_apEV_action_Erdeklodes_OfficeUtcapontos = log(apEV_action_Erdeklodes_OfficeUtcapontos + 1)
, log2w_everEV_trafficSource_Hirdetesfigyelo_email = log(everEV_trafficSource_Hirdetesfigyelo_email + 1)
, log2w_wEV_channel_Email = log(wEV_channel_Email + 1)
, log2w_wEV_channel_Organic_Search = log(wEV_channel_Organic_Search + 1)
, log2w_apEV_channel_Email = log(apEV_channel_Email + 1)
, log2w_wEV_referer_google = log(wEV_referer_google + 1)
, log2w_wEV_trafficSource_google_organic = log(wEV_trafficSource_google_organic + 1)
, log2w_everEV_event_AI = log(everEV_event_AI + 1)
, log2w_apEV_action_Erdeklodes_OfficeNemUtcapontos = log(apEV_action_Erdeklodes_OfficeNemUtcapontos + 1)
, log2w_everEV_action_M1X_talalkozottDuplikaltHirdetessel = log(everEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
, log2w_everEV_action_FrontendBelepes_Sikeres = log(everEV_action_FrontendBelepes_Sikeres + 1)
, log2w_everEV_label_FrontendBelepes_Sikeres_fejlec = log(everEV_label_FrontendBelepes_Sikeres_fejlec + 1)
, log2w_everEV_trafficSource_google_cpc = log(everEV_trafficSource_google_cpc + 1)
, log2w_wEV_trafficSource_Hirdetesfigyelo_email = log(wEV_trafficSource_Hirdetesfigyelo_email + 1)
, log2w_everEV_channel_Paid_Search = log(everEV_channel_Paid_Search + 1)
, log2w_everEV_event_OfficeLogo = log(everEV_event_OfficeLogo + 1)
, log2w_everEV_event_lakoparkListaKattintas = log(everEV_event_lakoparkListaKattintas + 1)
, log2w_wEV_action_M1X_talalkozottDuplikaltHirdetessel = log(wEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
, log2w_apEV_trafficSource_Hirdetesfigyelo_email = log(apEV_trafficSource_Hirdetesfigyelo_email + 1)
, log2w_everEV_referer_icom = log(everEV_referer_icom + 1)
, log2w_wEV_trafficSource_google_cpc = log(wEV_trafficSource_google_cpc + 1)
, log2w_everEV_Felhasznalo_Magan = log(everEV_Felhasznalo_Magan + 1)
, log2w_everEV_event_Piac = log(everEV_event_Piac + 1)
, log2w_everEV_action_Piac_undefined = log(everEV_action_Piac_undefined + 1)
, log2w_everEV_referer_google = log(everEV_referer_google + 1)
, log2w_everEV_channel_Direct = log(everEV_channel_Direct + 1)
, log2w_everRE_AVG_Kepgaleria = log(everRE_AVG_Kepgaleria + 1)
, log2w_wEV_channel_Paid_Search = log(wEV_channel_Paid_Search + 1)
, log2w_everRE_MAX_Kepgaleria = log(everRE_MAX_Kepgaleria + 1)
, log2w_everRE_SUM_Kepgaleria = log(everRE_SUM_Kepgaleria + 1)
, log2w_everRE_megye_Budapest = log(everRE_megye_Budapest + 1)
, log2w_everEV_event_Mobilreszletek = log(everEV_event_Mobilreszletek + 1)
, log2w_apRE_AVG_Kepgaleria = log(apRE_AVG_Kepgaleria + 1)
, log2w_apRE_MAX_Kepgaleria = log(apRE_MAX_Kepgaleria + 1)
, log2w_apRE_SUM_Kepgaleria = log(apRE_SUM_Kepgaleria + 1)
, log2w_apEV_action_Mobilreszletek_Kepgaleria = log(apEV_action_Mobilreszletek_Kepgaleria + 1)
, log2w_everEV_action_Erdeklodes_OfficeUtcapontos = log(everEV_action_Erdeklodes_OfficeUtcapontos + 1)
, log2w_apEV_trafficSource_rtbhouse_retargeting = log(apEV_trafficSource_rtbhouse_retargeting + 1)
, log2w_everRE_ikerhaz = log(everRE_ikerhaz + 1)
, log2w_wEV_action_Mobilreszletek_Kepgaleria = log(wEV_action_Mobilreszletek_Kepgaleria + 1)
, log2w_wEV_trafficSource_rtbhouse_retargeting = log(wEV_trafficSource_rtbhouse_retargeting + 1)
, log2w_everEV_Visits = log(everEV_Visits + 1)
, log2w_everEV_action_Mobilreszletek_Kepgaleria = log(everEV_action_Mobilreszletek_Kepgaleria + 1)
, log2w_apEV_event_Mobilreszletek = log(apEV_event_Mobilreszletek + 1)
, log2w_everEV_action_Erdeklodes_OfficeNemUtcapontos = log(everEV_action_Erdeklodes_OfficeNemUtcapontos + 1)
, log2w_wEV_event_Mobilreszletek = log(wEV_event_Mobilreszletek + 1)
, log2w_everEV_action_Erdeklodes_MaganFizetos = log(everEV_action_Erdeklodes_MaganFizetos + 1)
, log2w_wRE_AVG_Kepgaleria = log(wRE_AVG_Kepgaleria + 1)
, log2w_wEV_referer_creativecdn = log(wEV_referer_creativecdn + 1)
, log2w_wRE_AVG_Hirdetesar = log(wRE_AVG_Hirdetesar + 1)
, log2w_wRE_MAX_Kepgaleria = log(wRE_MAX_Kepgaleria + 1)
, log2w_wRE_SUM_Kepgaleria = log(wRE_SUM_Kepgaleria + 1)
, log2w_apRE_megye_Budapest = log(apRE_megye_Budapest + 1)
, log2w_everRE_megye_Pest = log(everRE_megye_Pest + 1)
, log2w_everEV_event_MaganhirdetoInfok = log(everEV_event_MaganhirdetoInfok + 1)
, log2w_wRE_nyaralo = log(wRE_nyaralo + 1)
, log2w_everRE_ujszeru = log(everRE_ujszeru + 1)
, log2w_apRE_AVG_Hirdetesar = log(apRE_AVG_Hirdetesar + 1)
, log2w_apEV_referer_creativecdn = log(apEV_referer_creativecdn + 1)
, log2w_wRE_megye_Budapest = log(wRE_megye_Budapest + 1)
, log2w_everEV_referer_creativecdn = log(everEV_referer_creativecdn + 1)
, log2w_everEV_Videkrol_Budapestre = log(everEV_Videkrol_Budapestre + 1)
, log2w_apEV_Videkrol_Budapestre = log(apEV_Videkrol_Budapestre + 1)
, log2w_wRE_MAX_Hirdetesar = log(wRE_MAX_Hirdetesar + 1)
, log2w_everRE_ker_11 = log(everRE_ker_11 + 1)
, log2w_everEV_channel_Display = log(everEV_channel_Display + 1)
, log2w_everEV_MAX_visitNumber = log(everEV_MAX_visitNumber + 1)
, log2w_everEV_MIN_visitNumber = log(everEV_MIN_visitNumber + 1)
, log2w_wRE_MAX_DIFF_Visit_per_RE = log(wRE_MAX_DIFF_Visit_per_RE + 1)
, log2w_everRE_AVG_Hirdetesar = log(everRE_AVG_Hirdetesar + 1)
, log2w_wRE_AVG_DIFF_Visit_per_RE = log(wRE_AVG_DIFF_Visit_per_RE + 1)
, log2w_everEV_channel_Referral = log(everEV_channel_Referral + 1)
, log2w_everEV_trafficSource_rtbhouse_retargeting = log(everEV_trafficSource_rtbhouse_retargeting + 1)
, log2w_everRE_tegla = log(everRE_tegla + 1)
, log2w_apEV_event_MaganhirdetoInfok = log(apEV_event_MaganhirdetoInfok + 1)
, log2w_everEV_action_Erdeklodes_MaganIngyenes = log(everEV_action_Erdeklodes_MaganIngyenes + 1)
, log2w_wEV_Videkrol_Budapestre = log(wEV_Videkrol_Budapestre + 1)
, log2w_apRE_tegla = log(apRE_tegla + 1)
, log2w_everRE_felujitott = log(everRE_felujitott + 1)
, log2w_everRE_lakas = log(everRE_lakas + 1)
, log2w_apRE_megye_Pest = log(apRE_megye_Pest + 1)
, log2w_apRE_lakas = log(apRE_lakas + 1)
, log2w_everEV_action_MaganToDo_Megjelenes = log(everEV_action_MaganToDo_Megjelenes + 1)
, log2w_everRE_SUM_DIFF_Visit_per_RE = log(everRE_SUM_DIFF_Visit_per_RE + 1)

),]

```

External
```{r}
JZ_external <- JZ_external[, `:=`(
log2w_apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
, log2w_everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
, log2w_everEV_event_Mobilelhelyezkedes = log(everEV_event_Mobilelhelyezkedes + 1)
, log2w_apEV_pageviews = log(apEV_pageviews + 1)
, log2w_apEV_HirdetesTipus_Lakossag = log(apEV_HirdetesTipus_Lakossag + 1)
, log2w_apEV_HirdetesEladasmodja_Elado = log(apEV_HirdetesEladasmodja_Elado + 1)
, log2w_apEV_event_Mobilelhelyezkedes = log(apEV_event_Mobilelhelyezkedes + 1)
, log2w_apEV_timeOnSite = log(apEV_timeOnSite + 1)
, log2w_everEV_event_autosuggest = log(everEV_event_autosuggest + 1)
, log2w_apEV_event_autosuggest = log(apEV_event_autosuggest + 1)
, log2w_apEV_hits = log(apEV_hits + 1)
, log2w_wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik = log(wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik + 1)
, log2w_apEV_device_desktop = log(apEV_device_desktop + 1)
, log2w_wEV_event_Mobilelhelyezkedes = log(wEV_event_Mobilelhelyezkedes + 1)
, log2w_wEV_pageviews = log(wEV_pageviews + 1)
, log2w_wEV_event_autosuggest = log(wEV_event_autosuggest + 1)
, log2w_wEV_timeOnSite = log(wEV_timeOnSite + 1)
, log2w_wEV_event_ReszletekOldalMegjelenes = log(wEV_event_ReszletekOldalMegjelenes + 1)
, log2w_apEV_event_ReszletekOldalMegjelenes = log(apEV_event_ReszletekOldalMegjelenes + 1)
, log2w_everEV_action_Kereso_ReszletesSzukitoLenyit = log(everEV_action_Kereso_ReszletesSzukitoLenyit + 1)
, log2w_wEV_HirdetesEladasmodja_Elado = log(wEV_HirdetesEladasmodja_Elado + 1)
, log2w_wEV_HirdetesTipus_Lakossag = log(wEV_HirdetesTipus_Lakossag + 1)
, log2w_everEV_event_Kereso = log(everEV_event_Kereso + 1)
, log2w_wEV_hits = log(wEV_hits + 1)
, log2w_apEV_Honnan_Hungary = log(apEV_Honnan_Hungary + 1)
, log2w_wEV_device_desktop = log(wEV_device_desktop + 1)
, log2w_apEV_event_Kereso = log(apEV_event_Kereso + 1)
, log2w_wEV_Honnan_Hungary = log(wEV_Honnan_Hungary + 1)
, log2w_apEV_action_Kereso_ReszletesSzukitoLenyit = log(apEV_action_Kereso_ReszletesSzukitoLenyit + 1)
, log2w_everEV_pageviews = log(everEV_pageviews + 1)
, log2w_apEV_event_Kepgaleria = log(apEV_event_Kepgaleria + 1)
, log2w_wEV_Visits = log(wEV_Visits + 1)
, log2w_everEV_action_Mobilelhelyezkedes_Linkethasznalt = log(everEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
, log2w_everEV_HirdetesEladasmodja_Elado = log(everEV_HirdetesEladasmodja_Elado + 1)
, log2w_everEV_HirdetesTipus_Lakossag = log(everEV_HirdetesTipus_Lakossag + 1)
, log2w_everEV_timeOnSite = log(everEV_timeOnSite + 1)
, log2w_apEV_Visits = log(apEV_Visits + 1)
, log2w_wEV_event_Kepgaleria = log(wEV_event_Kepgaleria + 1)
, log2w_everEV_hits = log(everEV_hits + 1)
, log2w_apEV_action_Mobilelhelyezkedes_Linkethasznalt = log(apEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
, log2w_apEV_Bejelentkezve_Magan = log(apEV_Bejelentkezve_Magan + 1)
, log2w_everEV_event_Eszkoztar = log(everEV_event_Eszkoztar + 1)
, log2w_wEV_action_Kereso_ReszletesSzukitoLenyit = log(wEV_action_Kereso_ReszletesSzukitoLenyit + 1)
, log2w_wEV_event_Kereso = log(wEV_event_Kereso + 1)
, log2w_apEV_referer_icom = log(apEV_referer_icom + 1)
, log2w_everEV_event_Kereses = log(everEV_event_Kereses + 1)
, log2w_everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(everEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
, log2w_apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(apEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
, log2w_everEV_device_desktop = log(everEV_device_desktop + 1)
, log2w_everEV_ReszletekOldalMegjelenes = log(everEV_ReszletekOldalMegjelenes + 1)
, log2w_everEV_event_ReszletekOldalMegjelenes = log(everEV_event_ReszletekOldalMegjelenes + 1)
, log2w_everEV_action_Eszkoztar_Kattintas = log(everEV_action_Eszkoztar_Kattintas + 1)
, log2w_wEV_Bejelentkezve_Magan = log(wEV_Bejelentkezve_Magan + 1)
, log2w_apEV_event_Erdeklodes = log(apEV_event_Erdeklodes + 1)
, log2w_apEV_event_Nincstalalatfrontend = log(apEV_event_Nincstalalatfrontend + 1)
, log2w_everEV_Bejelentkezve_Magan = log(everEV_Bejelentkezve_Magan + 1)
, log2w_everEV_Honnan_Hungary = log(everEV_Honnan_Hungary + 1)
, log2w_everEV_event_Kepgaleria = log(everEV_event_Kepgaleria + 1)
, log2w_everEV_action_Mentes_Sikeres = log(everEV_action_Mentes_Sikeres + 1)
, log2w_wEV_event_Erdeklodes = log(wEV_event_Erdeklodes + 1)
, log2w_everEV_event_Mentes = log(everEV_event_Mentes + 1)
, log2w_wEV_action_Mobilelhelyezkedes_Linkethasznalt = log(wEV_action_Mobilelhelyezkedes_Linkethasznalt + 1)
, log2w_apEV_event_Eszkoztar = log(apEV_event_Eszkoztar + 1)
, log2w_apEV_Budapestrol_Budapestre = log(apEV_Budapestrol_Budapestre + 1)
, log2w_wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest = log(wEV_label_Eszkoztar_Kattintas_Jelentemahirdetest + 1)
, log2w_wEV_referer_icom = log(wEV_referer_icom + 1)
, log2w_apEV_event_FrontendBelepes = log(apEV_event_FrontendBelepes + 1)
, log2w_wEV_action_Eszkoztar_Kattintas = log(wEV_action_Eszkoztar_Kattintas + 1)
, log2w_everEV_action_Kereses_ujEpitesuSzuro = log(everEV_action_Kereses_ujEpitesuSzuro + 1)
, log2w_wEV_event_Eszkoztar = log(wEV_event_Eszkoztar + 1)
, log2w_wEV_Budapestrol_Budapestre = log(wEV_Budapestrol_Budapestre + 1)
, log2w_apEV_channel_Organic_Search = log(apEV_channel_Organic_Search + 1)
, log2w_apEV_action_Eszkoztar_Kattintas = log(apEV_action_Eszkoztar_Kattintas + 1)
, log2w_apEV_action_FrontendBelepes_Megjelenes = log(apEV_action_FrontendBelepes_Megjelenes + 1)
, log2w_apEV_trafficSource_google_organic = log(apEV_trafficSource_google_organic + 1)
, log2w_apEV_event_M1X = log(apEV_event_M1X + 1)
, log2w_everEV_socialNetwork_Facebook = log(everEV_socialNetwork_Facebook + 1)
, log2w_apEV_referer_google = log(apEV_referer_google + 1)
, log2w_everEV_SocialSourceReferral = log(everEV_SocialSourceReferral + 1)
, log2w_everEV_channel_Social = log(everEV_channel_Social + 1)
, log2w_everEV_action_FrontendBelepes_Megjelenes = log(everEV_action_FrontendBelepes_Megjelenes + 1)
, log2w_everEV_event_FrontendBelepes = log(everEV_event_FrontendBelepes + 1)
, log2w_everEV_trafficSource_facebook_referral = log(everEV_trafficSource_facebook_referral + 1)
, log2w_everEV_event_Nincstalalatfrontend = log(everEV_event_Nincstalalatfrontend + 1)
, log2w_wEV_event_FrontendBelepes = log(wEV_event_FrontendBelepes + 1)
, log2w_everEV_channel_Organic_Search = log(everEV_channel_Organic_Search + 1)
, log2w_everEV_trafficSource_google_organic = log(everEV_trafficSource_google_organic + 1)
, log2w_everEV_Budapestrol_Budapestre = log(everEV_Budapestrol_Budapestre + 1)
, log2w_everEV_event_lakoparkMegjelenes = log(everEV_event_lakoparkMegjelenes + 1)
, log2w_wEV_trafficSource_facebook_referral = log(wEV_trafficSource_facebook_referral + 1)
, log2w_wEV_socialNetwork_Facebook = log(wEV_socialNetwork_Facebook + 1)
, log2w_wEV_SocialSourceReferral = log(wEV_SocialSourceReferral + 1)
, log2w_wEV_channel_Social = log(wEV_channel_Social + 1)
, log2w_everEV_action_Kereses_mobil = log(everEV_action_Kereses_mobil + 1)
, log2w_everEV_label_FrontendBelepes_Megjelenes_fejlec = log(everEV_label_FrontendBelepes_Megjelenes_fejlec + 1)
, log2w_everEV_event_Erdeklodes = log(everEV_event_Erdeklodes + 1)
, log2w_everEV_event_M1X = log(everEV_event_M1X + 1)
, log2w_everEV_channel_Email = log(everEV_channel_Email + 1)
, log2w_apEV_trafficSource_google_cpc = log(apEV_trafficSource_google_cpc + 1)
, log2w_apEV_trafficSource_facebook_referral = log(apEV_trafficSource_facebook_referral + 1)
, log2w_apEV_socialNetwork_Facebook = log(apEV_socialNetwork_Facebook + 1)
, log2w_apEV_SocialSourceReferral = log(apEV_SocialSourceReferral + 1)
, log2w_apEV_channel_Social = log(apEV_channel_Social + 1)
, log2w_wEV_event_M1X = log(wEV_event_M1X + 1)
, log2w_apEV_channel_Paid_Search = log(apEV_channel_Paid_Search + 1)
, log2w_apEV_action_M1X_talalkozottDuplikaltHirdetessel = log(apEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
, log2w_apEV_action_Erdeklodes_OfficeUtcapontos = log(apEV_action_Erdeklodes_OfficeUtcapontos + 1)
, log2w_everEV_trafficSource_Hirdetesfigyelo_email = log(everEV_trafficSource_Hirdetesfigyelo_email + 1)
, log2w_wEV_channel_Email = log(wEV_channel_Email + 1)
, log2w_wEV_channel_Organic_Search = log(wEV_channel_Organic_Search + 1)
, log2w_apEV_channel_Email = log(apEV_channel_Email + 1)
, log2w_wEV_referer_google = log(wEV_referer_google + 1)
, log2w_wEV_trafficSource_google_organic = log(wEV_trafficSource_google_organic + 1)
, log2w_everEV_event_AI = log(everEV_event_AI + 1)
, log2w_apEV_action_Erdeklodes_OfficeNemUtcapontos = log(apEV_action_Erdeklodes_OfficeNemUtcapontos + 1)
, log2w_everEV_action_M1X_talalkozottDuplikaltHirdetessel = log(everEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
, log2w_everEV_action_FrontendBelepes_Sikeres = log(everEV_action_FrontendBelepes_Sikeres + 1)
, log2w_everEV_label_FrontendBelepes_Sikeres_fejlec = log(everEV_label_FrontendBelepes_Sikeres_fejlec + 1)
, log2w_everEV_trafficSource_google_cpc = log(everEV_trafficSource_google_cpc + 1)
, log2w_wEV_trafficSource_Hirdetesfigyelo_email = log(wEV_trafficSource_Hirdetesfigyelo_email + 1)
, log2w_everEV_channel_Paid_Search = log(everEV_channel_Paid_Search + 1)
, log2w_everEV_event_OfficeLogo = log(everEV_event_OfficeLogo + 1)
, log2w_everEV_event_lakoparkListaKattintas = log(everEV_event_lakoparkListaKattintas + 1)
, log2w_wEV_action_M1X_talalkozottDuplikaltHirdetessel = log(wEV_action_M1X_talalkozottDuplikaltHirdetessel + 1)
, log2w_apEV_trafficSource_Hirdetesfigyelo_email = log(apEV_trafficSource_Hirdetesfigyelo_email + 1)
, log2w_everEV_referer_icom = log(everEV_referer_icom + 1)
, log2w_wEV_trafficSource_google_cpc = log(wEV_trafficSource_google_cpc + 1)
, log2w_everEV_Felhasznalo_Magan = log(everEV_Felhasznalo_Magan + 1)
, log2w_everEV_event_Piac = log(everEV_event_Piac + 1)
, log2w_everEV_action_Piac_undefined = log(everEV_action_Piac_undefined + 1)
, log2w_everEV_referer_google = log(everEV_referer_google + 1)
, log2w_everEV_channel_Direct = log(everEV_channel_Direct + 1)
, log2w_everRE_AVG_Kepgaleria = log(everRE_AVG_Kepgaleria + 1)
, log2w_wEV_channel_Paid_Search = log(wEV_channel_Paid_Search + 1)
, log2w_everRE_MAX_Kepgaleria = log(everRE_MAX_Kepgaleria + 1)
, log2w_everRE_SUM_Kepgaleria = log(everRE_SUM_Kepgaleria + 1)
, log2w_everRE_megye_Budapest = log(everRE_megye_Budapest + 1)
, log2w_everEV_event_Mobilreszletek = log(everEV_event_Mobilreszletek + 1)
, log2w_apRE_AVG_Kepgaleria = log(apRE_AVG_Kepgaleria + 1)
, log2w_apRE_MAX_Kepgaleria = log(apRE_MAX_Kepgaleria + 1)
, log2w_apRE_SUM_Kepgaleria = log(apRE_SUM_Kepgaleria + 1)
, log2w_apEV_action_Mobilreszletek_Kepgaleria = log(apEV_action_Mobilreszletek_Kepgaleria + 1)
, log2w_everEV_action_Erdeklodes_OfficeUtcapontos = log(everEV_action_Erdeklodes_OfficeUtcapontos + 1)
, log2w_apEV_trafficSource_rtbhouse_retargeting = log(apEV_trafficSource_rtbhouse_retargeting + 1)
, log2w_everRE_ikerhaz = log(everRE_ikerhaz + 1)
, log2w_wEV_action_Mobilreszletek_Kepgaleria = log(wEV_action_Mobilreszletek_Kepgaleria + 1)
, log2w_wEV_trafficSource_rtbhouse_retargeting = log(wEV_trafficSource_rtbhouse_retargeting + 1)
, log2w_everEV_Visits = log(everEV_Visits + 1)
, log2w_everEV_action_Mobilreszletek_Kepgaleria = log(everEV_action_Mobilreszletek_Kepgaleria + 1)
, log2w_apEV_event_Mobilreszletek = log(apEV_event_Mobilreszletek + 1)
, log2w_everEV_action_Erdeklodes_OfficeNemUtcapontos = log(everEV_action_Erdeklodes_OfficeNemUtcapontos + 1)
, log2w_wEV_event_Mobilreszletek = log(wEV_event_Mobilreszletek + 1)
, log2w_everEV_action_Erdeklodes_MaganFizetos = log(everEV_action_Erdeklodes_MaganFizetos + 1)
, log2w_wRE_AVG_Kepgaleria = log(wRE_AVG_Kepgaleria + 1)
, log2w_wEV_referer_creativecdn = log(wEV_referer_creativecdn + 1)
, log2w_wRE_AVG_Hirdetesar = log(wRE_AVG_Hirdetesar + 1)
, log2w_wRE_MAX_Kepgaleria = log(wRE_MAX_Kepgaleria + 1)
, log2w_wRE_SUM_Kepgaleria = log(wRE_SUM_Kepgaleria + 1)
, log2w_apRE_megye_Budapest = log(apRE_megye_Budapest + 1)
, log2w_everRE_megye_Pest = log(everRE_megye_Pest + 1)
, log2w_everEV_event_MaganhirdetoInfok = log(everEV_event_MaganhirdetoInfok + 1)
, log2w_wRE_nyaralo = log(wRE_nyaralo + 1)
, log2w_everRE_ujszeru = log(everRE_ujszeru + 1)
, log2w_apRE_AVG_Hirdetesar = log(apRE_AVG_Hirdetesar + 1)
, log2w_apEV_referer_creativecdn = log(apEV_referer_creativecdn + 1)
, log2w_wRE_megye_Budapest = log(wRE_megye_Budapest + 1)
, log2w_everEV_referer_creativecdn = log(everEV_referer_creativecdn + 1)
, log2w_everEV_Videkrol_Budapestre = log(everEV_Videkrol_Budapestre + 1)
, log2w_apEV_Videkrol_Budapestre = log(apEV_Videkrol_Budapestre + 1)
, log2w_wRE_MAX_Hirdetesar = log(wRE_MAX_Hirdetesar + 1)
, log2w_everRE_ker_11 = log(everRE_ker_11 + 1)
, log2w_everEV_channel_Display = log(everEV_channel_Display + 1)
, log2w_everEV_MAX_visitNumber = log(everEV_MAX_visitNumber + 1)
, log2w_everEV_MIN_visitNumber = log(everEV_MIN_visitNumber + 1)
, log2w_wRE_MAX_DIFF_Visit_per_RE = log(wRE_MAX_DIFF_Visit_per_RE + 1)
, log2w_everRE_AVG_Hirdetesar = log(everRE_AVG_Hirdetesar + 1)
, log2w_wRE_AVG_DIFF_Visit_per_RE = log(wRE_AVG_DIFF_Visit_per_RE + 1)
, log2w_everEV_channel_Referral = log(everEV_channel_Referral + 1)
, log2w_everEV_trafficSource_rtbhouse_retargeting = log(everEV_trafficSource_rtbhouse_retargeting + 1)
, log2w_everRE_tegla = log(everRE_tegla + 1)
, log2w_apEV_event_MaganhirdetoInfok = log(apEV_event_MaganhirdetoInfok + 1)
, log2w_everEV_action_Erdeklodes_MaganIngyenes = log(everEV_action_Erdeklodes_MaganIngyenes + 1)
, log2w_wEV_Videkrol_Budapestre = log(wEV_Videkrol_Budapestre + 1)
, log2w_apRE_tegla = log(apRE_tegla + 1)
, log2w_everRE_felujitott = log(everRE_felujitott + 1)
, log2w_everRE_lakas = log(everRE_lakas + 1)
, log2w_apRE_megye_Pest = log(apRE_megye_Pest + 1)
, log2w_apRE_lakas = log(apRE_lakas + 1)
, log2w_everEV_action_MaganToDo_Megjelenes = log(everEV_action_MaganToDo_Megjelenes + 1)
, log2w_everRE_SUM_DIFF_Visit_per_RE = log(everRE_SUM_DIFF_Visit_per_RE + 1)

),]

```

# Variables 1 week

```{r}
var_20_log_simple <- c("log_apEV_event_autosuggest", "log_everEV_event_Mobilelhelyezkedes", "log_everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik", "log_apEV_event_Mobilelhelyezkedes", "log_everEV_event_autosuggest", "log_wEV_event_autosuggest", "log_wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik", "log_apEV_timeOnSite", "log_apEV_pageviews")

var_21_log_all <- c(names(JZ_train)[grep("^log_.*", names(JZ_train))])
var_21_dropped <- c("xoxo")
var_21_log_all <- var_21_log_all[! var_21_log_all %in% var_21_dropped]
```

# Variables 2 weeks

```{r}
var_20_log_simple_2w <- c("log_apEV_event_autosuggest", "log_everEV_event_Mobilelhelyezkedes", "log_everEV_action_Mobilelhelyezkedes_Autocompletemegjelenik"
, "log_apEV_event_Mobilelhelyezkedes", "log_everEV_event_autosuggest", "log_wEV_event_autosuggest"
, "log_wEV_action_Mobilelhelyezkedes_Autocompletemegjelenik", "log_apEV_timeOnSite", "log_apEV_pageviews"
, "log_apEV_action_Mobilelhelyezkedes_Autocompletemegjelenik", "log_wEV_event_Mobilelhelyezkedes", "log_apEV_HirdetesTipus_Lakossag"
, "log_apEV_HirdetesEladasmodja_Elado", "log_apEV_device_desktop", "log_apEV_hits", "log_wEV_timeOnSite", "log_wEV_HirdetesTipus_Lakossag"
, "log_apEV_Honnan_Hungary")

var_21_log_all_2w <- c(names(JZ_train)[grep("^log2w_.*", names(JZ_train))])
var_21_dropped_2w <- c("xoxo")
var_21_log_all_2w <- var_21_log_all_2w[! var_21_log_all_2w %in% var_21_dropped_2w]
```
